Covariance/correlation for joint distribution of an mcgf object

Covariance/correlation for joint distribution of an mcgf object

Usage

# S3 method for mcgf
ccov(x, model = c("all", "base", "empirical"), cor = FALSE, ...)

Arguments

x

An mcgf object.

model

Which model to use. One of all, base, and empirical.

cor

Logical; if TRUE, correlation is outputted.

...

Additional arguments. Not in use.

Value

Joint covariance/correlation matrix.

Examples

data(sim1)
sim1_mcgf <- mcgf(sim1$data, dists = sim1$dists)
#> `time` is not provided, assuming rows are equally spaced temporally.
sim1_mcgf <- add_acfs(sim1_mcgf, lag_max = 5)
sim1_mcgf <- add_ccfs(sim1_mcgf, lag_max = 5)

# Fit a separable model and store it to 'sim1_mcgf'
fit_sep <- fit_base(
    sim1_mcgf,
    model = "sep",
    lag = 5,
    par_init = c(
        c = 0.001,
        gamma = 0.5,
        a = 0.3,
        alpha = 0.5
    ),
    par_fixed = c(nugget = 0)
)
sim1_mcgf <- add_base(sim1_mcgf, fit_base = fit_sep)

ccov(sim1_mcgf, model = "base")
#>              1:t+0      2:t+0      3:t+0      4:t+0      5:t+0      6:t+0
#> 1:t+0   0.97269389 0.85890922 0.78778667 0.82315843 0.75041556 0.80835264
#> 2:t+0   0.85890922 0.95033944 0.82344463 0.90844780 0.83029571 0.76617248
#> 3:t+0   0.78778667 0.82344463 0.95462464 0.83500131 0.79293041 0.81877287
#> 4:t+0   0.82315843 0.90844780 0.83500131 0.94060539 0.85550880 0.75409224
#> 5:t+0   0.75041556 0.83029571 0.79293041 0.85550880 0.92155031 0.69410042
#> 6:t+0   0.80835264 0.76617248 0.81877287 0.75409224 0.69410042 0.99883117
#> 7:t+0   0.84048905 0.88667429 0.86132193 0.88249365 0.80814368 0.79522031
#> 8:t+0   0.62124560 0.67986717 0.63455838 0.69129711 0.73611793 0.55529707
#> 9:t+0   0.91277332 0.91129982 0.80240103 0.87115486 0.79946060 0.77888512
#> 10:t+0  0.77584973 0.85459385 0.78273214 0.86905778 0.89271746 0.69677882
#> 1:lag1  0.59765450 0.52774153 0.48404154 0.50577509 0.46107952 0.49667794
#> 2:lag1  0.52774153 0.58391921 0.50595094 0.55817963 0.51016047 0.47076108
#> 3:lag1  0.48404154 0.50595094 0.58655217 0.51305174 0.48720202 0.50308046
#> 4:lag1  0.50577509 0.55817963 0.51305174 0.57793829 0.52565220 0.46333860
#> 5:lag1  0.46107952 0.51016047 0.48720202 0.52565220 0.56623024 0.42647769
#> 6:lag1  0.49667794 0.47076108 0.50308046 0.46333860 0.42647769 0.61371409
#> 7:lag1  0.51642358 0.54480128 0.52922397 0.54223257 0.49654954 0.48860901
#> 8:lag1  0.38171333 0.41773232 0.38989314 0.42475524 0.45229460 0.34119243
#> 9:lag1  0.56083737 0.55993200 0.49302107 0.53526565 0.49121438 0.47857214
#> 10:lag1 0.47670710 0.52509003 0.48093587 0.53397713 0.54851440 0.42812338
#> 1:lag2  0.35540230 0.31382773 0.28784102 0.30076513 0.27418638 0.29535540
#> 2:lag2  0.31382773 0.34723445 0.30086970 0.33192811 0.30337295 0.27994364
#> 3:lag2  0.28784102 0.30086970 0.34880017 0.30509227 0.28972043 0.29916274
#> 4:lag2  0.30076513 0.33192811 0.30509227 0.34367783 0.31258529 0.27552977
#> 5:lag2  0.27418638 0.30337295 0.28972043 0.31258529 0.33671550 0.25360999
#> 6:lag2  0.29535540 0.27994364 0.29916274 0.27552977 0.25360999 0.36495233
#> 7:lag2  0.30709738 0.32397252 0.31470928 0.32244500 0.29527905 0.29055712
#> 8:lag2  0.22699034 0.24840945 0.23185456 0.25258572 0.26896232 0.20289411
#> 9:lag2  0.33350856 0.33297017 0.29318080 0.31830204 0.29210643 0.28458857
#> 10:lag2 0.28347951 0.31225098 0.28599420 0.31753581 0.32618056 0.25458862
#> 1:lag3  0.23434438 0.20693103 0.18979597 0.19831784 0.18079241 0.19475079
#> 2:lag3  0.20693103 0.22895868 0.19838679 0.21886602 0.20003738 0.18458861
#> 3:lag3  0.18979597 0.19838679 0.22999108 0.20117106 0.19103521 0.19726126
#> 4:lag3  0.19831784 0.21886602 0.20117106 0.22661352 0.20611179 0.18167821
#> 5:lag3  0.18079241 0.20003738 0.19103521 0.20611179 0.22202272 0.16722479
#> 6:lag3  0.19475079 0.18458861 0.19726126 0.18167821 0.16722479 0.24064145
#> 7:lag3  0.20249319 0.21362028 0.20751231 0.21261307 0.19470044 0.19158691
#> 8:lag3  0.14967239 0.16379567 0.15287974 0.16654941 0.17734778 0.13378387
#> 9:lag3  0.21990814 0.21955313 0.19331691 0.20988129 0.19260849 0.18765138
#> 10:lag3 0.18692009 0.20589136 0.18857822 0.20937605 0.21507622 0.16787008
#> 1:lag4  0.16749110 0.14789817 0.13565136 0.14174214 0.12921632 0.13919268
#> 2:lag4  0.14789817 0.16364182 0.14179142 0.15642837 0.14297113 0.13192955
#> 3:lag4  0.13565136 0.14179142 0.16437970 0.14378140 0.13653708 0.14098697
#> 4:lag4  0.14174214 0.15642837 0.14378140 0.16196568 0.14731264 0.12984942
#> 5:lag4  0.12921632 0.14297113 0.13653708 0.14731264 0.15868453 0.11951925
#> 6:lag4  0.13919268 0.13192955 0.14098697 0.12984942 0.11951925 0.17199175
#> 7:lag4  0.14472635 0.15267912 0.14831362 0.15195925 0.13915670 0.13693139
#> 8:lag4  0.10697415 0.11706838 0.10926652 0.11903654 0.12675437 0.09561828
#> 9:lag4  0.15717319 0.15691947 0.13816785 0.15000678 0.13766153 0.13411858
#> 10:lag4 0.13359591 0.14715510 0.13478101 0.14964568 0.15371971 0.11998045
#> 1:lag5  0.12678596 0.11195468 0.10268419 0.10729473 0.09781305 0.10536487
#> 2:lag5  0.11195468 0.12387216 0.10733204 0.11841179 0.10822504 0.09986689
#> 3:lag5  0.10268419 0.10733204 0.12443072 0.10883839 0.10335466 0.10672310
#> 4:lag5  0.10729473 0.11841179 0.10883839 0.12260338 0.11151145 0.09829229
#> 5:lag5  0.09781305 0.10822504 0.10335466 0.11151145 0.12011964 0.09047264
#> 6:lag5  0.10536487 0.09986689 0.10672310 0.09829229 0.09047264 0.13019283
#> 7:lag5  0.10955369 0.11557372 0.11226916 0.11502879 0.10533763 0.10365313
#> 8:lag5  0.08097637 0.08861741 0.08271163 0.09010724 0.09594942 0.07238030
#> 9:lag5  0.11897560 0.11878354 0.10458911 0.11355084 0.10420583 0.10152392
#> 10:lag5 0.10112827 0.11139219 0.10202536 0.11327749 0.11636142 0.09082176
#>              7:t+0      8:t+0      9:t+0     10:t+0    1:lag1    2:lag1
#> 1:t+0   0.84048905 0.62124560 0.91277332 0.77584973 0.5976545 0.5277415
#> 2:t+0   0.88667429 0.67986717 0.91129982 0.85459385 0.5277415 0.5839192
#> 3:t+0   0.86132193 0.63455838 0.80240103 0.78273214 0.4840415 0.5059509
#> 4:t+0   0.88249365 0.69129711 0.87115486 0.86905778 0.5057751 0.5581796
#> 5:t+0   0.80814368 0.73611793 0.79946060 0.89271746 0.4610795 0.5101605
#> 6:t+0   0.79522031 0.55529707 0.77888512 0.69677882 0.4966779 0.4707611
#> 7:t+0   0.93063097 0.64911407 0.86660864 0.81541392 0.5164236 0.5448013
#> 8:t+0   0.64911407 0.89790658 0.66662312 0.75638841 0.3817133 0.4177323
#> 9:t+0   0.86660864 0.66662312 0.97945162 0.83096437 0.5608374 0.5599320
#> 10:t+0  0.81541392 0.75638841 0.83096437 0.95178452 0.4767071 0.5250900
#> 1:lag1  0.51642358 0.38171333 0.56083737 0.47670710 0.9726939 0.8589092
#> 2:lag1  0.54480128 0.41773232 0.55993200 0.52509003 0.8589092 0.9503394
#> 3:lag1  0.52922397 0.38989314 0.49302107 0.48093587 0.7877867 0.8234446
#> 4:lag1  0.54223257 0.42475524 0.53526565 0.53397713 0.8231584 0.9084478
#> 5:lag1  0.49654954 0.45229460 0.49121438 0.54851440 0.7504156 0.8302957
#> 6:lag1  0.48860901 0.34119243 0.47857214 0.42812338 0.8083526 0.7661725
#> 7:lag1  0.57180969 0.39883662 0.53247230 0.50101662 0.8404890 0.8866743
#> 8:lag1  0.39883662 0.55170277 0.40959475 0.46474944 0.6212456 0.6798672
#> 9:lag1  0.53247230 0.40959475 0.60180667 0.51057131 0.9127733 0.9112998
#> 10:lag1 0.50101662 0.46474944 0.51057131 0.58480711 0.7758497 0.8545939
#> 1:lag2  0.30709738 0.22699034 0.33350856 0.28347951 0.5976545 0.5277415
#> 2:lag2  0.32397252 0.24840945 0.33297017 0.31225098 0.5277415 0.5839192
#> 3:lag2  0.31470928 0.23185456 0.29318080 0.28599420 0.4840415 0.5059509
#> 4:lag2  0.32244500 0.25258572 0.31830204 0.31753581 0.5057751 0.5581796
#> 5:lag2  0.29527905 0.26896232 0.29210643 0.32618056 0.4610795 0.5101605
#> 6:lag2  0.29055712 0.20289411 0.28458857 0.25458862 0.4966779 0.4707611
#> 7:lag2  0.34003338 0.23717291 0.31664094 0.29793545 0.5164236 0.5448013
#> 8:lag2  0.23717291 0.32807656 0.24357035 0.27636874 0.3817133 0.4177323
#> 9:lag2  0.31664094 0.24357035 0.35787144 0.30361726 0.5608374 0.5599320
#> 10:lag2 0.29793545 0.27636874 0.30361726 0.34776245 0.4767071 0.5250900
#> 1:lag3  0.20249319 0.14967239 0.21990814 0.18692009 0.3554023 0.3138277
#> 2:lag3  0.21362028 0.16379567 0.21955313 0.20589136 0.3138277 0.3472345
#> 3:lag3  0.20751231 0.15287974 0.19331691 0.18857822 0.2878410 0.3008697
#> 4:lag3  0.21261307 0.16654941 0.20988129 0.20937605 0.3007651 0.3319281
#> 5:lag3  0.19470044 0.17734778 0.19260849 0.21507622 0.2741864 0.3033729
#> 6:lag3  0.19158691 0.13378387 0.18765138 0.16787008 0.2953554 0.2799436
#> 7:lag3  0.22421046 0.15638654 0.20878600 0.19645201 0.3070974 0.3239725
#> 8:lag3  0.15638654 0.21632640 0.16060488 0.18223141 0.2269903 0.2484095
#> 9:lag3  0.20878600 0.16060488 0.23597247 0.20019847 0.3335086 0.3329702
#> 10:lag3 0.19645201 0.18223141 0.20019847 0.22930683 0.2834795 0.3122510
#> 1:lag4  0.14472635 0.10697415 0.15717319 0.13359591 0.2343444 0.2069310
#> 2:lag4  0.15267912 0.11706838 0.15691947 0.14715510 0.2069310 0.2289587
#> 3:lag4  0.14831362 0.10926652 0.13816785 0.13478101 0.1897960 0.1983868
#> 4:lag4  0.15195925 0.11903654 0.15000678 0.14964568 0.1983178 0.2188660
#> 5:lag4  0.13915670 0.12675437 0.13766153 0.15371971 0.1807924 0.2000374
#> 6:lag4  0.13693139 0.09561828 0.13411858 0.11998045 0.1947508 0.1845886
#> 7:lag4  0.16024816 0.11177291 0.14922396 0.14040858 0.2024932 0.2136203
#> 8:lag4  0.11177291 0.15461324 0.11478785 0.13024480 0.1496724 0.1637957
#> 9:lag4  0.14922396 0.11478785 0.16865473 0.14308626 0.2199081 0.2195531
#> 10:lag4 0.14040858 0.13024480 0.14308626 0.16389065 0.1869201 0.2058914
#> 1:lag5  0.10955369 0.08097637 0.11897560 0.10112827 0.1674911 0.1478982
#> 2:lag5  0.11557372 0.08861741 0.11878354 0.11139219 0.1478982 0.1636418
#> 3:lag5  0.11226916 0.08271163 0.10458911 0.10202536 0.1356514 0.1417914
#> 4:lag5  0.11502879 0.09010724 0.11355084 0.11327749 0.1417421 0.1564284
#> 5:lag5  0.10533763 0.09594942 0.10420583 0.11636142 0.1292163 0.1429711
#> 6:lag5  0.10365313 0.07238030 0.10152392 0.09082176 0.1391927 0.1319296
#> 7:lag5  0.12130326 0.08460889 0.11295826 0.10628527 0.1447263 0.1526791
#> 8:lag5  0.08460889 0.11703779 0.08689111 0.09859158 0.1069742 0.1170684
#> 9:lag5  0.11295826 0.08689111 0.12766680 0.10831220 0.1571732 0.1569195
#> 10:lag5 0.10628527 0.09859158 0.10831220 0.12406052 0.1335959 0.1471551
#>            3:lag1    4:lag1    5:lag1     6:lag1    7:lag1     8:lag1    9:lag1
#> 1:t+0   0.4840415 0.5057751 0.4610795 0.49667794 0.5164236 0.38171333 0.5608374
#> 2:t+0   0.5059509 0.5581796 0.5101605 0.47076108 0.5448013 0.41773232 0.5599320
#> 3:t+0   0.5865522 0.5130517 0.4872020 0.50308046 0.5292240 0.38989314 0.4930211
#> 4:t+0   0.5130517 0.5779383 0.5256522 0.46333860 0.5422326 0.42475524 0.5352656
#> 5:t+0   0.4872020 0.5256522 0.5662302 0.42647769 0.4965495 0.45229460 0.4912144
#> 6:t+0   0.5030805 0.4633386 0.4264777 0.61371409 0.4886090 0.34119243 0.4785721
#> 7:t+0   0.5292240 0.5422326 0.4965495 0.48860901 0.5718097 0.39883662 0.5324723
#> 8:t+0   0.3898931 0.4247552 0.4522946 0.34119243 0.3988366 0.55170277 0.4095947
#> 9:t+0   0.4930211 0.5352656 0.4912144 0.47857214 0.5324723 0.40959475 0.6018067
#> 10:t+0  0.4809359 0.5339771 0.5485144 0.42812338 0.5010166 0.46474944 0.5105713
#> 1:lag1  0.7877867 0.8231584 0.7504156 0.80835264 0.8404890 0.62124560 0.9127733
#> 2:lag1  0.8234446 0.9084478 0.8302957 0.76617248 0.8866743 0.67986717 0.9112998
#> 3:lag1  0.9546246 0.8350013 0.7929304 0.81877287 0.8613219 0.63455838 0.8024010
#> 4:lag1  0.8350013 0.9406054 0.8555088 0.75409224 0.8824937 0.69129711 0.8711549
#> 5:lag1  0.7929304 0.8555088 0.9215503 0.69410042 0.8081437 0.73611793 0.7994606
#> 6:lag1  0.8187729 0.7540922 0.6941004 0.99883117 0.7952203 0.55529707 0.7788851
#> 7:lag1  0.8613219 0.8824937 0.8081437 0.79522031 0.9306310 0.64911407 0.8666086
#> 8:lag1  0.6345584 0.6912971 0.7361179 0.55529707 0.6491141 0.89790658 0.6666231
#> 9:lag1  0.8024010 0.8711549 0.7994606 0.77888512 0.8666086 0.66662312 0.9794516
#> 10:lag1 0.7827321 0.8690578 0.8927175 0.69677882 0.8154139 0.75638841 0.8309644
#> 1:lag2  0.4840415 0.5057751 0.4610795 0.49667794 0.5164236 0.38171333 0.5608374
#> 2:lag2  0.5059509 0.5581796 0.5101605 0.47076108 0.5448013 0.41773232 0.5599320
#> 3:lag2  0.5865522 0.5130517 0.4872020 0.50308046 0.5292240 0.38989314 0.4930211
#> 4:lag2  0.5130517 0.5779383 0.5256522 0.46333860 0.5422326 0.42475524 0.5352656
#> 5:lag2  0.4872020 0.5256522 0.5662302 0.42647769 0.4965495 0.45229460 0.4912144
#> 6:lag2  0.5030805 0.4633386 0.4264777 0.61371409 0.4886090 0.34119243 0.4785721
#> 7:lag2  0.5292240 0.5422326 0.4965495 0.48860901 0.5718097 0.39883662 0.5324723
#> 8:lag2  0.3898931 0.4247552 0.4522946 0.34119243 0.3988366 0.55170277 0.4095947
#> 9:lag2  0.4930211 0.5352656 0.4912144 0.47857214 0.5324723 0.40959475 0.6018067
#> 10:lag2 0.4809359 0.5339771 0.5485144 0.42812338 0.5010166 0.46474944 0.5105713
#> 1:lag3  0.2878410 0.3007651 0.2741864 0.29535540 0.3070974 0.22699034 0.3335086
#> 2:lag3  0.3008697 0.3319281 0.3033729 0.27994364 0.3239725 0.24840945 0.3329702
#> 3:lag3  0.3488002 0.3050923 0.2897204 0.29916274 0.3147093 0.23185456 0.2931808
#> 4:lag3  0.3050923 0.3436778 0.3125853 0.27552977 0.3224450 0.25258572 0.3183020
#> 5:lag3  0.2897204 0.3125853 0.3367155 0.25360999 0.2952790 0.26896232 0.2921064
#> 6:lag3  0.2991627 0.2755298 0.2536100 0.36495233 0.2905571 0.20289411 0.2845886
#> 7:lag3  0.3147093 0.3224450 0.2952790 0.29055712 0.3400334 0.23717291 0.3166409
#> 8:lag3  0.2318546 0.2525857 0.2689623 0.20289411 0.2371729 0.32807656 0.2435704
#> 9:lag3  0.2931808 0.3183020 0.2921064 0.28458857 0.3166409 0.24357035 0.3578714
#> 10:lag3 0.2859942 0.3175358 0.3261806 0.25458862 0.2979354 0.27636874 0.3036173
#> 1:lag4  0.1897960 0.1983178 0.1807924 0.19475079 0.2024932 0.14967239 0.2199081
#> 2:lag4  0.1983868 0.2188660 0.2000374 0.18458861 0.2136203 0.16379567 0.2195531
#> 3:lag4  0.2299911 0.2011711 0.1910352 0.19726126 0.2075123 0.15287974 0.1933169
#> 4:lag4  0.2011711 0.2266135 0.2061118 0.18167821 0.2126131 0.16654941 0.2098813
#> 5:lag4  0.1910352 0.2061118 0.2220227 0.16722479 0.1947004 0.17734778 0.1926085
#> 6:lag4  0.1972613 0.1816782 0.1672248 0.24064145 0.1915869 0.13378387 0.1876514
#> 7:lag4  0.2075123 0.2126131 0.1947004 0.19158691 0.2242105 0.15638654 0.2087860
#> 8:lag4  0.1528797 0.1665494 0.1773478 0.13378387 0.1563865 0.21632640 0.1606049
#> 9:lag4  0.1933169 0.2098813 0.1926085 0.18765138 0.2087860 0.16060488 0.2359725
#> 10:lag4 0.1885782 0.2093761 0.2150762 0.16787008 0.1964520 0.18223141 0.2001985
#> 1:lag5  0.1356514 0.1417421 0.1292163 0.13919268 0.1447263 0.10697415 0.1571732
#> 2:lag5  0.1417914 0.1564284 0.1429711 0.13192955 0.1526791 0.11706838 0.1569195
#> 3:lag5  0.1643797 0.1437814 0.1365371 0.14098697 0.1483136 0.10926652 0.1381679
#> 4:lag5  0.1437814 0.1619657 0.1473126 0.12984942 0.1519592 0.11903654 0.1500068
#> 5:lag5  0.1365371 0.1473126 0.1586845 0.11951925 0.1391567 0.12675437 0.1376615
#> 6:lag5  0.1409870 0.1298494 0.1195192 0.17199175 0.1369314 0.09561828 0.1341186
#> 7:lag5  0.1483136 0.1519592 0.1391567 0.13693139 0.1602482 0.11177291 0.1492240
#> 8:lag5  0.1092665 0.1190365 0.1267544 0.09561828 0.1117729 0.15461324 0.1147878
#> 9:lag5  0.1381679 0.1500068 0.1376615 0.13411858 0.1492240 0.11478785 0.1686547
#> 10:lag5 0.1347810 0.1496457 0.1537197 0.11998045 0.1404086 0.13024480 0.1430863
#>           10:lag1    1:lag2    2:lag2    3:lag2    4:lag2    5:lag2    6:lag2
#> 1:t+0   0.4767071 0.3554023 0.3138277 0.2878410 0.3007651 0.2741864 0.2953554
#> 2:t+0   0.5250900 0.3138277 0.3472345 0.3008697 0.3319281 0.3033729 0.2799436
#> 3:t+0   0.4809359 0.2878410 0.3008697 0.3488002 0.3050923 0.2897204 0.2991627
#> 4:t+0   0.5339771 0.3007651 0.3319281 0.3050923 0.3436778 0.3125853 0.2755298
#> 5:t+0   0.5485144 0.2741864 0.3033729 0.2897204 0.3125853 0.3367155 0.2536100
#> 6:t+0   0.4281234 0.2953554 0.2799436 0.2991627 0.2755298 0.2536100 0.3649523
#> 7:t+0   0.5010166 0.3070974 0.3239725 0.3147093 0.3224450 0.2952790 0.2905571
#> 8:t+0   0.4647494 0.2269903 0.2484095 0.2318546 0.2525857 0.2689623 0.2028941
#> 9:t+0   0.5105713 0.3335086 0.3329702 0.2931808 0.3183020 0.2921064 0.2845886
#> 10:t+0  0.5848071 0.2834795 0.3122510 0.2859942 0.3175358 0.3261806 0.2545886
#> 1:lag1  0.7758497 0.5976545 0.5277415 0.4840415 0.5057751 0.4610795 0.4966779
#> 2:lag1  0.8545939 0.5277415 0.5839192 0.5059509 0.5581796 0.5101605 0.4707611
#> 3:lag1  0.7827321 0.4840415 0.5059509 0.5865522 0.5130517 0.4872020 0.5030805
#> 4:lag1  0.8690578 0.5057751 0.5581796 0.5130517 0.5779383 0.5256522 0.4633386
#> 5:lag1  0.8927175 0.4610795 0.5101605 0.4872020 0.5256522 0.5662302 0.4264777
#> 6:lag1  0.6967788 0.4966779 0.4707611 0.5030805 0.4633386 0.4264777 0.6137141
#> 7:lag1  0.8154139 0.5164236 0.5448013 0.5292240 0.5422326 0.4965495 0.4886090
#> 8:lag1  0.7563884 0.3817133 0.4177323 0.3898931 0.4247552 0.4522946 0.3411924
#> 9:lag1  0.8309644 0.5608374 0.5599320 0.4930211 0.5352656 0.4912144 0.4785721
#> 10:lag1 0.9517845 0.4767071 0.5250900 0.4809359 0.5339771 0.5485144 0.4281234
#> 1:lag2  0.4767071 0.9726939 0.8589092 0.7877867 0.8231584 0.7504156 0.8083526
#> 2:lag2  0.5250900 0.8589092 0.9503394 0.8234446 0.9084478 0.8302957 0.7661725
#> 3:lag2  0.4809359 0.7877867 0.8234446 0.9546246 0.8350013 0.7929304 0.8187729
#> 4:lag2  0.5339771 0.8231584 0.9084478 0.8350013 0.9406054 0.8555088 0.7540922
#> 5:lag2  0.5485144 0.7504156 0.8302957 0.7929304 0.8555088 0.9215503 0.6941004
#> 6:lag2  0.4281234 0.8083526 0.7661725 0.8187729 0.7540922 0.6941004 0.9988312
#> 7:lag2  0.5010166 0.8404890 0.8866743 0.8613219 0.8824937 0.8081437 0.7952203
#> 8:lag2  0.4647494 0.6212456 0.6798672 0.6345584 0.6912971 0.7361179 0.5552971
#> 9:lag2  0.5105713 0.9127733 0.9112998 0.8024010 0.8711549 0.7994606 0.7788851
#> 10:lag2 0.5848071 0.7758497 0.8545939 0.7827321 0.8690578 0.8927175 0.6967788
#> 1:lag3  0.2834795 0.5976545 0.5277415 0.4840415 0.5057751 0.4610795 0.4966779
#> 2:lag3  0.3122510 0.5277415 0.5839192 0.5059509 0.5581796 0.5101605 0.4707611
#> 3:lag3  0.2859942 0.4840415 0.5059509 0.5865522 0.5130517 0.4872020 0.5030805
#> 4:lag3  0.3175358 0.5057751 0.5581796 0.5130517 0.5779383 0.5256522 0.4633386
#> 5:lag3  0.3261806 0.4610795 0.5101605 0.4872020 0.5256522 0.5662302 0.4264777
#> 6:lag3  0.2545886 0.4966779 0.4707611 0.5030805 0.4633386 0.4264777 0.6137141
#> 7:lag3  0.2979354 0.5164236 0.5448013 0.5292240 0.5422326 0.4965495 0.4886090
#> 8:lag3  0.2763687 0.3817133 0.4177323 0.3898931 0.4247552 0.4522946 0.3411924
#> 9:lag3  0.3036173 0.5608374 0.5599320 0.4930211 0.5352656 0.4912144 0.4785721
#> 10:lag3 0.3477625 0.4767071 0.5250900 0.4809359 0.5339771 0.5485144 0.4281234
#> 1:lag4  0.1869201 0.3554023 0.3138277 0.2878410 0.3007651 0.2741864 0.2953554
#> 2:lag4  0.2058914 0.3138277 0.3472345 0.3008697 0.3319281 0.3033729 0.2799436
#> 3:lag4  0.1885782 0.2878410 0.3008697 0.3488002 0.3050923 0.2897204 0.2991627
#> 4:lag4  0.2093761 0.3007651 0.3319281 0.3050923 0.3436778 0.3125853 0.2755298
#> 5:lag4  0.2150762 0.2741864 0.3033729 0.2897204 0.3125853 0.3367155 0.2536100
#> 6:lag4  0.1678701 0.2953554 0.2799436 0.2991627 0.2755298 0.2536100 0.3649523
#> 7:lag4  0.1964520 0.3070974 0.3239725 0.3147093 0.3224450 0.2952790 0.2905571
#> 8:lag4  0.1822314 0.2269903 0.2484095 0.2318546 0.2525857 0.2689623 0.2028941
#> 9:lag4  0.2001985 0.3335086 0.3329702 0.2931808 0.3183020 0.2921064 0.2845886
#> 10:lag4 0.2293068 0.2834795 0.3122510 0.2859942 0.3175358 0.3261806 0.2545886
#> 1:lag5  0.1335959 0.2343444 0.2069310 0.1897960 0.1983178 0.1807924 0.1947508
#> 2:lag5  0.1471551 0.2069310 0.2289587 0.1983868 0.2188660 0.2000374 0.1845886
#> 3:lag5  0.1347810 0.1897960 0.1983868 0.2299911 0.2011711 0.1910352 0.1972613
#> 4:lag5  0.1496457 0.1983178 0.2188660 0.2011711 0.2266135 0.2061118 0.1816782
#> 5:lag5  0.1537197 0.1807924 0.2000374 0.1910352 0.2061118 0.2220227 0.1672248
#> 6:lag5  0.1199804 0.1947508 0.1845886 0.1972613 0.1816782 0.1672248 0.2406415
#> 7:lag5  0.1404086 0.2024932 0.2136203 0.2075123 0.2126131 0.1947004 0.1915869
#> 8:lag5  0.1302448 0.1496724 0.1637957 0.1528797 0.1665494 0.1773478 0.1337839
#> 9:lag5  0.1430863 0.2199081 0.2195531 0.1933169 0.2098813 0.1926085 0.1876514
#> 10:lag5 0.1638906 0.1869201 0.2058914 0.1885782 0.2093761 0.2150762 0.1678701
#>            7:lag2    8:lag2    9:lag2   10:lag2    1:lag3    2:lag3    3:lag3
#> 1:t+0   0.3070974 0.2269903 0.3335086 0.2834795 0.2343444 0.2069310 0.1897960
#> 2:t+0   0.3239725 0.2484095 0.3329702 0.3122510 0.2069310 0.2289587 0.1983868
#> 3:t+0   0.3147093 0.2318546 0.2931808 0.2859942 0.1897960 0.1983868 0.2299911
#> 4:t+0   0.3224450 0.2525857 0.3183020 0.3175358 0.1983178 0.2188660 0.2011711
#> 5:t+0   0.2952790 0.2689623 0.2921064 0.3261806 0.1807924 0.2000374 0.1910352
#> 6:t+0   0.2905571 0.2028941 0.2845886 0.2545886 0.1947508 0.1845886 0.1972613
#> 7:t+0   0.3400334 0.2371729 0.3166409 0.2979354 0.2024932 0.2136203 0.2075123
#> 8:t+0   0.2371729 0.3280766 0.2435704 0.2763687 0.1496724 0.1637957 0.1528797
#> 9:t+0   0.3166409 0.2435704 0.3578714 0.3036173 0.2199081 0.2195531 0.1933169
#> 10:t+0  0.2979354 0.2763687 0.3036173 0.3477625 0.1869201 0.2058914 0.1885782
#> 1:lag1  0.5164236 0.3817133 0.5608374 0.4767071 0.3554023 0.3138277 0.2878410
#> 2:lag1  0.5448013 0.4177323 0.5599320 0.5250900 0.3138277 0.3472345 0.3008697
#> 3:lag1  0.5292240 0.3898931 0.4930211 0.4809359 0.2878410 0.3008697 0.3488002
#> 4:lag1  0.5422326 0.4247552 0.5352656 0.5339771 0.3007651 0.3319281 0.3050923
#> 5:lag1  0.4965495 0.4522946 0.4912144 0.5485144 0.2741864 0.3033729 0.2897204
#> 6:lag1  0.4886090 0.3411924 0.4785721 0.4281234 0.2953554 0.2799436 0.2991627
#> 7:lag1  0.5718097 0.3988366 0.5324723 0.5010166 0.3070974 0.3239725 0.3147093
#> 8:lag1  0.3988366 0.5517028 0.4095947 0.4647494 0.2269903 0.2484095 0.2318546
#> 9:lag1  0.5324723 0.4095947 0.6018067 0.5105713 0.3335086 0.3329702 0.2931808
#> 10:lag1 0.5010166 0.4647494 0.5105713 0.5848071 0.2834795 0.3122510 0.2859942
#> 1:lag2  0.8404890 0.6212456 0.9127733 0.7758497 0.5976545 0.5277415 0.4840415
#> 2:lag2  0.8866743 0.6798672 0.9112998 0.8545939 0.5277415 0.5839192 0.5059509
#> 3:lag2  0.8613219 0.6345584 0.8024010 0.7827321 0.4840415 0.5059509 0.5865522
#> 4:lag2  0.8824937 0.6912971 0.8711549 0.8690578 0.5057751 0.5581796 0.5130517
#> 5:lag2  0.8081437 0.7361179 0.7994606 0.8927175 0.4610795 0.5101605 0.4872020
#> 6:lag2  0.7952203 0.5552971 0.7788851 0.6967788 0.4966779 0.4707611 0.5030805
#> 7:lag2  0.9306310 0.6491141 0.8666086 0.8154139 0.5164236 0.5448013 0.5292240
#> 8:lag2  0.6491141 0.8979066 0.6666231 0.7563884 0.3817133 0.4177323 0.3898931
#> 9:lag2  0.8666086 0.6666231 0.9794516 0.8309644 0.5608374 0.5599320 0.4930211
#> 10:lag2 0.8154139 0.7563884 0.8309644 0.9517845 0.4767071 0.5250900 0.4809359
#> 1:lag3  0.5164236 0.3817133 0.5608374 0.4767071 0.9726939 0.8589092 0.7877867
#> 2:lag3  0.5448013 0.4177323 0.5599320 0.5250900 0.8589092 0.9503394 0.8234446
#> 3:lag3  0.5292240 0.3898931 0.4930211 0.4809359 0.7877867 0.8234446 0.9546246
#> 4:lag3  0.5422326 0.4247552 0.5352656 0.5339771 0.8231584 0.9084478 0.8350013
#> 5:lag3  0.4965495 0.4522946 0.4912144 0.5485144 0.7504156 0.8302957 0.7929304
#> 6:lag3  0.4886090 0.3411924 0.4785721 0.4281234 0.8083526 0.7661725 0.8187729
#> 7:lag3  0.5718097 0.3988366 0.5324723 0.5010166 0.8404890 0.8866743 0.8613219
#> 8:lag3  0.3988366 0.5517028 0.4095947 0.4647494 0.6212456 0.6798672 0.6345584
#> 9:lag3  0.5324723 0.4095947 0.6018067 0.5105713 0.9127733 0.9112998 0.8024010
#> 10:lag3 0.5010166 0.4647494 0.5105713 0.5848071 0.7758497 0.8545939 0.7827321
#> 1:lag4  0.3070974 0.2269903 0.3335086 0.2834795 0.5976545 0.5277415 0.4840415
#> 2:lag4  0.3239725 0.2484095 0.3329702 0.3122510 0.5277415 0.5839192 0.5059509
#> 3:lag4  0.3147093 0.2318546 0.2931808 0.2859942 0.4840415 0.5059509 0.5865522
#> 4:lag4  0.3224450 0.2525857 0.3183020 0.3175358 0.5057751 0.5581796 0.5130517
#> 5:lag4  0.2952790 0.2689623 0.2921064 0.3261806 0.4610795 0.5101605 0.4872020
#> 6:lag4  0.2905571 0.2028941 0.2845886 0.2545886 0.4966779 0.4707611 0.5030805
#> 7:lag4  0.3400334 0.2371729 0.3166409 0.2979354 0.5164236 0.5448013 0.5292240
#> 8:lag4  0.2371729 0.3280766 0.2435704 0.2763687 0.3817133 0.4177323 0.3898931
#> 9:lag4  0.3166409 0.2435704 0.3578714 0.3036173 0.5608374 0.5599320 0.4930211
#> 10:lag4 0.2979354 0.2763687 0.3036173 0.3477625 0.4767071 0.5250900 0.4809359
#> 1:lag5  0.2024932 0.1496724 0.2199081 0.1869201 0.3554023 0.3138277 0.2878410
#> 2:lag5  0.2136203 0.1637957 0.2195531 0.2058914 0.3138277 0.3472345 0.3008697
#> 3:lag5  0.2075123 0.1528797 0.1933169 0.1885782 0.2878410 0.3008697 0.3488002
#> 4:lag5  0.2126131 0.1665494 0.2098813 0.2093761 0.3007651 0.3319281 0.3050923
#> 5:lag5  0.1947004 0.1773478 0.1926085 0.2150762 0.2741864 0.3033729 0.2897204
#> 6:lag5  0.1915869 0.1337839 0.1876514 0.1678701 0.2953554 0.2799436 0.2991627
#> 7:lag5  0.2242105 0.1563865 0.2087860 0.1964520 0.3070974 0.3239725 0.3147093
#> 8:lag5  0.1563865 0.2163264 0.1606049 0.1822314 0.2269903 0.2484095 0.2318546
#> 9:lag5  0.2087860 0.1606049 0.2359725 0.2001985 0.3335086 0.3329702 0.2931808
#> 10:lag5 0.1964520 0.1822314 0.2001985 0.2293068 0.2834795 0.3122510 0.2859942
#>            4:lag3    5:lag3    6:lag3    7:lag3    8:lag3    9:lag3   10:lag3
#> 1:t+0   0.1983178 0.1807924 0.1947508 0.2024932 0.1496724 0.2199081 0.1869201
#> 2:t+0   0.2188660 0.2000374 0.1845886 0.2136203 0.1637957 0.2195531 0.2058914
#> 3:t+0   0.2011711 0.1910352 0.1972613 0.2075123 0.1528797 0.1933169 0.1885782
#> 4:t+0   0.2266135 0.2061118 0.1816782 0.2126131 0.1665494 0.2098813 0.2093761
#> 5:t+0   0.2061118 0.2220227 0.1672248 0.1947004 0.1773478 0.1926085 0.2150762
#> 6:t+0   0.1816782 0.1672248 0.2406415 0.1915869 0.1337839 0.1876514 0.1678701
#> 7:t+0   0.2126131 0.1947004 0.1915869 0.2242105 0.1563865 0.2087860 0.1964520
#> 8:t+0   0.1665494 0.1773478 0.1337839 0.1563865 0.2163264 0.1606049 0.1822314
#> 9:t+0   0.2098813 0.1926085 0.1876514 0.2087860 0.1606049 0.2359725 0.2001985
#> 10:t+0  0.2093761 0.2150762 0.1678701 0.1964520 0.1822314 0.2001985 0.2293068
#> 1:lag1  0.3007651 0.2741864 0.2953554 0.3070974 0.2269903 0.3335086 0.2834795
#> 2:lag1  0.3319281 0.3033729 0.2799436 0.3239725 0.2484095 0.3329702 0.3122510
#> 3:lag1  0.3050923 0.2897204 0.2991627 0.3147093 0.2318546 0.2931808 0.2859942
#> 4:lag1  0.3436778 0.3125853 0.2755298 0.3224450 0.2525857 0.3183020 0.3175358
#> 5:lag1  0.3125853 0.3367155 0.2536100 0.2952790 0.2689623 0.2921064 0.3261806
#> 6:lag1  0.2755298 0.2536100 0.3649523 0.2905571 0.2028941 0.2845886 0.2545886
#> 7:lag1  0.3224450 0.2952790 0.2905571 0.3400334 0.2371729 0.3166409 0.2979354
#> 8:lag1  0.2525857 0.2689623 0.2028941 0.2371729 0.3280766 0.2435704 0.2763687
#> 9:lag1  0.3183020 0.2921064 0.2845886 0.3166409 0.2435704 0.3578714 0.3036173
#> 10:lag1 0.3175358 0.3261806 0.2545886 0.2979354 0.2763687 0.3036173 0.3477625
#> 1:lag2  0.5057751 0.4610795 0.4966779 0.5164236 0.3817133 0.5608374 0.4767071
#> 2:lag2  0.5581796 0.5101605 0.4707611 0.5448013 0.4177323 0.5599320 0.5250900
#> 3:lag2  0.5130517 0.4872020 0.5030805 0.5292240 0.3898931 0.4930211 0.4809359
#> 4:lag2  0.5779383 0.5256522 0.4633386 0.5422326 0.4247552 0.5352656 0.5339771
#> 5:lag2  0.5256522 0.5662302 0.4264777 0.4965495 0.4522946 0.4912144 0.5485144
#> 6:lag2  0.4633386 0.4264777 0.6137141 0.4886090 0.3411924 0.4785721 0.4281234
#> 7:lag2  0.5422326 0.4965495 0.4886090 0.5718097 0.3988366 0.5324723 0.5010166
#> 8:lag2  0.4247552 0.4522946 0.3411924 0.3988366 0.5517028 0.4095947 0.4647494
#> 9:lag2  0.5352656 0.4912144 0.4785721 0.5324723 0.4095947 0.6018067 0.5105713
#> 10:lag2 0.5339771 0.5485144 0.4281234 0.5010166 0.4647494 0.5105713 0.5848071
#> 1:lag3  0.8231584 0.7504156 0.8083526 0.8404890 0.6212456 0.9127733 0.7758497
#> 2:lag3  0.9084478 0.8302957 0.7661725 0.8866743 0.6798672 0.9112998 0.8545939
#> 3:lag3  0.8350013 0.7929304 0.8187729 0.8613219 0.6345584 0.8024010 0.7827321
#> 4:lag3  0.9406054 0.8555088 0.7540922 0.8824937 0.6912971 0.8711549 0.8690578
#> 5:lag3  0.8555088 0.9215503 0.6941004 0.8081437 0.7361179 0.7994606 0.8927175
#> 6:lag3  0.7540922 0.6941004 0.9988312 0.7952203 0.5552971 0.7788851 0.6967788
#> 7:lag3  0.8824937 0.8081437 0.7952203 0.9306310 0.6491141 0.8666086 0.8154139
#> 8:lag3  0.6912971 0.7361179 0.5552971 0.6491141 0.8979066 0.6666231 0.7563884
#> 9:lag3  0.8711549 0.7994606 0.7788851 0.8666086 0.6666231 0.9794516 0.8309644
#> 10:lag3 0.8690578 0.8927175 0.6967788 0.8154139 0.7563884 0.8309644 0.9517845
#> 1:lag4  0.5057751 0.4610795 0.4966779 0.5164236 0.3817133 0.5608374 0.4767071
#> 2:lag4  0.5581796 0.5101605 0.4707611 0.5448013 0.4177323 0.5599320 0.5250900
#> 3:lag4  0.5130517 0.4872020 0.5030805 0.5292240 0.3898931 0.4930211 0.4809359
#> 4:lag4  0.5779383 0.5256522 0.4633386 0.5422326 0.4247552 0.5352656 0.5339771
#> 5:lag4  0.5256522 0.5662302 0.4264777 0.4965495 0.4522946 0.4912144 0.5485144
#> 6:lag4  0.4633386 0.4264777 0.6137141 0.4886090 0.3411924 0.4785721 0.4281234
#> 7:lag4  0.5422326 0.4965495 0.4886090 0.5718097 0.3988366 0.5324723 0.5010166
#> 8:lag4  0.4247552 0.4522946 0.3411924 0.3988366 0.5517028 0.4095947 0.4647494
#> 9:lag4  0.5352656 0.4912144 0.4785721 0.5324723 0.4095947 0.6018067 0.5105713
#> 10:lag4 0.5339771 0.5485144 0.4281234 0.5010166 0.4647494 0.5105713 0.5848071
#> 1:lag5  0.3007651 0.2741864 0.2953554 0.3070974 0.2269903 0.3335086 0.2834795
#> 2:lag5  0.3319281 0.3033729 0.2799436 0.3239725 0.2484095 0.3329702 0.3122510
#> 3:lag5  0.3050923 0.2897204 0.2991627 0.3147093 0.2318546 0.2931808 0.2859942
#> 4:lag5  0.3436778 0.3125853 0.2755298 0.3224450 0.2525857 0.3183020 0.3175358
#> 5:lag5  0.3125853 0.3367155 0.2536100 0.2952790 0.2689623 0.2921064 0.3261806
#> 6:lag5  0.2755298 0.2536100 0.3649523 0.2905571 0.2028941 0.2845886 0.2545886
#> 7:lag5  0.3224450 0.2952790 0.2905571 0.3400334 0.2371729 0.3166409 0.2979354
#> 8:lag5  0.2525857 0.2689623 0.2028941 0.2371729 0.3280766 0.2435704 0.2763687
#> 9:lag5  0.3183020 0.2921064 0.2845886 0.3166409 0.2435704 0.3578714 0.3036173
#> 10:lag5 0.3175358 0.3261806 0.2545886 0.2979354 0.2763687 0.3036173 0.3477625
#>            1:lag4    2:lag4    3:lag4    4:lag4    5:lag4     6:lag4    7:lag4
#> 1:t+0   0.1674911 0.1478982 0.1356514 0.1417421 0.1292163 0.13919268 0.1447263
#> 2:t+0   0.1478982 0.1636418 0.1417914 0.1564284 0.1429711 0.13192955 0.1526791
#> 3:t+0   0.1356514 0.1417914 0.1643797 0.1437814 0.1365371 0.14098697 0.1483136
#> 4:t+0   0.1417421 0.1564284 0.1437814 0.1619657 0.1473126 0.12984942 0.1519592
#> 5:t+0   0.1292163 0.1429711 0.1365371 0.1473126 0.1586845 0.11951925 0.1391567
#> 6:t+0   0.1391927 0.1319296 0.1409870 0.1298494 0.1195192 0.17199175 0.1369314
#> 7:t+0   0.1447263 0.1526791 0.1483136 0.1519592 0.1391567 0.13693139 0.1602482
#> 8:t+0   0.1069742 0.1170684 0.1092665 0.1190365 0.1267544 0.09561828 0.1117729
#> 9:t+0   0.1571732 0.1569195 0.1381679 0.1500068 0.1376615 0.13411858 0.1492240
#> 10:t+0  0.1335959 0.1471551 0.1347810 0.1496457 0.1537197 0.11998045 0.1404086
#> 1:lag1  0.2343444 0.2069310 0.1897960 0.1983178 0.1807924 0.19475079 0.2024932
#> 2:lag1  0.2069310 0.2289587 0.1983868 0.2188660 0.2000374 0.18458861 0.2136203
#> 3:lag1  0.1897960 0.1983868 0.2299911 0.2011711 0.1910352 0.19726126 0.2075123
#> 4:lag1  0.1983178 0.2188660 0.2011711 0.2266135 0.2061118 0.18167821 0.2126131
#> 5:lag1  0.1807924 0.2000374 0.1910352 0.2061118 0.2220227 0.16722479 0.1947004
#> 6:lag1  0.1947508 0.1845886 0.1972613 0.1816782 0.1672248 0.24064145 0.1915869
#> 7:lag1  0.2024932 0.2136203 0.2075123 0.2126131 0.1947004 0.19158691 0.2242105
#> 8:lag1  0.1496724 0.1637957 0.1528797 0.1665494 0.1773478 0.13378387 0.1563865
#> 9:lag1  0.2199081 0.2195531 0.1933169 0.2098813 0.1926085 0.18765138 0.2087860
#> 10:lag1 0.1869201 0.2058914 0.1885782 0.2093761 0.2150762 0.16787008 0.1964520
#> 1:lag2  0.3554023 0.3138277 0.2878410 0.3007651 0.2741864 0.29535540 0.3070974
#> 2:lag2  0.3138277 0.3472345 0.3008697 0.3319281 0.3033729 0.27994364 0.3239725
#> 3:lag2  0.2878410 0.3008697 0.3488002 0.3050923 0.2897204 0.29916274 0.3147093
#> 4:lag2  0.3007651 0.3319281 0.3050923 0.3436778 0.3125853 0.27552977 0.3224450
#> 5:lag2  0.2741864 0.3033729 0.2897204 0.3125853 0.3367155 0.25360999 0.2952790
#> 6:lag2  0.2953554 0.2799436 0.2991627 0.2755298 0.2536100 0.36495233 0.2905571
#> 7:lag2  0.3070974 0.3239725 0.3147093 0.3224450 0.2952790 0.29055712 0.3400334
#> 8:lag2  0.2269903 0.2484095 0.2318546 0.2525857 0.2689623 0.20289411 0.2371729
#> 9:lag2  0.3335086 0.3329702 0.2931808 0.3183020 0.2921064 0.28458857 0.3166409
#> 10:lag2 0.2834795 0.3122510 0.2859942 0.3175358 0.3261806 0.25458862 0.2979354
#> 1:lag3  0.5976545 0.5277415 0.4840415 0.5057751 0.4610795 0.49667794 0.5164236
#> 2:lag3  0.5277415 0.5839192 0.5059509 0.5581796 0.5101605 0.47076108 0.5448013
#> 3:lag3  0.4840415 0.5059509 0.5865522 0.5130517 0.4872020 0.50308046 0.5292240
#> 4:lag3  0.5057751 0.5581796 0.5130517 0.5779383 0.5256522 0.46333860 0.5422326
#> 5:lag3  0.4610795 0.5101605 0.4872020 0.5256522 0.5662302 0.42647769 0.4965495
#> 6:lag3  0.4966779 0.4707611 0.5030805 0.4633386 0.4264777 0.61371409 0.4886090
#> 7:lag3  0.5164236 0.5448013 0.5292240 0.5422326 0.4965495 0.48860901 0.5718097
#> 8:lag3  0.3817133 0.4177323 0.3898931 0.4247552 0.4522946 0.34119243 0.3988366
#> 9:lag3  0.5608374 0.5599320 0.4930211 0.5352656 0.4912144 0.47857214 0.5324723
#> 10:lag3 0.4767071 0.5250900 0.4809359 0.5339771 0.5485144 0.42812338 0.5010166
#> 1:lag4  0.9726939 0.8589092 0.7877867 0.8231584 0.7504156 0.80835264 0.8404890
#> 2:lag4  0.8589092 0.9503394 0.8234446 0.9084478 0.8302957 0.76617248 0.8866743
#> 3:lag4  0.7877867 0.8234446 0.9546246 0.8350013 0.7929304 0.81877287 0.8613219
#> 4:lag4  0.8231584 0.9084478 0.8350013 0.9406054 0.8555088 0.75409224 0.8824937
#> 5:lag4  0.7504156 0.8302957 0.7929304 0.8555088 0.9215503 0.69410042 0.8081437
#> 6:lag4  0.8083526 0.7661725 0.8187729 0.7540922 0.6941004 0.99883117 0.7952203
#> 7:lag4  0.8404890 0.8866743 0.8613219 0.8824937 0.8081437 0.79522031 0.9306310
#> 8:lag4  0.6212456 0.6798672 0.6345584 0.6912971 0.7361179 0.55529707 0.6491141
#> 9:lag4  0.9127733 0.9112998 0.8024010 0.8711549 0.7994606 0.77888512 0.8666086
#> 10:lag4 0.7758497 0.8545939 0.7827321 0.8690578 0.8927175 0.69677882 0.8154139
#> 1:lag5  0.5976545 0.5277415 0.4840415 0.5057751 0.4610795 0.49667794 0.5164236
#> 2:lag5  0.5277415 0.5839192 0.5059509 0.5581796 0.5101605 0.47076108 0.5448013
#> 3:lag5  0.4840415 0.5059509 0.5865522 0.5130517 0.4872020 0.50308046 0.5292240
#> 4:lag5  0.5057751 0.5581796 0.5130517 0.5779383 0.5256522 0.46333860 0.5422326
#> 5:lag5  0.4610795 0.5101605 0.4872020 0.5256522 0.5662302 0.42647769 0.4965495
#> 6:lag5  0.4966779 0.4707611 0.5030805 0.4633386 0.4264777 0.61371409 0.4886090
#> 7:lag5  0.5164236 0.5448013 0.5292240 0.5422326 0.4965495 0.48860901 0.5718097
#> 8:lag5  0.3817133 0.4177323 0.3898931 0.4247552 0.4522946 0.34119243 0.3988366
#> 9:lag5  0.5608374 0.5599320 0.4930211 0.5352656 0.4912144 0.47857214 0.5324723
#> 10:lag5 0.4767071 0.5250900 0.4809359 0.5339771 0.5485144 0.42812338 0.5010166
#>             8:lag4    9:lag4   10:lag4     1:lag5     2:lag5     3:lag5
#> 1:t+0   0.10697415 0.1571732 0.1335959 0.12678596 0.11195468 0.10268419
#> 2:t+0   0.11706838 0.1569195 0.1471551 0.11195468 0.12387216 0.10733204
#> 3:t+0   0.10926652 0.1381679 0.1347810 0.10268419 0.10733204 0.12443072
#> 4:t+0   0.11903654 0.1500068 0.1496457 0.10729473 0.11841179 0.10883839
#> 5:t+0   0.12675437 0.1376615 0.1537197 0.09781305 0.10822504 0.10335466
#> 6:t+0   0.09561828 0.1341186 0.1199804 0.10536487 0.09986689 0.10672310
#> 7:t+0   0.11177291 0.1492240 0.1404086 0.10955369 0.11557372 0.11226916
#> 8:t+0   0.15461324 0.1147878 0.1302448 0.08097637 0.08861741 0.08271163
#> 9:t+0   0.11478785 0.1686547 0.1430863 0.11897560 0.11878354 0.10458911
#> 10:t+0  0.13024480 0.1430863 0.1638906 0.10112827 0.11139219 0.10202536
#> 1:lag1  0.14967239 0.2199081 0.1869201 0.16749110 0.14789817 0.13565136
#> 2:lag1  0.16379567 0.2195531 0.2058914 0.14789817 0.16364182 0.14179142
#> 3:lag1  0.15287974 0.1933169 0.1885782 0.13565136 0.14179142 0.16437970
#> 4:lag1  0.16654941 0.2098813 0.2093761 0.14174214 0.15642837 0.14378140
#> 5:lag1  0.17734778 0.1926085 0.2150762 0.12921632 0.14297113 0.13653708
#> 6:lag1  0.13378387 0.1876514 0.1678701 0.13919268 0.13192955 0.14098697
#> 7:lag1  0.15638654 0.2087860 0.1964520 0.14472635 0.15267912 0.14831362
#> 8:lag1  0.21632640 0.1606049 0.1822314 0.10697415 0.11706838 0.10926652
#> 9:lag1  0.16060488 0.2359725 0.2001985 0.15717319 0.15691947 0.13816785
#> 10:lag1 0.18223141 0.2001985 0.2293068 0.13359591 0.14715510 0.13478101
#> 1:lag2  0.22699034 0.3335086 0.2834795 0.23434438 0.20693103 0.18979597
#> 2:lag2  0.24840945 0.3329702 0.3122510 0.20693103 0.22895868 0.19838679
#> 3:lag2  0.23185456 0.2931808 0.2859942 0.18979597 0.19838679 0.22999108
#> 4:lag2  0.25258572 0.3183020 0.3175358 0.19831784 0.21886602 0.20117106
#> 5:lag2  0.26896232 0.2921064 0.3261806 0.18079241 0.20003738 0.19103521
#> 6:lag2  0.20289411 0.2845886 0.2545886 0.19475079 0.18458861 0.19726126
#> 7:lag2  0.23717291 0.3166409 0.2979354 0.20249319 0.21362028 0.20751231
#> 8:lag2  0.32807656 0.2435704 0.2763687 0.14967239 0.16379567 0.15287974
#> 9:lag2  0.24357035 0.3578714 0.3036173 0.21990814 0.21955313 0.19331691
#> 10:lag2 0.27636874 0.3036173 0.3477625 0.18692009 0.20589136 0.18857822
#> 1:lag3  0.38171333 0.5608374 0.4767071 0.35540230 0.31382773 0.28784102
#> 2:lag3  0.41773232 0.5599320 0.5250900 0.31382773 0.34723445 0.30086970
#> 3:lag3  0.38989314 0.4930211 0.4809359 0.28784102 0.30086970 0.34880017
#> 4:lag3  0.42475524 0.5352656 0.5339771 0.30076513 0.33192811 0.30509227
#> 5:lag3  0.45229460 0.4912144 0.5485144 0.27418638 0.30337295 0.28972043
#> 6:lag3  0.34119243 0.4785721 0.4281234 0.29535540 0.27994364 0.29916274
#> 7:lag3  0.39883662 0.5324723 0.5010166 0.30709738 0.32397252 0.31470928
#> 8:lag3  0.55170277 0.4095947 0.4647494 0.22699034 0.24840945 0.23185456
#> 9:lag3  0.40959475 0.6018067 0.5105713 0.33350856 0.33297017 0.29318080
#> 10:lag3 0.46474944 0.5105713 0.5848071 0.28347951 0.31225098 0.28599420
#> 1:lag4  0.62124560 0.9127733 0.7758497 0.59765450 0.52774153 0.48404154
#> 2:lag4  0.67986717 0.9112998 0.8545939 0.52774153 0.58391921 0.50595094
#> 3:lag4  0.63455838 0.8024010 0.7827321 0.48404154 0.50595094 0.58655217
#> 4:lag4  0.69129711 0.8711549 0.8690578 0.50577509 0.55817963 0.51305174
#> 5:lag4  0.73611793 0.7994606 0.8927175 0.46107952 0.51016047 0.48720202
#> 6:lag4  0.55529707 0.7788851 0.6967788 0.49667794 0.47076108 0.50308046
#> 7:lag4  0.64911407 0.8666086 0.8154139 0.51642358 0.54480128 0.52922397
#> 8:lag4  0.89790658 0.6666231 0.7563884 0.38171333 0.41773232 0.38989314
#> 9:lag4  0.66662312 0.9794516 0.8309644 0.56083737 0.55993200 0.49302107
#> 10:lag4 0.75638841 0.8309644 0.9517845 0.47670710 0.52509003 0.48093587
#> 1:lag5  0.38171333 0.5608374 0.4767071 0.97269389 0.85890922 0.78778667
#> 2:lag5  0.41773232 0.5599320 0.5250900 0.85890922 0.95033944 0.82344463
#> 3:lag5  0.38989314 0.4930211 0.4809359 0.78778667 0.82344463 0.95462464
#> 4:lag5  0.42475524 0.5352656 0.5339771 0.82315843 0.90844780 0.83500131
#> 5:lag5  0.45229460 0.4912144 0.5485144 0.75041556 0.83029571 0.79293041
#> 6:lag5  0.34119243 0.4785721 0.4281234 0.80835264 0.76617248 0.81877287
#> 7:lag5  0.39883662 0.5324723 0.5010166 0.84048905 0.88667429 0.86132193
#> 8:lag5  0.55170277 0.4095947 0.4647494 0.62124560 0.67986717 0.63455838
#> 9:lag5  0.40959475 0.6018067 0.5105713 0.91277332 0.91129982 0.80240103
#> 10:lag5 0.46474944 0.5105713 0.5848071 0.77584973 0.85459385 0.78273214
#>             4:lag5     5:lag5     6:lag5     7:lag5     8:lag5     9:lag5
#> 1:t+0   0.10729473 0.09781305 0.10536487 0.10955369 0.08097637 0.11897560
#> 2:t+0   0.11841179 0.10822504 0.09986689 0.11557372 0.08861741 0.11878354
#> 3:t+0   0.10883839 0.10335466 0.10672310 0.11226916 0.08271163 0.10458911
#> 4:t+0   0.12260338 0.11151145 0.09829229 0.11502879 0.09010724 0.11355084
#> 5:t+0   0.11151145 0.12011964 0.09047264 0.10533763 0.09594942 0.10420583
#> 6:t+0   0.09829229 0.09047264 0.13019283 0.10365313 0.07238030 0.10152392
#> 7:t+0   0.11502879 0.10533763 0.10365313 0.12130326 0.08460889 0.11295826
#> 8:t+0   0.09010724 0.09594942 0.07238030 0.08460889 0.11703779 0.08689111
#> 9:t+0   0.11355084 0.10420583 0.10152392 0.11295826 0.08689111 0.12766680
#> 10:t+0  0.11327749 0.11636142 0.09082176 0.10628527 0.09859158 0.10831220
#> 1:lag1  0.14174214 0.12921632 0.13919268 0.14472635 0.10697415 0.15717319
#> 2:lag1  0.15642837 0.14297113 0.13192955 0.15267912 0.11706838 0.15691947
#> 3:lag1  0.14378140 0.13653708 0.14098697 0.14831362 0.10926652 0.13816785
#> 4:lag1  0.16196568 0.14731264 0.12984942 0.15195925 0.11903654 0.15000678
#> 5:lag1  0.14731264 0.15868453 0.11951925 0.13915670 0.12675437 0.13766153
#> 6:lag1  0.12984942 0.11951925 0.17199175 0.13693139 0.09561828 0.13411858
#> 7:lag1  0.15195925 0.13915670 0.13693139 0.16024816 0.11177291 0.14922396
#> 8:lag1  0.11903654 0.12675437 0.09561828 0.11177291 0.15461324 0.11478785
#> 9:lag1  0.15000678 0.13766153 0.13411858 0.14922396 0.11478785 0.16865473
#> 10:lag1 0.14964568 0.15371971 0.11998045 0.14040858 0.13024480 0.14308626
#> 1:lag2  0.19831784 0.18079241 0.19475079 0.20249319 0.14967239 0.21990814
#> 2:lag2  0.21886602 0.20003738 0.18458861 0.21362028 0.16379567 0.21955313
#> 3:lag2  0.20117106 0.19103521 0.19726126 0.20751231 0.15287974 0.19331691
#> 4:lag2  0.22661352 0.20611179 0.18167821 0.21261307 0.16654941 0.20988129
#> 5:lag2  0.20611179 0.22202272 0.16722479 0.19470044 0.17734778 0.19260849
#> 6:lag2  0.18167821 0.16722479 0.24064145 0.19158691 0.13378387 0.18765138
#> 7:lag2  0.21261307 0.19470044 0.19158691 0.22421046 0.15638654 0.20878600
#> 8:lag2  0.16654941 0.17734778 0.13378387 0.15638654 0.21632640 0.16060488
#> 9:lag2  0.20988129 0.19260849 0.18765138 0.20878600 0.16060488 0.23597247
#> 10:lag2 0.20937605 0.21507622 0.16787008 0.19645201 0.18223141 0.20019847
#> 1:lag3  0.30076513 0.27418638 0.29535540 0.30709738 0.22699034 0.33350856
#> 2:lag3  0.33192811 0.30337295 0.27994364 0.32397252 0.24840945 0.33297017
#> 3:lag3  0.30509227 0.28972043 0.29916274 0.31470928 0.23185456 0.29318080
#> 4:lag3  0.34367783 0.31258529 0.27552977 0.32244500 0.25258572 0.31830204
#> 5:lag3  0.31258529 0.33671550 0.25360999 0.29527905 0.26896232 0.29210643
#> 6:lag3  0.27552977 0.25360999 0.36495233 0.29055712 0.20289411 0.28458857
#> 7:lag3  0.32244500 0.29527905 0.29055712 0.34003338 0.23717291 0.31664094
#> 8:lag3  0.25258572 0.26896232 0.20289411 0.23717291 0.32807656 0.24357035
#> 9:lag3  0.31830204 0.29210643 0.28458857 0.31664094 0.24357035 0.35787144
#> 10:lag3 0.31753581 0.32618056 0.25458862 0.29793545 0.27636874 0.30361726
#> 1:lag4  0.50577509 0.46107952 0.49667794 0.51642358 0.38171333 0.56083737
#> 2:lag4  0.55817963 0.51016047 0.47076108 0.54480128 0.41773232 0.55993200
#> 3:lag4  0.51305174 0.48720202 0.50308046 0.52922397 0.38989314 0.49302107
#> 4:lag4  0.57793829 0.52565220 0.46333860 0.54223257 0.42475524 0.53526565
#> 5:lag4  0.52565220 0.56623024 0.42647769 0.49654954 0.45229460 0.49121438
#> 6:lag4  0.46333860 0.42647769 0.61371409 0.48860901 0.34119243 0.47857214
#> 7:lag4  0.54223257 0.49654954 0.48860901 0.57180969 0.39883662 0.53247230
#> 8:lag4  0.42475524 0.45229460 0.34119243 0.39883662 0.55170277 0.40959475
#> 9:lag4  0.53526565 0.49121438 0.47857214 0.53247230 0.40959475 0.60180667
#> 10:lag4 0.53397713 0.54851440 0.42812338 0.50101662 0.46474944 0.51057131
#> 1:lag5  0.82315843 0.75041556 0.80835264 0.84048905 0.62124560 0.91277332
#> 2:lag5  0.90844780 0.83029571 0.76617248 0.88667429 0.67986717 0.91129982
#> 3:lag5  0.83500131 0.79293041 0.81877287 0.86132193 0.63455838 0.80240103
#> 4:lag5  0.94060539 0.85550880 0.75409224 0.88249365 0.69129711 0.87115486
#> 5:lag5  0.85550880 0.92155031 0.69410042 0.80814368 0.73611793 0.79946060
#> 6:lag5  0.75409224 0.69410042 0.99883117 0.79522031 0.55529707 0.77888512
#> 7:lag5  0.88249365 0.80814368 0.79522031 0.93063097 0.64911407 0.86660864
#> 8:lag5  0.69129711 0.73611793 0.55529707 0.64911407 0.89790658 0.66662312
#> 9:lag5  0.87115486 0.79946060 0.77888512 0.86660864 0.66662312 0.97945162
#> 10:lag5 0.86905778 0.89271746 0.69677882 0.81541392 0.75638841 0.83096437
#>            10:lag5
#> 1:t+0   0.10112827
#> 2:t+0   0.11139219
#> 3:t+0   0.10202536
#> 4:t+0   0.11327749
#> 5:t+0   0.11636142
#> 6:t+0   0.09082176
#> 7:t+0   0.10628527
#> 8:t+0   0.09859158
#> 9:t+0   0.10831220
#> 10:t+0  0.12406052
#> 1:lag1  0.13359591
#> 2:lag1  0.14715510
#> 3:lag1  0.13478101
#> 4:lag1  0.14964568
#> 5:lag1  0.15371971
#> 6:lag1  0.11998045
#> 7:lag1  0.14040858
#> 8:lag1  0.13024480
#> 9:lag1  0.14308626
#> 10:lag1 0.16389065
#> 1:lag2  0.18692009
#> 2:lag2  0.20589136
#> 3:lag2  0.18857822
#> 4:lag2  0.20937605
#> 5:lag2  0.21507622
#> 6:lag2  0.16787008
#> 7:lag2  0.19645201
#> 8:lag2  0.18223141
#> 9:lag2  0.20019847
#> 10:lag2 0.22930683
#> 1:lag3  0.28347951
#> 2:lag3  0.31225098
#> 3:lag3  0.28599420
#> 4:lag3  0.31753581
#> 5:lag3  0.32618056
#> 6:lag3  0.25458862
#> 7:lag3  0.29793545
#> 8:lag3  0.27636874
#> 9:lag3  0.30361726
#> 10:lag3 0.34776245
#> 1:lag4  0.47670710
#> 2:lag4  0.52509003
#> 3:lag4  0.48093587
#> 4:lag4  0.53397713
#> 5:lag4  0.54851440
#> 6:lag4  0.42812338
#> 7:lag4  0.50101662
#> 8:lag4  0.46474944
#> 9:lag4  0.51057131
#> 10:lag4 0.58480711
#> 1:lag5  0.77584973
#> 2:lag5  0.85459385
#> 3:lag5  0.78273214
#> 4:lag5  0.86905778
#> 5:lag5  0.89271746
#> 6:lag5  0.69677882
#> 7:lag5  0.81541392
#> 8:lag5  0.75638841
#> 9:lag5  0.83096437
#> 10:lag5 0.95178452