Covariance/correlation for joint distribution of an mcgf object

Covariance/correlation for joint distribution of an mcgf object

Usage

# S3 method for mcgf
ccov(x, model = c("all", "base", "empirical"), cor = FALSE, ...)

Arguments

x

An mcgf object.

model

Which model to use. One of all, base, and empirical.

cor

Logical; if TRUE, correlation is outputted.

...

Additional arguments. Not in use.

Value

Joint covariance/correlation matrix.

Examples

data(sim1)
sim1_mcgf <- mcgf(sim1$data, dists = sim1$dists)
#> `time` is not provided, assuming rows are equally spaced temporally.
sim1_mcgf <- add_acfs(sim1_mcgf, lag_max = 5)
sim1_mcgf <- add_ccfs(sim1_mcgf, lag_max = 5)

# Fit a separable model and store it to 'sim1_mcgf'
fit_sep <- fit_base(
    sim1_mcgf,
    model = "sep",
    lag = 5,
    par_init = c(
        c = 0.001,
        gamma = 0.5,
        a = 0.3,
        alpha = 0.5
    ),
    par_fixed = c(nugget = 0)
)
sim1_mcgf <- add_base(sim1_mcgf, fit_base = fit_sep)

ccov(sim1_mcgf, model = "base")
#>              1:t+0      2:t+0      3:t+0      4:t+0      5:t+0      6:t+0
#> 1:t+0   0.97269380 0.85890916 0.78778712 0.82315831 0.75041561 0.80835281
#> 2:t+0   0.85890916 0.95033925 0.82344690 0.90844760 0.83029569 0.76617397
#> 3:t+0   0.78778712 0.82344690 0.95462554 0.83500450 0.79293293 0.81877354
#> 4:t+0   0.82315831 0.90844760 0.83500450 0.94060499 0.85550863 0.75409404
#> 5:t+0   0.75041561 0.83029569 0.79293293 0.85550863 0.92155023 0.69410189
#> 6:t+0   0.80835281 0.76617397 0.81877354 0.75409404 0.69410189 0.99883141
#> 7:t+0   0.84048909 0.88667461 0.86132611 0.88249374 0.80814380 0.79522277
#> 8:t+0   0.62124582 0.67986724 0.63455595 0.69129689 0.73611760 0.55529521
#> 9:t+0   0.91277334 0.91129976 0.80240066 0.87115469 0.79946063 0.77888455
#> 10:t+0  0.77584953 0.85459351 0.78273290 0.86905736 0.89271719 0.69677919
#> 1:lag1  0.59765440 0.52774146 0.48404178 0.50577498 0.46107953 0.49667801
#> 2:lag1  0.52774146 0.58391905 0.50595230 0.55817947 0.51016042 0.47076197
#> 3:lag1  0.48404178 0.50595230 0.58655268 0.51305367 0.48720354 0.50308084
#> 4:lag1  0.50577498 0.55817947 0.51305367 0.57793801 0.52565207 0.46333967
#> 5:lag1  0.46107953 0.51016042 0.48720354 0.52565207 0.56623015 0.42647856
#> 6:lag1  0.49667801 0.47076197 0.50308084 0.46333967 0.42647856 0.61371419
#> 7:lag1  0.51642358 0.54480144 0.52922651 0.54223258 0.49654959 0.48861049
#> 8:lag1  0.38171345 0.41773233 0.38989162 0.42475508 0.45229437 0.34119127
#> 9:lag1  0.56083734 0.55993193 0.49302082 0.53526551 0.49121437 0.47857176
#> 10:lag1 0.47670694 0.52508979 0.48093631 0.53397684 0.54851420 0.42812358
#> 1:lag2  0.35540229 0.31382773 0.28784120 0.30076510 0.27418642 0.29535548
#> 2:lag2  0.31382773 0.34723440 0.30087055 0.33192806 0.30337296 0.27994420
#> 3:lag2  0.28784120 0.30087055 0.34880052 0.30509346 0.28972138 0.29916300
#> 4:lag2  0.30076510 0.33192806 0.30509346 0.34367771 0.31258525 0.27553044
#> 5:lag2  0.27418642 0.30337296 0.28972138 0.31258525 0.33671549 0.25361054
#> 6:lag2  0.29535548 0.27994420 0.29916300 0.27553044 0.25361054 0.36495244
#> 7:lag2  0.30709742 0.32397265 0.31471083 0.32244505 0.29527911 0.29055803
#> 8:lag2  0.22699044 0.24840950 0.23185368 0.25258566 0.26896222 0.20289344
#> 9:lag2  0.33350859 0.33297018 0.29318069 0.31830200 0.29210646 0.28458839
#> 10:lag2 0.28347945 0.31225088 0.28599449 0.31753567 0.32618049 0.25458878
#> 1:lag3  0.23434441 0.20693106 0.18979612 0.19831785 0.18079246 0.19475087
#> 2:lag3  0.20693106 0.22895868 0.19838738 0.21886601 0.20003741 0.18458901
#> 3:lag3  0.18979612 0.19838738 0.22999135 0.20117188 0.19103586 0.19726146
#> 4:lag3  0.19831785 0.21886601 0.20117188 0.22661347 0.20611179 0.18167868
#> 5:lag3  0.18079246 0.20003741 0.19103586 0.20611179 0.22202274 0.16722518
#> 6:lag3  0.19475087 0.18458901 0.19726146 0.18167868 0.16722518 0.24064156
#> 7:lag3  0.20249324 0.21362040 0.20751336 0.21261313 0.19470051 0.19158754
#> 8:lag3  0.14967247 0.16379572 0.15287919 0.16654939 0.17734774 0.13378344
#> 9:lag3  0.21990818 0.21955316 0.19331686 0.20988129 0.19260854 0.18765128
#> 10:lag3 0.18692008 0.20589132 0.18857844 0.20937600 0.21507619 0.16787020
#> 1:lag4  0.16749113 0.14789821 0.13565148 0.14174216 0.12921637 0.13919275
#> 2:lag4  0.14789821 0.16364184 0.14179185 0.15642838 0.14297117 0.13192985
#> 3:lag4  0.13565148 0.14179185 0.16437990 0.14378199 0.13653756 0.14098713
#> 4:lag4  0.14174216 0.15642838 0.14378199 0.16196566 0.14731266 0.12984977
#> 5:lag4  0.12921637 0.14297117 0.13653756 0.14731266 0.15868457 0.11951954
#> 6:lag4  0.13919275 0.13192985 0.14098713 0.12984977 0.11951954 0.17199185
#> 7:lag4  0.14472640 0.15267923 0.14831439 0.15195931 0.13915676 0.13693185
#> 8:lag4  0.10697423 0.11706843 0.10926614 0.11903654 0.12675435 0.09561799
#> 9:lag4  0.15717325 0.15691951 0.13816784 0.15000680 0.13766158 0.13411853
#> 10:lag4 0.13359591 0.14715508 0.13478118 0.14964566 0.15371972 0.11998055
#> 1:lag5  0.12678600 0.11195471 0.10268429 0.10729476 0.09781310 0.10536493
#> 2:lag5  0.11195471 0.12387219 0.10733238 0.11841181 0.10822508 0.09986712
#> 3:lag5  0.10268429 0.10733238 0.12443089 0.10883886 0.10335503 0.10672323
#> 4:lag5  0.10729476 0.11841181 0.10883886 0.12260338 0.11151147 0.09829256
#> 5:lag5  0.09781310 0.10822508 0.10335503 0.11151147 0.12011968 0.09047287
#> 6:lag5  0.10536493 0.09986712 0.10672323 0.09829256 0.09047287 0.13019291
#> 7:lag5  0.10955375 0.11557381 0.11226975 0.11502885 0.10533769 0.10365350
#> 8:lag5  0.08097643 0.08861745 0.08271134 0.09010725 0.09594942 0.07238008
#> 9:lag5  0.11897565 0.11878358 0.10458910 0.11355086 0.10420588 0.10152389
#> 10:lag5 0.10112829 0.11139219 0.10202550 0.11327748 0.11636143 0.09082185
#>              7:t+0      8:t+0      9:t+0     10:t+0    1:lag1    2:lag1
#> 1:t+0   0.84048909 0.62124582 0.91277334 0.77584953 0.5976544 0.5277415
#> 2:t+0   0.88667461 0.67986724 0.91129976 0.85459351 0.5277415 0.5839190
#> 3:t+0   0.86132611 0.63455595 0.80240066 0.78273290 0.4840418 0.5059523
#> 4:t+0   0.88249374 0.69129689 0.87115469 0.86905736 0.5057750 0.5581795
#> 5:t+0   0.80814380 0.73611760 0.79946063 0.89271719 0.4610795 0.5101604
#> 6:t+0   0.79522277 0.55529521 0.77888455 0.69677919 0.4966780 0.4707620
#> 7:t+0   0.93063100 0.64911371 0.86660850 0.81541380 0.5164236 0.5448014
#> 8:t+0   0.64911371 0.89790679 0.66662338 0.75638830 0.3817134 0.4177323
#> 9:t+0   0.86660850 0.66662338 0.97945166 0.83096417 0.5608373 0.5599319
#> 10:t+0  0.81541380 0.75638830 0.83096417 0.95178382 0.4767069 0.5250898
#> 1:lag1  0.51642358 0.38171345 0.56083734 0.47670694 0.9726938 0.8589092
#> 2:lag1  0.54480144 0.41773233 0.55993193 0.52508979 0.8589092 0.9503392
#> 3:lag1  0.52922651 0.38989162 0.49302082 0.48093631 0.7877871 0.8234469
#> 4:lag1  0.54223258 0.42475508 0.53526551 0.53397684 0.8231583 0.9084476
#> 5:lag1  0.49654959 0.45229437 0.49121437 0.54851420 0.7504156 0.8302957
#> 6:lag1  0.48861049 0.34119127 0.47857176 0.42812358 0.8083528 0.7661740
#> 7:lag1  0.57180967 0.39883637 0.53247218 0.50101651 0.8404891 0.8866746
#> 8:lag1  0.39883637 0.55170286 0.40959488 0.46474934 0.6212458 0.6798672
#> 9:lag1  0.53247218 0.40959488 0.60180665 0.51057115 0.9127733 0.9112998
#> 10:lag1 0.50101651 0.46474934 0.51057115 0.58480664 0.7758495 0.8545935
#> 1:lag2  0.30709742 0.22699044 0.33350859 0.28347945 0.5976544 0.5277415
#> 2:lag2  0.32397265 0.24840950 0.33297018 0.31225088 0.5277415 0.5839190
#> 3:lag2  0.31471083 0.23185368 0.29318069 0.28599449 0.4840418 0.5059523
#> 4:lag2  0.32244505 0.25258566 0.31830200 0.31753567 0.5057750 0.5581795
#> 5:lag2  0.29527911 0.26896222 0.29210646 0.32618049 0.4610795 0.5101604
#> 6:lag2  0.29055803 0.20289344 0.28458839 0.25458878 0.4966780 0.4707620
#> 7:lag2  0.34003341 0.23717279 0.31664091 0.29793542 0.5164236 0.5448014
#> 8:lag2  0.23717279 0.32807666 0.24357046 0.27636872 0.3817134 0.4177323
#> 9:lag2  0.31664091 0.24357046 0.35787148 0.30361720 0.5608373 0.5599319
#> 10:lag2 0.29793542 0.27636872 0.30361720 0.34776222 0.4767069 0.5250898
#> 1:lag3  0.20249324 0.14967247 0.21990818 0.18692008 0.3554023 0.3138277
#> 2:lag3  0.21362040 0.16379572 0.21955316 0.20589132 0.3138277 0.3472344
#> 3:lag3  0.20751336 0.15287919 0.19331686 0.18857844 0.2878412 0.3008706
#> 4:lag3  0.21261313 0.16654939 0.20988129 0.20937600 0.3007651 0.3319281
#> 5:lag3  0.19470051 0.17734774 0.19260854 0.21507619 0.2741864 0.3033730
#> 6:lag3  0.19158754 0.13378344 0.18765128 0.16787020 0.2953555 0.2799442
#> 7:lag3  0.22421051 0.15638649 0.20878601 0.19645202 0.3070974 0.3239727
#> 8:lag3  0.15638649 0.21632649 0.16060497 0.18223142 0.2269904 0.2484095
#> 9:lag3  0.20878601 0.16060497 0.23597253 0.20019847 0.3335086 0.3329702
#> 10:lag3 0.19645202 0.18223142 0.20019847 0.22930671 0.2834794 0.3122509
#> 1:lag4  0.14472640 0.10697423 0.15717325 0.13359591 0.2343444 0.2069311
#> 2:lag4  0.15267923 0.11706843 0.15691951 0.14715508 0.2069311 0.2289587
#> 3:lag4  0.14831439 0.10926614 0.13816784 0.13478118 0.1897961 0.1983874
#> 4:lag4  0.15195931 0.11903654 0.15000680 0.14964566 0.1983179 0.2188660
#> 5:lag4  0.13915676 0.12675435 0.13766158 0.15371972 0.1807925 0.2000374
#> 6:lag4  0.13693185 0.09561799 0.13411853 0.11998055 0.1947509 0.1845890
#> 7:lag4  0.16024821 0.11177288 0.14922398 0.14040861 0.2024932 0.2136204
#> 8:lag4  0.11177288 0.15461333 0.11478793 0.13024483 0.1496725 0.1637957
#> 9:lag4  0.14922398 0.11478793 0.16865479 0.14308627 0.2199082 0.2195532
#> 10:lag4 0.14040861 0.13024483 0.14308627 0.16389058 0.1869201 0.2058913
#> 1:lag5  0.10955375 0.08097643 0.11897565 0.10112829 0.1674911 0.1478982
#> 2:lag5  0.11557381 0.08861745 0.11878358 0.11139219 0.1478982 0.1636418
#> 3:lag5  0.11226975 0.08271134 0.10458910 0.10202550 0.1356515 0.1417919
#> 4:lag5  0.11502885 0.09010725 0.11355086 0.11327748 0.1417422 0.1564284
#> 5:lag5  0.10533769 0.09594942 0.10420588 0.11636143 0.1292164 0.1429712
#> 6:lag5  0.10365350 0.07238008 0.10152389 0.09082185 0.1391928 0.1319298
#> 7:lag5  0.12130331 0.08460888 0.11295829 0.10628530 0.1447264 0.1526792
#> 8:lag5  0.08460888 0.11703787 0.08689118 0.09859161 0.1069742 0.1170684
#> 9:lag5  0.11295829 0.08689118 0.12766685 0.10831222 0.1571732 0.1569195
#> 10:lag5 0.10628530 0.09859161 0.10831222 0.12406048 0.1335959 0.1471551
#>            3:lag1    4:lag1    5:lag1     6:lag1    7:lag1     8:lag1    9:lag1
#> 1:t+0   0.4840418 0.5057750 0.4610795 0.49667801 0.5164236 0.38171345 0.5608373
#> 2:t+0   0.5059523 0.5581795 0.5101604 0.47076197 0.5448014 0.41773233 0.5599319
#> 3:t+0   0.5865527 0.5130537 0.4872035 0.50308084 0.5292265 0.38989162 0.4930208
#> 4:t+0   0.5130537 0.5779380 0.5256521 0.46333967 0.5422326 0.42475508 0.5352655
#> 5:t+0   0.4872035 0.5256521 0.5662302 0.42647856 0.4965496 0.45229437 0.4912144
#> 6:t+0   0.5030808 0.4633397 0.4264786 0.61371419 0.4886105 0.34119127 0.4785718
#> 7:t+0   0.5292265 0.5422326 0.4965496 0.48861049 0.5718097 0.39883637 0.5324722
#> 8:t+0   0.3898916 0.4247551 0.4522944 0.34119127 0.3988364 0.55170286 0.4095949
#> 9:t+0   0.4930208 0.5352655 0.4912144 0.47857176 0.5324722 0.40959488 0.6018067
#> 10:t+0  0.4809363 0.5339768 0.5485142 0.42812358 0.5010165 0.46474934 0.5105712
#> 1:lag1  0.7877871 0.8231583 0.7504156 0.80835281 0.8404891 0.62124582 0.9127733
#> 2:lag1  0.8234469 0.9084476 0.8302957 0.76617397 0.8866746 0.67986724 0.9112998
#> 3:lag1  0.9546255 0.8350045 0.7929329 0.81877354 0.8613261 0.63455595 0.8024007
#> 4:lag1  0.8350045 0.9406050 0.8555086 0.75409404 0.8824937 0.69129689 0.8711547
#> 5:lag1  0.7929329 0.8555086 0.9215502 0.69410189 0.8081438 0.73611760 0.7994606
#> 6:lag1  0.8187735 0.7540940 0.6941019 0.99883141 0.7952228 0.55529521 0.7788846
#> 7:lag1  0.8613261 0.8824937 0.8081438 0.79522277 0.9306310 0.64911371 0.8666085
#> 8:lag1  0.6345559 0.6912969 0.7361176 0.55529521 0.6491137 0.89790679 0.6666234
#> 9:lag1  0.8024007 0.8711547 0.7994606 0.77888455 0.8666085 0.66662338 0.9794517
#> 10:lag1 0.7827329 0.8690574 0.8927172 0.69677919 0.8154138 0.75638830 0.8309642
#> 1:lag2  0.4840418 0.5057750 0.4610795 0.49667801 0.5164236 0.38171345 0.5608373
#> 2:lag2  0.5059523 0.5581795 0.5101604 0.47076197 0.5448014 0.41773233 0.5599319
#> 3:lag2  0.5865527 0.5130537 0.4872035 0.50308084 0.5292265 0.38989162 0.4930208
#> 4:lag2  0.5130537 0.5779380 0.5256521 0.46333967 0.5422326 0.42475508 0.5352655
#> 5:lag2  0.4872035 0.5256521 0.5662302 0.42647856 0.4965496 0.45229437 0.4912144
#> 6:lag2  0.5030808 0.4633397 0.4264786 0.61371419 0.4886105 0.34119127 0.4785718
#> 7:lag2  0.5292265 0.5422326 0.4965496 0.48861049 0.5718097 0.39883637 0.5324722
#> 8:lag2  0.3898916 0.4247551 0.4522944 0.34119127 0.3988364 0.55170286 0.4095949
#> 9:lag2  0.4930208 0.5352655 0.4912144 0.47857176 0.5324722 0.40959488 0.6018067
#> 10:lag2 0.4809363 0.5339768 0.5485142 0.42812358 0.5010165 0.46474934 0.5105712
#> 1:lag3  0.2878412 0.3007651 0.2741864 0.29535548 0.3070974 0.22699044 0.3335086
#> 2:lag3  0.3008706 0.3319281 0.3033730 0.27994420 0.3239727 0.24840950 0.3329702
#> 3:lag3  0.3488005 0.3050935 0.2897214 0.29916300 0.3147108 0.23185368 0.2931807
#> 4:lag3  0.3050935 0.3436777 0.3125852 0.27553044 0.3224450 0.25258566 0.3183020
#> 5:lag3  0.2897214 0.3125852 0.3367155 0.25361054 0.2952791 0.26896222 0.2921065
#> 6:lag3  0.2991630 0.2755304 0.2536105 0.36495244 0.2905580 0.20289344 0.2845884
#> 7:lag3  0.3147108 0.3224450 0.2952791 0.29055803 0.3400334 0.23717279 0.3166409
#> 8:lag3  0.2318537 0.2525857 0.2689622 0.20289344 0.2371728 0.32807666 0.2435705
#> 9:lag3  0.2931807 0.3183020 0.2921065 0.28458839 0.3166409 0.24357046 0.3578715
#> 10:lag3 0.2859945 0.3175357 0.3261805 0.25458878 0.2979354 0.27636872 0.3036172
#> 1:lag4  0.1897961 0.1983179 0.1807925 0.19475087 0.2024932 0.14967247 0.2199082
#> 2:lag4  0.1983874 0.2188660 0.2000374 0.18458901 0.2136204 0.16379572 0.2195532
#> 3:lag4  0.2299913 0.2011719 0.1910359 0.19726146 0.2075134 0.15287919 0.1933169
#> 4:lag4  0.2011719 0.2266135 0.2061118 0.18167868 0.2126131 0.16654939 0.2098813
#> 5:lag4  0.1910359 0.2061118 0.2220227 0.16722518 0.1947005 0.17734774 0.1926085
#> 6:lag4  0.1972615 0.1816787 0.1672252 0.24064156 0.1915875 0.13378344 0.1876513
#> 7:lag4  0.2075134 0.2126131 0.1947005 0.19158754 0.2242105 0.15638649 0.2087860
#> 8:lag4  0.1528792 0.1665494 0.1773477 0.13378344 0.1563865 0.21632649 0.1606050
#> 9:lag4  0.1933169 0.2098813 0.1926085 0.18765128 0.2087860 0.16060497 0.2359725
#> 10:lag4 0.1885784 0.2093760 0.2150762 0.16787020 0.1964520 0.18223142 0.2001985
#> 1:lag5  0.1356515 0.1417422 0.1292164 0.13919275 0.1447264 0.10697423 0.1571732
#> 2:lag5  0.1417919 0.1564284 0.1429712 0.13192985 0.1526792 0.11706843 0.1569195
#> 3:lag5  0.1643799 0.1437820 0.1365376 0.14098713 0.1483144 0.10926614 0.1381678
#> 4:lag5  0.1437820 0.1619657 0.1473127 0.12984977 0.1519593 0.11903654 0.1500068
#> 5:lag5  0.1365376 0.1473127 0.1586846 0.11951954 0.1391568 0.12675435 0.1376616
#> 6:lag5  0.1409871 0.1298498 0.1195195 0.17199185 0.1369319 0.09561799 0.1341185
#> 7:lag5  0.1483144 0.1519593 0.1391568 0.13693185 0.1602482 0.11177288 0.1492240
#> 8:lag5  0.1092661 0.1190365 0.1267544 0.09561799 0.1117729 0.15461333 0.1147879
#> 9:lag5  0.1381678 0.1500068 0.1376616 0.13411853 0.1492240 0.11478793 0.1686548
#> 10:lag5 0.1347812 0.1496457 0.1537197 0.11998055 0.1404086 0.13024483 0.1430863
#>           10:lag1    1:lag2    2:lag2    3:lag2    4:lag2    5:lag2    6:lag2
#> 1:t+0   0.4767069 0.3554023 0.3138277 0.2878412 0.3007651 0.2741864 0.2953555
#> 2:t+0   0.5250898 0.3138277 0.3472344 0.3008706 0.3319281 0.3033730 0.2799442
#> 3:t+0   0.4809363 0.2878412 0.3008706 0.3488005 0.3050935 0.2897214 0.2991630
#> 4:t+0   0.5339768 0.3007651 0.3319281 0.3050935 0.3436777 0.3125852 0.2755304
#> 5:t+0   0.5485142 0.2741864 0.3033730 0.2897214 0.3125852 0.3367155 0.2536105
#> 6:t+0   0.4281236 0.2953555 0.2799442 0.2991630 0.2755304 0.2536105 0.3649524
#> 7:t+0   0.5010165 0.3070974 0.3239727 0.3147108 0.3224450 0.2952791 0.2905580
#> 8:t+0   0.4647493 0.2269904 0.2484095 0.2318537 0.2525857 0.2689622 0.2028934
#> 9:t+0   0.5105712 0.3335086 0.3329702 0.2931807 0.3183020 0.2921065 0.2845884
#> 10:t+0  0.5848066 0.2834794 0.3122509 0.2859945 0.3175357 0.3261805 0.2545888
#> 1:lag1  0.7758495 0.5976544 0.5277415 0.4840418 0.5057750 0.4610795 0.4966780
#> 2:lag1  0.8545935 0.5277415 0.5839190 0.5059523 0.5581795 0.5101604 0.4707620
#> 3:lag1  0.7827329 0.4840418 0.5059523 0.5865527 0.5130537 0.4872035 0.5030808
#> 4:lag1  0.8690574 0.5057750 0.5581795 0.5130537 0.5779380 0.5256521 0.4633397
#> 5:lag1  0.8927172 0.4610795 0.5101604 0.4872035 0.5256521 0.5662302 0.4264786
#> 6:lag1  0.6967792 0.4966780 0.4707620 0.5030808 0.4633397 0.4264786 0.6137142
#> 7:lag1  0.8154138 0.5164236 0.5448014 0.5292265 0.5422326 0.4965496 0.4886105
#> 8:lag1  0.7563883 0.3817134 0.4177323 0.3898916 0.4247551 0.4522944 0.3411913
#> 9:lag1  0.8309642 0.5608373 0.5599319 0.4930208 0.5352655 0.4912144 0.4785718
#> 10:lag1 0.9517838 0.4767069 0.5250898 0.4809363 0.5339768 0.5485142 0.4281236
#> 1:lag2  0.4767069 0.9726938 0.8589092 0.7877871 0.8231583 0.7504156 0.8083528
#> 2:lag2  0.5250898 0.8589092 0.9503392 0.8234469 0.9084476 0.8302957 0.7661740
#> 3:lag2  0.4809363 0.7877871 0.8234469 0.9546255 0.8350045 0.7929329 0.8187735
#> 4:lag2  0.5339768 0.8231583 0.9084476 0.8350045 0.9406050 0.8555086 0.7540940
#> 5:lag2  0.5485142 0.7504156 0.8302957 0.7929329 0.8555086 0.9215502 0.6941019
#> 6:lag2  0.4281236 0.8083528 0.7661740 0.8187735 0.7540940 0.6941019 0.9988314
#> 7:lag2  0.5010165 0.8404891 0.8866746 0.8613261 0.8824937 0.8081438 0.7952228
#> 8:lag2  0.4647493 0.6212458 0.6798672 0.6345559 0.6912969 0.7361176 0.5552952
#> 9:lag2  0.5105712 0.9127733 0.9112998 0.8024007 0.8711547 0.7994606 0.7788846
#> 10:lag2 0.5848066 0.7758495 0.8545935 0.7827329 0.8690574 0.8927172 0.6967792
#> 1:lag3  0.2834794 0.5976544 0.5277415 0.4840418 0.5057750 0.4610795 0.4966780
#> 2:lag3  0.3122509 0.5277415 0.5839190 0.5059523 0.5581795 0.5101604 0.4707620
#> 3:lag3  0.2859945 0.4840418 0.5059523 0.5865527 0.5130537 0.4872035 0.5030808
#> 4:lag3  0.3175357 0.5057750 0.5581795 0.5130537 0.5779380 0.5256521 0.4633397
#> 5:lag3  0.3261805 0.4610795 0.5101604 0.4872035 0.5256521 0.5662302 0.4264786
#> 6:lag3  0.2545888 0.4966780 0.4707620 0.5030808 0.4633397 0.4264786 0.6137142
#> 7:lag3  0.2979354 0.5164236 0.5448014 0.5292265 0.5422326 0.4965496 0.4886105
#> 8:lag3  0.2763687 0.3817134 0.4177323 0.3898916 0.4247551 0.4522944 0.3411913
#> 9:lag3  0.3036172 0.5608373 0.5599319 0.4930208 0.5352655 0.4912144 0.4785718
#> 10:lag3 0.3477622 0.4767069 0.5250898 0.4809363 0.5339768 0.5485142 0.4281236
#> 1:lag4  0.1869201 0.3554023 0.3138277 0.2878412 0.3007651 0.2741864 0.2953555
#> 2:lag4  0.2058913 0.3138277 0.3472344 0.3008706 0.3319281 0.3033730 0.2799442
#> 3:lag4  0.1885784 0.2878412 0.3008706 0.3488005 0.3050935 0.2897214 0.2991630
#> 4:lag4  0.2093760 0.3007651 0.3319281 0.3050935 0.3436777 0.3125852 0.2755304
#> 5:lag4  0.2150762 0.2741864 0.3033730 0.2897214 0.3125852 0.3367155 0.2536105
#> 6:lag4  0.1678702 0.2953555 0.2799442 0.2991630 0.2755304 0.2536105 0.3649524
#> 7:lag4  0.1964520 0.3070974 0.3239727 0.3147108 0.3224450 0.2952791 0.2905580
#> 8:lag4  0.1822314 0.2269904 0.2484095 0.2318537 0.2525857 0.2689622 0.2028934
#> 9:lag4  0.2001985 0.3335086 0.3329702 0.2931807 0.3183020 0.2921065 0.2845884
#> 10:lag4 0.2293067 0.2834794 0.3122509 0.2859945 0.3175357 0.3261805 0.2545888
#> 1:lag5  0.1335959 0.2343444 0.2069311 0.1897961 0.1983179 0.1807925 0.1947509
#> 2:lag5  0.1471551 0.2069311 0.2289587 0.1983874 0.2188660 0.2000374 0.1845890
#> 3:lag5  0.1347812 0.1897961 0.1983874 0.2299913 0.2011719 0.1910359 0.1972615
#> 4:lag5  0.1496457 0.1983179 0.2188660 0.2011719 0.2266135 0.2061118 0.1816787
#> 5:lag5  0.1537197 0.1807925 0.2000374 0.1910359 0.2061118 0.2220227 0.1672252
#> 6:lag5  0.1199805 0.1947509 0.1845890 0.1972615 0.1816787 0.1672252 0.2406416
#> 7:lag5  0.1404086 0.2024932 0.2136204 0.2075134 0.2126131 0.1947005 0.1915875
#> 8:lag5  0.1302448 0.1496725 0.1637957 0.1528792 0.1665494 0.1773477 0.1337834
#> 9:lag5  0.1430863 0.2199082 0.2195532 0.1933169 0.2098813 0.1926085 0.1876513
#> 10:lag5 0.1638906 0.1869201 0.2058913 0.1885784 0.2093760 0.2150762 0.1678702
#>            7:lag2    8:lag2    9:lag2   10:lag2    1:lag3    2:lag3    3:lag3
#> 1:t+0   0.3070974 0.2269904 0.3335086 0.2834794 0.2343444 0.2069311 0.1897961
#> 2:t+0   0.3239727 0.2484095 0.3329702 0.3122509 0.2069311 0.2289587 0.1983874
#> 3:t+0   0.3147108 0.2318537 0.2931807 0.2859945 0.1897961 0.1983874 0.2299913
#> 4:t+0   0.3224450 0.2525857 0.3183020 0.3175357 0.1983179 0.2188660 0.2011719
#> 5:t+0   0.2952791 0.2689622 0.2921065 0.3261805 0.1807925 0.2000374 0.1910359
#> 6:t+0   0.2905580 0.2028934 0.2845884 0.2545888 0.1947509 0.1845890 0.1972615
#> 7:t+0   0.3400334 0.2371728 0.3166409 0.2979354 0.2024932 0.2136204 0.2075134
#> 8:t+0   0.2371728 0.3280767 0.2435705 0.2763687 0.1496725 0.1637957 0.1528792
#> 9:t+0   0.3166409 0.2435705 0.3578715 0.3036172 0.2199082 0.2195532 0.1933169
#> 10:t+0  0.2979354 0.2763687 0.3036172 0.3477622 0.1869201 0.2058913 0.1885784
#> 1:lag1  0.5164236 0.3817134 0.5608373 0.4767069 0.3554023 0.3138277 0.2878412
#> 2:lag1  0.5448014 0.4177323 0.5599319 0.5250898 0.3138277 0.3472344 0.3008706
#> 3:lag1  0.5292265 0.3898916 0.4930208 0.4809363 0.2878412 0.3008706 0.3488005
#> 4:lag1  0.5422326 0.4247551 0.5352655 0.5339768 0.3007651 0.3319281 0.3050935
#> 5:lag1  0.4965496 0.4522944 0.4912144 0.5485142 0.2741864 0.3033730 0.2897214
#> 6:lag1  0.4886105 0.3411913 0.4785718 0.4281236 0.2953555 0.2799442 0.2991630
#> 7:lag1  0.5718097 0.3988364 0.5324722 0.5010165 0.3070974 0.3239727 0.3147108
#> 8:lag1  0.3988364 0.5517029 0.4095949 0.4647493 0.2269904 0.2484095 0.2318537
#> 9:lag1  0.5324722 0.4095949 0.6018067 0.5105712 0.3335086 0.3329702 0.2931807
#> 10:lag1 0.5010165 0.4647493 0.5105712 0.5848066 0.2834794 0.3122509 0.2859945
#> 1:lag2  0.8404891 0.6212458 0.9127733 0.7758495 0.5976544 0.5277415 0.4840418
#> 2:lag2  0.8866746 0.6798672 0.9112998 0.8545935 0.5277415 0.5839190 0.5059523
#> 3:lag2  0.8613261 0.6345559 0.8024007 0.7827329 0.4840418 0.5059523 0.5865527
#> 4:lag2  0.8824937 0.6912969 0.8711547 0.8690574 0.5057750 0.5581795 0.5130537
#> 5:lag2  0.8081438 0.7361176 0.7994606 0.8927172 0.4610795 0.5101604 0.4872035
#> 6:lag2  0.7952228 0.5552952 0.7788846 0.6967792 0.4966780 0.4707620 0.5030808
#> 7:lag2  0.9306310 0.6491137 0.8666085 0.8154138 0.5164236 0.5448014 0.5292265
#> 8:lag2  0.6491137 0.8979068 0.6666234 0.7563883 0.3817134 0.4177323 0.3898916
#> 9:lag2  0.8666085 0.6666234 0.9794517 0.8309642 0.5608373 0.5599319 0.4930208
#> 10:lag2 0.8154138 0.7563883 0.8309642 0.9517838 0.4767069 0.5250898 0.4809363
#> 1:lag3  0.5164236 0.3817134 0.5608373 0.4767069 0.9726938 0.8589092 0.7877871
#> 2:lag3  0.5448014 0.4177323 0.5599319 0.5250898 0.8589092 0.9503392 0.8234469
#> 3:lag3  0.5292265 0.3898916 0.4930208 0.4809363 0.7877871 0.8234469 0.9546255
#> 4:lag3  0.5422326 0.4247551 0.5352655 0.5339768 0.8231583 0.9084476 0.8350045
#> 5:lag3  0.4965496 0.4522944 0.4912144 0.5485142 0.7504156 0.8302957 0.7929329
#> 6:lag3  0.4886105 0.3411913 0.4785718 0.4281236 0.8083528 0.7661740 0.8187735
#> 7:lag3  0.5718097 0.3988364 0.5324722 0.5010165 0.8404891 0.8866746 0.8613261
#> 8:lag3  0.3988364 0.5517029 0.4095949 0.4647493 0.6212458 0.6798672 0.6345559
#> 9:lag3  0.5324722 0.4095949 0.6018067 0.5105712 0.9127733 0.9112998 0.8024007
#> 10:lag3 0.5010165 0.4647493 0.5105712 0.5848066 0.7758495 0.8545935 0.7827329
#> 1:lag4  0.3070974 0.2269904 0.3335086 0.2834794 0.5976544 0.5277415 0.4840418
#> 2:lag4  0.3239727 0.2484095 0.3329702 0.3122509 0.5277415 0.5839190 0.5059523
#> 3:lag4  0.3147108 0.2318537 0.2931807 0.2859945 0.4840418 0.5059523 0.5865527
#> 4:lag4  0.3224450 0.2525857 0.3183020 0.3175357 0.5057750 0.5581795 0.5130537
#> 5:lag4  0.2952791 0.2689622 0.2921065 0.3261805 0.4610795 0.5101604 0.4872035
#> 6:lag4  0.2905580 0.2028934 0.2845884 0.2545888 0.4966780 0.4707620 0.5030808
#> 7:lag4  0.3400334 0.2371728 0.3166409 0.2979354 0.5164236 0.5448014 0.5292265
#> 8:lag4  0.2371728 0.3280767 0.2435705 0.2763687 0.3817134 0.4177323 0.3898916
#> 9:lag4  0.3166409 0.2435705 0.3578715 0.3036172 0.5608373 0.5599319 0.4930208
#> 10:lag4 0.2979354 0.2763687 0.3036172 0.3477622 0.4767069 0.5250898 0.4809363
#> 1:lag5  0.2024932 0.1496725 0.2199082 0.1869201 0.3554023 0.3138277 0.2878412
#> 2:lag5  0.2136204 0.1637957 0.2195532 0.2058913 0.3138277 0.3472344 0.3008706
#> 3:lag5  0.2075134 0.1528792 0.1933169 0.1885784 0.2878412 0.3008706 0.3488005
#> 4:lag5  0.2126131 0.1665494 0.2098813 0.2093760 0.3007651 0.3319281 0.3050935
#> 5:lag5  0.1947005 0.1773477 0.1926085 0.2150762 0.2741864 0.3033730 0.2897214
#> 6:lag5  0.1915875 0.1337834 0.1876513 0.1678702 0.2953555 0.2799442 0.2991630
#> 7:lag5  0.2242105 0.1563865 0.2087860 0.1964520 0.3070974 0.3239727 0.3147108
#> 8:lag5  0.1563865 0.2163265 0.1606050 0.1822314 0.2269904 0.2484095 0.2318537
#> 9:lag5  0.2087860 0.1606050 0.2359725 0.2001985 0.3335086 0.3329702 0.2931807
#> 10:lag5 0.1964520 0.1822314 0.2001985 0.2293067 0.2834794 0.3122509 0.2859945
#>            4:lag3    5:lag3    6:lag3    7:lag3    8:lag3    9:lag3   10:lag3
#> 1:t+0   0.1983179 0.1807925 0.1947509 0.2024932 0.1496725 0.2199082 0.1869201
#> 2:t+0   0.2188660 0.2000374 0.1845890 0.2136204 0.1637957 0.2195532 0.2058913
#> 3:t+0   0.2011719 0.1910359 0.1972615 0.2075134 0.1528792 0.1933169 0.1885784
#> 4:t+0   0.2266135 0.2061118 0.1816787 0.2126131 0.1665494 0.2098813 0.2093760
#> 5:t+0   0.2061118 0.2220227 0.1672252 0.1947005 0.1773477 0.1926085 0.2150762
#> 6:t+0   0.1816787 0.1672252 0.2406416 0.1915875 0.1337834 0.1876513 0.1678702
#> 7:t+0   0.2126131 0.1947005 0.1915875 0.2242105 0.1563865 0.2087860 0.1964520
#> 8:t+0   0.1665494 0.1773477 0.1337834 0.1563865 0.2163265 0.1606050 0.1822314
#> 9:t+0   0.2098813 0.1926085 0.1876513 0.2087860 0.1606050 0.2359725 0.2001985
#> 10:t+0  0.2093760 0.2150762 0.1678702 0.1964520 0.1822314 0.2001985 0.2293067
#> 1:lag1  0.3007651 0.2741864 0.2953555 0.3070974 0.2269904 0.3335086 0.2834794
#> 2:lag1  0.3319281 0.3033730 0.2799442 0.3239727 0.2484095 0.3329702 0.3122509
#> 3:lag1  0.3050935 0.2897214 0.2991630 0.3147108 0.2318537 0.2931807 0.2859945
#> 4:lag1  0.3436777 0.3125852 0.2755304 0.3224450 0.2525857 0.3183020 0.3175357
#> 5:lag1  0.3125852 0.3367155 0.2536105 0.2952791 0.2689622 0.2921065 0.3261805
#> 6:lag1  0.2755304 0.2536105 0.3649524 0.2905580 0.2028934 0.2845884 0.2545888
#> 7:lag1  0.3224450 0.2952791 0.2905580 0.3400334 0.2371728 0.3166409 0.2979354
#> 8:lag1  0.2525857 0.2689622 0.2028934 0.2371728 0.3280767 0.2435705 0.2763687
#> 9:lag1  0.3183020 0.2921065 0.2845884 0.3166409 0.2435705 0.3578715 0.3036172
#> 10:lag1 0.3175357 0.3261805 0.2545888 0.2979354 0.2763687 0.3036172 0.3477622
#> 1:lag2  0.5057750 0.4610795 0.4966780 0.5164236 0.3817134 0.5608373 0.4767069
#> 2:lag2  0.5581795 0.5101604 0.4707620 0.5448014 0.4177323 0.5599319 0.5250898
#> 3:lag2  0.5130537 0.4872035 0.5030808 0.5292265 0.3898916 0.4930208 0.4809363
#> 4:lag2  0.5779380 0.5256521 0.4633397 0.5422326 0.4247551 0.5352655 0.5339768
#> 5:lag2  0.5256521 0.5662302 0.4264786 0.4965496 0.4522944 0.4912144 0.5485142
#> 6:lag2  0.4633397 0.4264786 0.6137142 0.4886105 0.3411913 0.4785718 0.4281236
#> 7:lag2  0.5422326 0.4965496 0.4886105 0.5718097 0.3988364 0.5324722 0.5010165
#> 8:lag2  0.4247551 0.4522944 0.3411913 0.3988364 0.5517029 0.4095949 0.4647493
#> 9:lag2  0.5352655 0.4912144 0.4785718 0.5324722 0.4095949 0.6018067 0.5105712
#> 10:lag2 0.5339768 0.5485142 0.4281236 0.5010165 0.4647493 0.5105712 0.5848066
#> 1:lag3  0.8231583 0.7504156 0.8083528 0.8404891 0.6212458 0.9127733 0.7758495
#> 2:lag3  0.9084476 0.8302957 0.7661740 0.8866746 0.6798672 0.9112998 0.8545935
#> 3:lag3  0.8350045 0.7929329 0.8187735 0.8613261 0.6345559 0.8024007 0.7827329
#> 4:lag3  0.9406050 0.8555086 0.7540940 0.8824937 0.6912969 0.8711547 0.8690574
#> 5:lag3  0.8555086 0.9215502 0.6941019 0.8081438 0.7361176 0.7994606 0.8927172
#> 6:lag3  0.7540940 0.6941019 0.9988314 0.7952228 0.5552952 0.7788846 0.6967792
#> 7:lag3  0.8824937 0.8081438 0.7952228 0.9306310 0.6491137 0.8666085 0.8154138
#> 8:lag3  0.6912969 0.7361176 0.5552952 0.6491137 0.8979068 0.6666234 0.7563883
#> 9:lag3  0.8711547 0.7994606 0.7788846 0.8666085 0.6666234 0.9794517 0.8309642
#> 10:lag3 0.8690574 0.8927172 0.6967792 0.8154138 0.7563883 0.8309642 0.9517838
#> 1:lag4  0.5057750 0.4610795 0.4966780 0.5164236 0.3817134 0.5608373 0.4767069
#> 2:lag4  0.5581795 0.5101604 0.4707620 0.5448014 0.4177323 0.5599319 0.5250898
#> 3:lag4  0.5130537 0.4872035 0.5030808 0.5292265 0.3898916 0.4930208 0.4809363
#> 4:lag4  0.5779380 0.5256521 0.4633397 0.5422326 0.4247551 0.5352655 0.5339768
#> 5:lag4  0.5256521 0.5662302 0.4264786 0.4965496 0.4522944 0.4912144 0.5485142
#> 6:lag4  0.4633397 0.4264786 0.6137142 0.4886105 0.3411913 0.4785718 0.4281236
#> 7:lag4  0.5422326 0.4965496 0.4886105 0.5718097 0.3988364 0.5324722 0.5010165
#> 8:lag4  0.4247551 0.4522944 0.3411913 0.3988364 0.5517029 0.4095949 0.4647493
#> 9:lag4  0.5352655 0.4912144 0.4785718 0.5324722 0.4095949 0.6018067 0.5105712
#> 10:lag4 0.5339768 0.5485142 0.4281236 0.5010165 0.4647493 0.5105712 0.5848066
#> 1:lag5  0.3007651 0.2741864 0.2953555 0.3070974 0.2269904 0.3335086 0.2834794
#> 2:lag5  0.3319281 0.3033730 0.2799442 0.3239727 0.2484095 0.3329702 0.3122509
#> 3:lag5  0.3050935 0.2897214 0.2991630 0.3147108 0.2318537 0.2931807 0.2859945
#> 4:lag5  0.3436777 0.3125852 0.2755304 0.3224450 0.2525857 0.3183020 0.3175357
#> 5:lag5  0.3125852 0.3367155 0.2536105 0.2952791 0.2689622 0.2921065 0.3261805
#> 6:lag5  0.2755304 0.2536105 0.3649524 0.2905580 0.2028934 0.2845884 0.2545888
#> 7:lag5  0.3224450 0.2952791 0.2905580 0.3400334 0.2371728 0.3166409 0.2979354
#> 8:lag5  0.2525857 0.2689622 0.2028934 0.2371728 0.3280767 0.2435705 0.2763687
#> 9:lag5  0.3183020 0.2921065 0.2845884 0.3166409 0.2435705 0.3578715 0.3036172
#> 10:lag5 0.3175357 0.3261805 0.2545888 0.2979354 0.2763687 0.3036172 0.3477622
#>            1:lag4    2:lag4    3:lag4    4:lag4    5:lag4     6:lag4    7:lag4
#> 1:t+0   0.1674911 0.1478982 0.1356515 0.1417422 0.1292164 0.13919275 0.1447264
#> 2:t+0   0.1478982 0.1636418 0.1417919 0.1564284 0.1429712 0.13192985 0.1526792
#> 3:t+0   0.1356515 0.1417919 0.1643799 0.1437820 0.1365376 0.14098713 0.1483144
#> 4:t+0   0.1417422 0.1564284 0.1437820 0.1619657 0.1473127 0.12984977 0.1519593
#> 5:t+0   0.1292164 0.1429712 0.1365376 0.1473127 0.1586846 0.11951954 0.1391568
#> 6:t+0   0.1391928 0.1319298 0.1409871 0.1298498 0.1195195 0.17199185 0.1369319
#> 7:t+0   0.1447264 0.1526792 0.1483144 0.1519593 0.1391568 0.13693185 0.1602482
#> 8:t+0   0.1069742 0.1170684 0.1092661 0.1190365 0.1267544 0.09561799 0.1117729
#> 9:t+0   0.1571732 0.1569195 0.1381678 0.1500068 0.1376616 0.13411853 0.1492240
#> 10:t+0  0.1335959 0.1471551 0.1347812 0.1496457 0.1537197 0.11998055 0.1404086
#> 1:lag1  0.2343444 0.2069311 0.1897961 0.1983179 0.1807925 0.19475087 0.2024932
#> 2:lag1  0.2069311 0.2289587 0.1983874 0.2188660 0.2000374 0.18458901 0.2136204
#> 3:lag1  0.1897961 0.1983874 0.2299913 0.2011719 0.1910359 0.19726146 0.2075134
#> 4:lag1  0.1983179 0.2188660 0.2011719 0.2266135 0.2061118 0.18167868 0.2126131
#> 5:lag1  0.1807925 0.2000374 0.1910359 0.2061118 0.2220227 0.16722518 0.1947005
#> 6:lag1  0.1947509 0.1845890 0.1972615 0.1816787 0.1672252 0.24064156 0.1915875
#> 7:lag1  0.2024932 0.2136204 0.2075134 0.2126131 0.1947005 0.19158754 0.2242105
#> 8:lag1  0.1496725 0.1637957 0.1528792 0.1665494 0.1773477 0.13378344 0.1563865
#> 9:lag1  0.2199082 0.2195532 0.1933169 0.2098813 0.1926085 0.18765128 0.2087860
#> 10:lag1 0.1869201 0.2058913 0.1885784 0.2093760 0.2150762 0.16787020 0.1964520
#> 1:lag2  0.3554023 0.3138277 0.2878412 0.3007651 0.2741864 0.29535548 0.3070974
#> 2:lag2  0.3138277 0.3472344 0.3008706 0.3319281 0.3033730 0.27994420 0.3239727
#> 3:lag2  0.2878412 0.3008706 0.3488005 0.3050935 0.2897214 0.29916300 0.3147108
#> 4:lag2  0.3007651 0.3319281 0.3050935 0.3436777 0.3125852 0.27553044 0.3224450
#> 5:lag2  0.2741864 0.3033730 0.2897214 0.3125852 0.3367155 0.25361054 0.2952791
#> 6:lag2  0.2953555 0.2799442 0.2991630 0.2755304 0.2536105 0.36495244 0.2905580
#> 7:lag2  0.3070974 0.3239727 0.3147108 0.3224450 0.2952791 0.29055803 0.3400334
#> 8:lag2  0.2269904 0.2484095 0.2318537 0.2525857 0.2689622 0.20289344 0.2371728
#> 9:lag2  0.3335086 0.3329702 0.2931807 0.3183020 0.2921065 0.28458839 0.3166409
#> 10:lag2 0.2834794 0.3122509 0.2859945 0.3175357 0.3261805 0.25458878 0.2979354
#> 1:lag3  0.5976544 0.5277415 0.4840418 0.5057750 0.4610795 0.49667801 0.5164236
#> 2:lag3  0.5277415 0.5839190 0.5059523 0.5581795 0.5101604 0.47076197 0.5448014
#> 3:lag3  0.4840418 0.5059523 0.5865527 0.5130537 0.4872035 0.50308084 0.5292265
#> 4:lag3  0.5057750 0.5581795 0.5130537 0.5779380 0.5256521 0.46333967 0.5422326
#> 5:lag3  0.4610795 0.5101604 0.4872035 0.5256521 0.5662302 0.42647856 0.4965496
#> 6:lag3  0.4966780 0.4707620 0.5030808 0.4633397 0.4264786 0.61371419 0.4886105
#> 7:lag3  0.5164236 0.5448014 0.5292265 0.5422326 0.4965496 0.48861049 0.5718097
#> 8:lag3  0.3817134 0.4177323 0.3898916 0.4247551 0.4522944 0.34119127 0.3988364
#> 9:lag3  0.5608373 0.5599319 0.4930208 0.5352655 0.4912144 0.47857176 0.5324722
#> 10:lag3 0.4767069 0.5250898 0.4809363 0.5339768 0.5485142 0.42812358 0.5010165
#> 1:lag4  0.9726938 0.8589092 0.7877871 0.8231583 0.7504156 0.80835281 0.8404891
#> 2:lag4  0.8589092 0.9503392 0.8234469 0.9084476 0.8302957 0.76617397 0.8866746
#> 3:lag4  0.7877871 0.8234469 0.9546255 0.8350045 0.7929329 0.81877354 0.8613261
#> 4:lag4  0.8231583 0.9084476 0.8350045 0.9406050 0.8555086 0.75409404 0.8824937
#> 5:lag4  0.7504156 0.8302957 0.7929329 0.8555086 0.9215502 0.69410189 0.8081438
#> 6:lag4  0.8083528 0.7661740 0.8187735 0.7540940 0.6941019 0.99883141 0.7952228
#> 7:lag4  0.8404891 0.8866746 0.8613261 0.8824937 0.8081438 0.79522277 0.9306310
#> 8:lag4  0.6212458 0.6798672 0.6345559 0.6912969 0.7361176 0.55529521 0.6491137
#> 9:lag4  0.9127733 0.9112998 0.8024007 0.8711547 0.7994606 0.77888455 0.8666085
#> 10:lag4 0.7758495 0.8545935 0.7827329 0.8690574 0.8927172 0.69677919 0.8154138
#> 1:lag5  0.5976544 0.5277415 0.4840418 0.5057750 0.4610795 0.49667801 0.5164236
#> 2:lag5  0.5277415 0.5839190 0.5059523 0.5581795 0.5101604 0.47076197 0.5448014
#> 3:lag5  0.4840418 0.5059523 0.5865527 0.5130537 0.4872035 0.50308084 0.5292265
#> 4:lag5  0.5057750 0.5581795 0.5130537 0.5779380 0.5256521 0.46333967 0.5422326
#> 5:lag5  0.4610795 0.5101604 0.4872035 0.5256521 0.5662302 0.42647856 0.4965496
#> 6:lag5  0.4966780 0.4707620 0.5030808 0.4633397 0.4264786 0.61371419 0.4886105
#> 7:lag5  0.5164236 0.5448014 0.5292265 0.5422326 0.4965496 0.48861049 0.5718097
#> 8:lag5  0.3817134 0.4177323 0.3898916 0.4247551 0.4522944 0.34119127 0.3988364
#> 9:lag5  0.5608373 0.5599319 0.4930208 0.5352655 0.4912144 0.47857176 0.5324722
#> 10:lag5 0.4767069 0.5250898 0.4809363 0.5339768 0.5485142 0.42812358 0.5010165
#>             8:lag4    9:lag4   10:lag4     1:lag5     2:lag5     3:lag5
#> 1:t+0   0.10697423 0.1571732 0.1335959 0.12678600 0.11195471 0.10268429
#> 2:t+0   0.11706843 0.1569195 0.1471551 0.11195471 0.12387219 0.10733238
#> 3:t+0   0.10926614 0.1381678 0.1347812 0.10268429 0.10733238 0.12443089
#> 4:t+0   0.11903654 0.1500068 0.1496457 0.10729476 0.11841181 0.10883886
#> 5:t+0   0.12675435 0.1376616 0.1537197 0.09781310 0.10822508 0.10335503
#> 6:t+0   0.09561799 0.1341185 0.1199805 0.10536493 0.09986712 0.10672323
#> 7:t+0   0.11177288 0.1492240 0.1404086 0.10955375 0.11557381 0.11226975
#> 8:t+0   0.15461333 0.1147879 0.1302448 0.08097643 0.08861745 0.08271134
#> 9:t+0   0.11478793 0.1686548 0.1430863 0.11897565 0.11878358 0.10458910
#> 10:t+0  0.13024483 0.1430863 0.1638906 0.10112829 0.11139219 0.10202550
#> 1:lag1  0.14967247 0.2199082 0.1869201 0.16749113 0.14789821 0.13565148
#> 2:lag1  0.16379572 0.2195532 0.2058913 0.14789821 0.16364184 0.14179185
#> 3:lag1  0.15287919 0.1933169 0.1885784 0.13565148 0.14179185 0.16437990
#> 4:lag1  0.16654939 0.2098813 0.2093760 0.14174216 0.15642838 0.14378199
#> 5:lag1  0.17734774 0.1926085 0.2150762 0.12921637 0.14297117 0.13653756
#> 6:lag1  0.13378344 0.1876513 0.1678702 0.13919275 0.13192985 0.14098713
#> 7:lag1  0.15638649 0.2087860 0.1964520 0.14472640 0.15267923 0.14831439
#> 8:lag1  0.21632649 0.1606050 0.1822314 0.10697423 0.11706843 0.10926614
#> 9:lag1  0.16060497 0.2359725 0.2001985 0.15717325 0.15691951 0.13816784
#> 10:lag1 0.18223142 0.2001985 0.2293067 0.13359591 0.14715508 0.13478118
#> 1:lag2  0.22699044 0.3335086 0.2834794 0.23434441 0.20693106 0.18979612
#> 2:lag2  0.24840950 0.3329702 0.3122509 0.20693106 0.22895868 0.19838738
#> 3:lag2  0.23185368 0.2931807 0.2859945 0.18979612 0.19838738 0.22999135
#> 4:lag2  0.25258566 0.3183020 0.3175357 0.19831785 0.21886601 0.20117188
#> 5:lag2  0.26896222 0.2921065 0.3261805 0.18079246 0.20003741 0.19103586
#> 6:lag2  0.20289344 0.2845884 0.2545888 0.19475087 0.18458901 0.19726146
#> 7:lag2  0.23717279 0.3166409 0.2979354 0.20249324 0.21362040 0.20751336
#> 8:lag2  0.32807666 0.2435705 0.2763687 0.14967247 0.16379572 0.15287919
#> 9:lag2  0.24357046 0.3578715 0.3036172 0.21990818 0.21955316 0.19331686
#> 10:lag2 0.27636872 0.3036172 0.3477622 0.18692008 0.20589132 0.18857844
#> 1:lag3  0.38171345 0.5608373 0.4767069 0.35540229 0.31382773 0.28784120
#> 2:lag3  0.41773233 0.5599319 0.5250898 0.31382773 0.34723440 0.30087055
#> 3:lag3  0.38989162 0.4930208 0.4809363 0.28784120 0.30087055 0.34880052
#> 4:lag3  0.42475508 0.5352655 0.5339768 0.30076510 0.33192806 0.30509346
#> 5:lag3  0.45229437 0.4912144 0.5485142 0.27418642 0.30337296 0.28972138
#> 6:lag3  0.34119127 0.4785718 0.4281236 0.29535548 0.27994420 0.29916300
#> 7:lag3  0.39883637 0.5324722 0.5010165 0.30709742 0.32397265 0.31471083
#> 8:lag3  0.55170286 0.4095949 0.4647493 0.22699044 0.24840950 0.23185368
#> 9:lag3  0.40959488 0.6018067 0.5105712 0.33350859 0.33297018 0.29318069
#> 10:lag3 0.46474934 0.5105712 0.5848066 0.28347945 0.31225088 0.28599449
#> 1:lag4  0.62124582 0.9127733 0.7758495 0.59765440 0.52774146 0.48404178
#> 2:lag4  0.67986724 0.9112998 0.8545935 0.52774146 0.58391905 0.50595230
#> 3:lag4  0.63455595 0.8024007 0.7827329 0.48404178 0.50595230 0.58655268
#> 4:lag4  0.69129689 0.8711547 0.8690574 0.50577498 0.55817947 0.51305367
#> 5:lag4  0.73611760 0.7994606 0.8927172 0.46107953 0.51016042 0.48720354
#> 6:lag4  0.55529521 0.7788846 0.6967792 0.49667801 0.47076197 0.50308084
#> 7:lag4  0.64911371 0.8666085 0.8154138 0.51642358 0.54480144 0.52922651
#> 8:lag4  0.89790679 0.6666234 0.7563883 0.38171345 0.41773233 0.38989162
#> 9:lag4  0.66662338 0.9794517 0.8309642 0.56083734 0.55993193 0.49302082
#> 10:lag4 0.75638830 0.8309642 0.9517838 0.47670694 0.52508979 0.48093631
#> 1:lag5  0.38171345 0.5608373 0.4767069 0.97269380 0.85890916 0.78778712
#> 2:lag5  0.41773233 0.5599319 0.5250898 0.85890916 0.95033925 0.82344690
#> 3:lag5  0.38989162 0.4930208 0.4809363 0.78778712 0.82344690 0.95462554
#> 4:lag5  0.42475508 0.5352655 0.5339768 0.82315831 0.90844760 0.83500450
#> 5:lag5  0.45229437 0.4912144 0.5485142 0.75041561 0.83029569 0.79293293
#> 6:lag5  0.34119127 0.4785718 0.4281236 0.80835281 0.76617397 0.81877354
#> 7:lag5  0.39883637 0.5324722 0.5010165 0.84048909 0.88667461 0.86132611
#> 8:lag5  0.55170286 0.4095949 0.4647493 0.62124582 0.67986724 0.63455595
#> 9:lag5  0.40959488 0.6018067 0.5105712 0.91277334 0.91129976 0.80240066
#> 10:lag5 0.46474934 0.5105712 0.5848066 0.77584953 0.85459351 0.78273290
#>             4:lag5     5:lag5     6:lag5     7:lag5     8:lag5     9:lag5
#> 1:t+0   0.10729476 0.09781310 0.10536493 0.10955375 0.08097643 0.11897565
#> 2:t+0   0.11841181 0.10822508 0.09986712 0.11557381 0.08861745 0.11878358
#> 3:t+0   0.10883886 0.10335503 0.10672323 0.11226975 0.08271134 0.10458910
#> 4:t+0   0.12260338 0.11151147 0.09829256 0.11502885 0.09010725 0.11355086
#> 5:t+0   0.11151147 0.12011968 0.09047287 0.10533769 0.09594942 0.10420588
#> 6:t+0   0.09829256 0.09047287 0.13019291 0.10365350 0.07238008 0.10152389
#> 7:t+0   0.11502885 0.10533769 0.10365350 0.12130331 0.08460888 0.11295829
#> 8:t+0   0.09010725 0.09594942 0.07238008 0.08460888 0.11703787 0.08689118
#> 9:t+0   0.11355086 0.10420588 0.10152389 0.11295829 0.08689118 0.12766685
#> 10:t+0  0.11327748 0.11636143 0.09082185 0.10628530 0.09859161 0.10831222
#> 1:lag1  0.14174216 0.12921637 0.13919275 0.14472640 0.10697423 0.15717325
#> 2:lag1  0.15642838 0.14297117 0.13192985 0.15267923 0.11706843 0.15691951
#> 3:lag1  0.14378199 0.13653756 0.14098713 0.14831439 0.10926614 0.13816784
#> 4:lag1  0.16196566 0.14731266 0.12984977 0.15195931 0.11903654 0.15000680
#> 5:lag1  0.14731266 0.15868457 0.11951954 0.13915676 0.12675435 0.13766158
#> 6:lag1  0.12984977 0.11951954 0.17199185 0.13693185 0.09561799 0.13411853
#> 7:lag1  0.15195931 0.13915676 0.13693185 0.16024821 0.11177288 0.14922398
#> 8:lag1  0.11903654 0.12675435 0.09561799 0.11177288 0.15461333 0.11478793
#> 9:lag1  0.15000680 0.13766158 0.13411853 0.14922398 0.11478793 0.16865479
#> 10:lag1 0.14964566 0.15371972 0.11998055 0.14040861 0.13024483 0.14308627
#> 1:lag2  0.19831785 0.18079246 0.19475087 0.20249324 0.14967247 0.21990818
#> 2:lag2  0.21886601 0.20003741 0.18458901 0.21362040 0.16379572 0.21955316
#> 3:lag2  0.20117188 0.19103586 0.19726146 0.20751336 0.15287919 0.19331686
#> 4:lag2  0.22661347 0.20611179 0.18167868 0.21261313 0.16654939 0.20988129
#> 5:lag2  0.20611179 0.22202274 0.16722518 0.19470051 0.17734774 0.19260854
#> 6:lag2  0.18167868 0.16722518 0.24064156 0.19158754 0.13378344 0.18765128
#> 7:lag2  0.21261313 0.19470051 0.19158754 0.22421051 0.15638649 0.20878601
#> 8:lag2  0.16654939 0.17734774 0.13378344 0.15638649 0.21632649 0.16060497
#> 9:lag2  0.20988129 0.19260854 0.18765128 0.20878601 0.16060497 0.23597253
#> 10:lag2 0.20937600 0.21507619 0.16787020 0.19645202 0.18223142 0.20019847
#> 1:lag3  0.30076510 0.27418642 0.29535548 0.30709742 0.22699044 0.33350859
#> 2:lag3  0.33192806 0.30337296 0.27994420 0.32397265 0.24840950 0.33297018
#> 3:lag3  0.30509346 0.28972138 0.29916300 0.31471083 0.23185368 0.29318069
#> 4:lag3  0.34367771 0.31258525 0.27553044 0.32244505 0.25258566 0.31830200
#> 5:lag3  0.31258525 0.33671549 0.25361054 0.29527911 0.26896222 0.29210646
#> 6:lag3  0.27553044 0.25361054 0.36495244 0.29055803 0.20289344 0.28458839
#> 7:lag3  0.32244505 0.29527911 0.29055803 0.34003341 0.23717279 0.31664091
#> 8:lag3  0.25258566 0.26896222 0.20289344 0.23717279 0.32807666 0.24357046
#> 9:lag3  0.31830200 0.29210646 0.28458839 0.31664091 0.24357046 0.35787148
#> 10:lag3 0.31753567 0.32618049 0.25458878 0.29793542 0.27636872 0.30361720
#> 1:lag4  0.50577498 0.46107953 0.49667801 0.51642358 0.38171345 0.56083734
#> 2:lag4  0.55817947 0.51016042 0.47076197 0.54480144 0.41773233 0.55993193
#> 3:lag4  0.51305367 0.48720354 0.50308084 0.52922651 0.38989162 0.49302082
#> 4:lag4  0.57793801 0.52565207 0.46333967 0.54223258 0.42475508 0.53526551
#> 5:lag4  0.52565207 0.56623015 0.42647856 0.49654959 0.45229437 0.49121437
#> 6:lag4  0.46333967 0.42647856 0.61371419 0.48861049 0.34119127 0.47857176
#> 7:lag4  0.54223258 0.49654959 0.48861049 0.57180967 0.39883637 0.53247218
#> 8:lag4  0.42475508 0.45229437 0.34119127 0.39883637 0.55170286 0.40959488
#> 9:lag4  0.53526551 0.49121437 0.47857176 0.53247218 0.40959488 0.60180665
#> 10:lag4 0.53397684 0.54851420 0.42812358 0.50101651 0.46474934 0.51057115
#> 1:lag5  0.82315831 0.75041561 0.80835281 0.84048909 0.62124582 0.91277334
#> 2:lag5  0.90844760 0.83029569 0.76617397 0.88667461 0.67986724 0.91129976
#> 3:lag5  0.83500450 0.79293293 0.81877354 0.86132611 0.63455595 0.80240066
#> 4:lag5  0.94060499 0.85550863 0.75409404 0.88249374 0.69129689 0.87115469
#> 5:lag5  0.85550863 0.92155023 0.69410189 0.80814380 0.73611760 0.79946063
#> 6:lag5  0.75409404 0.69410189 0.99883141 0.79522277 0.55529521 0.77888455
#> 7:lag5  0.88249374 0.80814380 0.79522277 0.93063100 0.64911371 0.86660850
#> 8:lag5  0.69129689 0.73611760 0.55529521 0.64911371 0.89790679 0.66662338
#> 9:lag5  0.87115469 0.79946063 0.77888455 0.86660850 0.66662338 0.97945166
#> 10:lag5 0.86905736 0.89271719 0.69677919 0.81541380 0.75638830 0.83096417
#>            10:lag5
#> 1:t+0   0.10112829
#> 2:t+0   0.11139219
#> 3:t+0   0.10202550
#> 4:t+0   0.11327748
#> 5:t+0   0.11636143
#> 6:t+0   0.09082185
#> 7:t+0   0.10628530
#> 8:t+0   0.09859161
#> 9:t+0   0.10831222
#> 10:t+0  0.12406048
#> 1:lag1  0.13359591
#> 2:lag1  0.14715508
#> 3:lag1  0.13478118
#> 4:lag1  0.14964566
#> 5:lag1  0.15371972
#> 6:lag1  0.11998055
#> 7:lag1  0.14040861
#> 8:lag1  0.13024483
#> 9:lag1  0.14308627
#> 10:lag1 0.16389058
#> 1:lag2  0.18692008
#> 2:lag2  0.20589132
#> 3:lag2  0.18857844
#> 4:lag2  0.20937600
#> 5:lag2  0.21507619
#> 6:lag2  0.16787020
#> 7:lag2  0.19645202
#> 8:lag2  0.18223142
#> 9:lag2  0.20019847
#> 10:lag2 0.22930671
#> 1:lag3  0.28347945
#> 2:lag3  0.31225088
#> 3:lag3  0.28599449
#> 4:lag3  0.31753567
#> 5:lag3  0.32618049
#> 6:lag3  0.25458878
#> 7:lag3  0.29793542
#> 8:lag3  0.27636872
#> 9:lag3  0.30361720
#> 10:lag3 0.34776222
#> 1:lag4  0.47670694
#> 2:lag4  0.52508979
#> 3:lag4  0.48093631
#> 4:lag4  0.53397684
#> 5:lag4  0.54851420
#> 6:lag4  0.42812358
#> 7:lag4  0.50101651
#> 8:lag4  0.46474934
#> 9:lag4  0.51057115
#> 10:lag4 0.58480664
#> 1:lag5  0.77584953
#> 2:lag5  0.85459351
#> 3:lag5  0.78273290
#> 4:lag5  0.86905736
#> 5:lag5  0.89271719
#> 6:lag5  0.69677919
#> 7:lag5  0.81541380
#> 8:lag5  0.75638830
#> 9:lag5  0.83096417
#> 10:lag5 0.95178382