Covariance/correlation for joint distribution of an mcgf object

Covariance/correlation for joint distribution of an mcgf object

Usage

# S3 method for mcgf
ccov(x, model = c("all", "base", "empirical"), cor = FALSE, ...)

Arguments

x

An mcgf object.

model

Which model to use. One of all, base, and empirical.

cor

Logical; if TRUE, correlation is outputted.

...

Additional arguments. Not in use.

Value

Joint covariance/correlation matrix.

Examples

data(sim1)
sim1_mcgf <- mcgf(sim1$data, dists = sim1$dists)
#> `time` is not provided, assuming rows are equally spaced temporally.
sim1_mcgf <- add_acfs(sim1_mcgf, lag_max = 5)
sim1_mcgf <- add_ccfs(sim1_mcgf, lag_max = 5)

# Fit a separable model and store it to 'sim1_mcgf'
fit_sep <- fit_base(
    sim1_mcgf,
    model = "sep",
    lag = 5,
    par_init = c(
        c = 0.001,
        gamma = 0.5,
        a = 0.3,
        alpha = 0.5
    ),
    par_fixed = c(nugget = 0)
)
sim1_mcgf <- add_base(sim1_mcgf, fit_base = fit_sep)

ccov(sim1_mcgf, model = "base")
#>              loc1:t+0   loc2:t+0   loc3:t+0   loc4:t+0   loc5:t+0   loc6:t+0
#> loc1:t+0   0.97049881 0.86397901 0.74316928 0.82576241 0.76077679 0.75096792
#> loc2:t+0   0.86397901 0.95240499 0.77854406 0.90650535 0.83863841 0.73807472
#> loc3:t+0   0.74316928 0.77854406 0.96213325 0.80004853 0.77419859 0.83434895
#> loc4:t+0   0.82576241 0.90650535 0.80004853 0.94425364 0.86593608 0.74038961
#> loc5:t+0   0.76077679 0.83863841 0.77419859 0.86593608 0.92683326 0.69391392
#> loc6:t+0   0.75096792 0.73807472 0.83434895 0.74038961 0.69391392 0.99858754
#> loc7:t+0   0.82177023 0.86731017 0.83263182 0.87917939 0.81697223 0.78270785
#> loc8:t+0   0.64163758 0.69084736 0.60880822 0.69399895 0.72726517 0.54574698
#> loc9:t+0   0.91718794 0.90519597 0.74803219 0.86160003 0.80215456 0.73017319
#> loc10:t+0  0.79424743 0.86703233 0.74767901 0.87142938 0.88333330 0.68334951
#> loc1:lag1  0.59739497 0.53182622 0.45746124 0.50830182 0.46829962 0.46226173
#> loc2:lag1  0.53182622 0.58625724 0.47923635 0.55800350 0.51622770 0.45432526
#> loc3:lag1  0.45746124 0.47923635 0.59224551 0.49247352 0.47656147 0.51358730
#> loc4:lag1  0.50830182 0.55800350 0.49247352 0.58123964 0.53303090 0.45575020
#> loc5:lag1  0.46829962 0.51622770 0.47656147 0.53303090 0.57051645 0.42714188
#> loc6:lag1  0.46226173 0.45432526 0.51358730 0.45575020 0.42714188 0.61468512
#> loc7:lag1  0.50584442 0.53387673 0.51253032 0.54118288 0.50289098 0.48179939
#> loc8:lag1  0.39496294 0.42525425 0.37475468 0.42719422 0.44767139 0.33593704
#> loc9:lag1  0.56457922 0.55719751 0.46045462 0.53036183 0.49376990 0.44946144
#> loc10:lag1 0.48890263 0.53370571 0.46023723 0.53641233 0.54373985 0.42063891
#> loc1:lag2  0.35532213 0.31632276 0.27209152 0.30233077 0.27853803 0.27494678
#> loc2:lag2  0.31632276 0.34869756 0.28504304 0.33189264 0.30704498 0.27022628
#> loc3:lag2  0.27209152 0.28504304 0.35225930 0.29291633 0.28345206 0.30547451
#> loc4:lag2  0.30233077 0.33189264 0.29291633 0.34571316 0.31703929 0.27107381
#> loc5:lag2  0.27853803 0.30704498 0.28345206 0.31703929 0.33933516 0.25405798
#> loc6:lag2  0.27494678 0.27022628 0.30547451 0.27107381 0.25405798 0.36560607
#> loc7:lag2  0.30086915 0.31754237 0.30484583 0.32188797 0.29911248 0.28656750
#> loc8:lag2  0.23491840 0.25293525 0.22289881 0.25408911 0.26626865 0.19981063
#> loc9:lag2  0.33580378 0.33141324 0.27387194 0.31545176 0.29368739 0.26733334
#> loc10:lag2 0.29079241 0.31744065 0.27374263 0.31905051 0.32340881 0.25019011
#> loc1:lag3  0.23422564 0.20851755 0.17936066 0.19929414 0.18361015 0.18124282
#> loc2:lag3  0.20851755 0.22985878 0.18789820 0.21878110 0.20240171 0.17813110
#> loc3:lag3  0.17936066 0.18789820 0.23220665 0.19308821 0.18684944 0.20136647
#> loc4:lag3  0.19929414 0.21878110 0.19308821 0.22789149 0.20898988 0.17868979
#> loc5:lag3  0.18361015 0.20240171 0.18684944 0.20898988 0.22368715 0.16747309
#> loc6:lag3  0.18124282 0.17813110 0.20136647 0.17868979 0.16747309 0.24100473
#> loc7:lag3  0.19833065 0.20932151 0.20095205 0.21218610 0.19717267 0.18890311
#> loc8:lag3  0.15485642 0.16673299 0.14693320 0.16749361 0.17552227 0.13171364
#> loc9:lag3  0.22135930 0.21846509 0.18053430 0.20794340 0.19359649 0.17622410
#> loc10:lag3 0.19168814 0.20925446 0.18044906 0.21031567 0.21318863 0.16492341
#> loc1:lag4  0.16734466 0.14897728 0.12814586 0.14238753 0.13118195 0.12949059
#> loc2:lag4  0.14897728 0.16422471 0.13424559 0.15631016 0.14460776 0.12726740
#> loc3:lag4  0.12814586 0.13424559 0.16590217 0.13795364 0.13349630 0.14386812
#> loc4:lag4  0.14238753 0.15631016 0.13795364 0.16281916 0.14931474 0.12766656
#> loc5:lag4  0.13118195 0.14460776 0.13349630 0.14931474 0.15981534 0.11965268
#> loc6:lag4  0.12949059 0.12726740 0.14386812 0.12766656 0.11965268 0.17218804
#> loc7:lag4  0.14169915 0.14955167 0.14357203 0.15159830 0.14087182 0.13496356
#> loc8:lag4  0.11063859 0.11912391 0.10497777 0.11966734 0.12540349 0.09410402
#> loc9:lag4  0.15815218 0.15608438 0.12898438 0.14856707 0.13831679 0.12590493
#> loc10:lag4 0.13695335 0.14950377 0.12892348 0.15026196 0.15231457 0.11783104
#> loc1:lag5  0.12662961 0.11273102 0.09696790 0.10774456 0.09926531 0.09798546
#> loc2:lag5  0.11273102 0.12426875 0.10158357 0.11827981 0.10942461 0.09630317
#> loc3:lag5  0.09696790 0.10158357 0.12553808 0.10438945 0.10101657 0.10886493
#> loc4:lag5  0.10774456 0.11827981 0.10438945 0.12320517 0.11298638 0.09660522
#> loc5:lag5  0.09926531 0.10942461 0.10101657 0.11298638 0.12093218 0.09054112
#> loc6:lag5  0.09798546 0.09630317 0.10886493 0.09660522 0.09054112 0.13029460
#> loc7:lag5  0.10722367 0.11316567 0.10864088 0.11471435 0.10659763 0.10212685
#> loc8:lag5  0.08372016 0.09014100 0.07943662 0.09055221 0.09489275 0.07120846
#> loc9:lag5  0.11967367 0.11810897 0.09760241 0.11242062 0.10466424 0.09527219
#> loc10:lag5 0.10363252 0.11312942 0.09755633 0.11370314 0.11525635 0.08916269
#>              loc7:t+0   loc8:t+0   loc9:t+0  loc10:t+0 loc1:lag1 loc2:lag1
#> loc1:t+0   0.82177023 0.64163758 0.91718794 0.79424743 0.5973950 0.5318262
#> loc2:t+0   0.86731017 0.69084736 0.90519597 0.86703233 0.5318262 0.5862572
#> loc3:t+0   0.83263182 0.60880822 0.74803219 0.74767901 0.4574612 0.4792364
#> loc4:t+0   0.87917939 0.69399895 0.86160003 0.87142938 0.5083018 0.5580035
#> loc5:t+0   0.81697223 0.72726517 0.80215456 0.88333330 0.4682996 0.5162277
#> loc6:t+0   0.78270785 0.54574698 0.73017319 0.68334951 0.4622617 0.4543253
#> loc7:t+0   0.92943762 0.64696869 0.83479401 0.81143827 0.5058444 0.5338767
#> loc8:t+0   0.64696869 0.90230824 0.68658155 0.76556666 0.3949629 0.4252542
#> loc9:t+0   0.83479401 0.68658155 0.98144745 0.84639711 0.5645792 0.5571975
#> loc10:t+0  0.81143827 0.76556666 0.84639711 0.96276319 0.4889026 0.5337057
#> loc1:lag1  0.50584442 0.39496294 0.56457922 0.48890263 0.9704988 0.8639790
#> loc2:lag1  0.53387673 0.42525425 0.55719751 0.53370571 0.8639790 0.9524050
#> loc3:lag1  0.51253032 0.37475468 0.46045462 0.46023723 0.7431693 0.7785441
#> loc4:lag1  0.54118288 0.42719422 0.53036183 0.53641233 0.8257624 0.9065053
#> loc5:lag1  0.50289098 0.44767139 0.49376990 0.54373985 0.7607768 0.8386384
#> loc6:lag1  0.48179939 0.33593704 0.44946144 0.42063891 0.7509679 0.7380747
#> loc7:lag1  0.57211957 0.39824453 0.51386126 0.49948453 0.8217702 0.8673102
#> loc8:lag1  0.39824453 0.55541996 0.42262841 0.47124805 0.6416376 0.6908474
#> loc9:lag1  0.51386126 0.42262841 0.60413446 0.52100360 0.9171879 0.9051960
#> loc10:lag1 0.49948453 0.47124805 0.52100360 0.59263327 0.7942474 0.8670323
#> loc1:lag2  0.30086915 0.23491840 0.33580378 0.29079241 0.5973950 0.5318262
#> loc2:lag2  0.31754237 0.25293525 0.33141324 0.31744065 0.5318262 0.5862572
#> loc3:lag2  0.30484583 0.22289881 0.27387194 0.27374263 0.4574612 0.4792364
#> loc4:lag2  0.32188797 0.25408911 0.31545176 0.31905051 0.5083018 0.5580035
#> loc5:lag2  0.29911248 0.26626865 0.29368739 0.32340881 0.4682996 0.5162277
#> loc6:lag2  0.28656750 0.19981063 0.26733334 0.25019011 0.4622617 0.4543253
#> loc7:lag2  0.34028868 0.23687025 0.30563745 0.29708637 0.5058444 0.5338767
#> loc8:lag2  0.23687025 0.33035598 0.25137343 0.28029171 0.3949629 0.4252542
#> loc9:lag2  0.30563745 0.25137343 0.35933068 0.30988562 0.5645792 0.5571975
#> loc10:lag2 0.29708637 0.28029171 0.30988562 0.35248994 0.4889026 0.5337057
#> loc1:lag3  0.19833065 0.15485642 0.22135930 0.19168814 0.3553221 0.3163228
#> loc2:lag3  0.20932151 0.16673299 0.21846509 0.20925446 0.3163228 0.3486976
#> loc3:lag3  0.20095205 0.14693320 0.18053430 0.18044906 0.2720915 0.2850430
#> loc4:lag3  0.21218610 0.16749361 0.20794340 0.21031567 0.3023308 0.3318926
#> loc5:lag3  0.19717267 0.17552227 0.19359649 0.21318863 0.2785380 0.3070450
#> loc6:lag3  0.18890311 0.13171364 0.17622410 0.16492341 0.2749468 0.2702263
#> loc7:lag3  0.22431570 0.15614306 0.20147388 0.19583708 0.3008691 0.3175424
#> loc8:lag3  0.15614306 0.21776815 0.16570345 0.18476616 0.2349184 0.2529352
#> loc9:lag3  0.20147388 0.16570345 0.23686805 0.20427424 0.3358038 0.3314132
#> loc10:lag3 0.19583708 0.18476616 0.20427424 0.23235868 0.2907924 0.3174407
#> loc1:lag4  0.14169915 0.11063859 0.15815218 0.13695335 0.2342256 0.2085176
#> loc2:lag4  0.14955167 0.11912391 0.15608438 0.14950377 0.2085176 0.2298588
#> loc3:lag4  0.14357203 0.10497777 0.12898438 0.12892348 0.1793607 0.1878982
#> loc4:lag4  0.15159830 0.11966734 0.14856707 0.15026196 0.1992941 0.2187811
#> loc5:lag4  0.14087182 0.12540349 0.13831679 0.15231457 0.1836101 0.2024017
#> loc6:lag4  0.13496356 0.09410402 0.12590493 0.11783104 0.1812428 0.1781311
#> loc7:lag4  0.16026441 0.11155784 0.14394486 0.13991759 0.1983306 0.2093215
#> loc8:lag4  0.11155784 0.15558645 0.11838835 0.13200788 0.1548564 0.1667330
#> loc9:lag4  0.14394486 0.11838835 0.16923255 0.14594560 0.2213593 0.2184651
#> loc10:lag4 0.13991759 0.13200788 0.14594560 0.16601079 0.1916881 0.2092545
#> loc1:lag5  0.10722367 0.08372016 0.11967367 0.10363252 0.1673447 0.1489773
#> loc2:lag5  0.11316567 0.09014100 0.11810897 0.11312942 0.1489773 0.1642247
#> loc3:lag5  0.10864088 0.07943662 0.09760241 0.09755633 0.1281459 0.1342456
#> loc4:lag5  0.11471435 0.09055221 0.11242062 0.11370314 0.1423875 0.1563102
#> loc5:lag5  0.10659763 0.09489275 0.10466424 0.11525635 0.1311820 0.1446078
#> loc6:lag5  0.10212685 0.07120846 0.09527219 0.08916269 0.1294906 0.1272674
#> loc7:lag5  0.12127199 0.08441576 0.10892300 0.10587557 0.1416991 0.1495517
#> loc8:lag5  0.08441576 0.11773218 0.08958440 0.09989029 0.1106386 0.1191239
#> loc9:lag5  0.10892300 0.08958440 0.12805818 0.11043696 0.1581522 0.1560844
#> loc10:lag5 0.10587557 0.09989029 0.11043696 0.12562027 0.1369533 0.1495038
#>            loc3:lag1 loc4:lag1 loc5:lag1  loc6:lag1 loc7:lag1  loc8:lag1
#> loc1:t+0   0.4574612 0.5083018 0.4682996 0.46226173 0.5058444 0.39496294
#> loc2:t+0   0.4792364 0.5580035 0.5162277 0.45432526 0.5338767 0.42525425
#> loc3:t+0   0.5922455 0.4924735 0.4765615 0.51358730 0.5125303 0.37475468
#> loc4:t+0   0.4924735 0.5812396 0.5330309 0.45575020 0.5411829 0.42719422
#> loc5:t+0   0.4765615 0.5330309 0.5705164 0.42714188 0.5028910 0.44767139
#> loc6:t+0   0.5135873 0.4557502 0.4271419 0.61468512 0.4817994 0.33593704
#> loc7:t+0   0.5125303 0.5411829 0.5028910 0.48179939 0.5721196 0.39824453
#> loc8:t+0   0.3747547 0.4271942 0.4476714 0.33593704 0.3982445 0.55541996
#> loc9:t+0   0.4604546 0.5303618 0.4937699 0.44946144 0.5138613 0.42262841
#> loc10:t+0  0.4602372 0.5364123 0.5437398 0.42063891 0.4994845 0.47124805
#> loc1:lag1  0.7431693 0.8257624 0.7607768 0.75096792 0.8217702 0.64163758
#> loc2:lag1  0.7785441 0.9065053 0.8386384 0.73807472 0.8673102 0.69084736
#> loc3:lag1  0.9621333 0.8000485 0.7741986 0.83434895 0.8326318 0.60880822
#> loc4:lag1  0.8000485 0.9442536 0.8659361 0.74038961 0.8791794 0.69399895
#> loc5:lag1  0.7741986 0.8659361 0.9268333 0.69391392 0.8169722 0.72726517
#> loc6:lag1  0.8343490 0.7403896 0.6939139 0.99858754 0.7827078 0.54574698
#> loc7:lag1  0.8326318 0.8791794 0.8169722 0.78270785 0.9294376 0.64696869
#> loc8:lag1  0.6088082 0.6939989 0.7272652 0.54574698 0.6469687 0.90230824
#> loc9:lag1  0.7480322 0.8616000 0.8021546 0.73017319 0.8347940 0.68658155
#> loc10:lag1 0.7476790 0.8714294 0.8833333 0.68334951 0.8114383 0.76556666
#> loc1:lag2  0.4574612 0.5083018 0.4682996 0.46226173 0.5058444 0.39496294
#> loc2:lag2  0.4792364 0.5580035 0.5162277 0.45432526 0.5338767 0.42525425
#> loc3:lag2  0.5922455 0.4924735 0.4765615 0.51358730 0.5125303 0.37475468
#> loc4:lag2  0.4924735 0.5812396 0.5330309 0.45575020 0.5411829 0.42719422
#> loc5:lag2  0.4765615 0.5330309 0.5705164 0.42714188 0.5028910 0.44767139
#> loc6:lag2  0.5135873 0.4557502 0.4271419 0.61468512 0.4817994 0.33593704
#> loc7:lag2  0.5125303 0.5411829 0.5028910 0.48179939 0.5721196 0.39824453
#> loc8:lag2  0.3747547 0.4271942 0.4476714 0.33593704 0.3982445 0.55541996
#> loc9:lag2  0.4604546 0.5303618 0.4937699 0.44946144 0.5138613 0.42262841
#> loc10:lag2 0.4602372 0.5364123 0.5437398 0.42063891 0.4994845 0.47124805
#> loc1:lag3  0.2720915 0.3023308 0.2785380 0.27494678 0.3008691 0.23491840
#> loc2:lag3  0.2850430 0.3318926 0.3070450 0.27022628 0.3175424 0.25293525
#> loc3:lag3  0.3522593 0.2929163 0.2834521 0.30547451 0.3048458 0.22289881
#> loc4:lag3  0.2929163 0.3457132 0.3170393 0.27107381 0.3218880 0.25408911
#> loc5:lag3  0.2834521 0.3170393 0.3393352 0.25405798 0.2991125 0.26626865
#> loc6:lag3  0.3054745 0.2710738 0.2540580 0.36560607 0.2865675 0.19981063
#> loc7:lag3  0.3048458 0.3218880 0.2991125 0.28656750 0.3402887 0.23687025
#> loc8:lag3  0.2228988 0.2540891 0.2662686 0.19981063 0.2368702 0.33035598
#> loc9:lag3  0.2738719 0.3154518 0.2936874 0.26733334 0.3056375 0.25137343
#> loc10:lag3 0.2737426 0.3190505 0.3234088 0.25019011 0.2970864 0.28029171
#> loc1:lag4  0.1793607 0.1992941 0.1836101 0.18124282 0.1983306 0.15485642
#> loc2:lag4  0.1878982 0.2187811 0.2024017 0.17813110 0.2093215 0.16673299
#> loc3:lag4  0.2322067 0.1930882 0.1868494 0.20136647 0.2009520 0.14693320
#> loc4:lag4  0.1930882 0.2278915 0.2089899 0.17868979 0.2121861 0.16749361
#> loc5:lag4  0.1868494 0.2089899 0.2236872 0.16747309 0.1971727 0.17552227
#> loc6:lag4  0.2013665 0.1786898 0.1674731 0.24100473 0.1889031 0.13171364
#> loc7:lag4  0.2009520 0.2121861 0.1971727 0.18890311 0.2243157 0.15614306
#> loc8:lag4  0.1469332 0.1674936 0.1755223 0.13171364 0.1561431 0.21776815
#> loc9:lag4  0.1805343 0.2079434 0.1935965 0.17622410 0.2014739 0.16570345
#> loc10:lag4 0.1804491 0.2103157 0.2131886 0.16492341 0.1958371 0.18476616
#> loc1:lag5  0.1281459 0.1423875 0.1311820 0.12949059 0.1416991 0.11063859
#> loc2:lag5  0.1342456 0.1563102 0.1446078 0.12726740 0.1495517 0.11912391
#> loc3:lag5  0.1659022 0.1379536 0.1334963 0.14386812 0.1435720 0.10497777
#> loc4:lag5  0.1379536 0.1628192 0.1493147 0.12766656 0.1515983 0.11966734
#> loc5:lag5  0.1334963 0.1493147 0.1598153 0.11965268 0.1408718 0.12540349
#> loc6:lag5  0.1438681 0.1276666 0.1196527 0.17218804 0.1349636 0.09410402
#> loc7:lag5  0.1435720 0.1515983 0.1408718 0.13496356 0.1602644 0.11155784
#> loc8:lag5  0.1049778 0.1196673 0.1254035 0.09410402 0.1115578 0.15558645
#> loc9:lag5  0.1289844 0.1485671 0.1383168 0.12590493 0.1439449 0.11838835
#> loc10:lag5 0.1289235 0.1502620 0.1523146 0.11783104 0.1399176 0.13200788
#>            loc9:lag1 loc10:lag1 loc1:lag2 loc2:lag2 loc3:lag2 loc4:lag2
#> loc1:t+0   0.5645792  0.4889026 0.3553221 0.3163228 0.2720915 0.3023308
#> loc2:t+0   0.5571975  0.5337057 0.3163228 0.3486976 0.2850430 0.3318926
#> loc3:t+0   0.4604546  0.4602372 0.2720915 0.2850430 0.3522593 0.2929163
#> loc4:t+0   0.5303618  0.5364123 0.3023308 0.3318926 0.2929163 0.3457132
#> loc5:t+0   0.4937699  0.5437398 0.2785380 0.3070450 0.2834521 0.3170393
#> loc6:t+0   0.4494614  0.4206389 0.2749468 0.2702263 0.3054745 0.2710738
#> loc7:t+0   0.5138613  0.4994845 0.3008691 0.3175424 0.3048458 0.3218880
#> loc8:t+0   0.4226284  0.4712480 0.2349184 0.2529352 0.2228988 0.2540891
#> loc9:t+0   0.6041345  0.5210036 0.3358038 0.3314132 0.2738719 0.3154518
#> loc10:t+0  0.5210036  0.5926333 0.2907924 0.3174407 0.2737426 0.3190505
#> loc1:lag1  0.9171879  0.7942474 0.5973950 0.5318262 0.4574612 0.5083018
#> loc2:lag1  0.9051960  0.8670323 0.5318262 0.5862572 0.4792364 0.5580035
#> loc3:lag1  0.7480322  0.7476790 0.4574612 0.4792364 0.5922455 0.4924735
#> loc4:lag1  0.8616000  0.8714294 0.5083018 0.5580035 0.4924735 0.5812396
#> loc5:lag1  0.8021546  0.8833333 0.4682996 0.5162277 0.4765615 0.5330309
#> loc6:lag1  0.7301732  0.6833495 0.4622617 0.4543253 0.5135873 0.4557502
#> loc7:lag1  0.8347940  0.8114383 0.5058444 0.5338767 0.5125303 0.5411829
#> loc8:lag1  0.6865816  0.7655667 0.3949629 0.4252542 0.3747547 0.4271942
#> loc9:lag1  0.9814475  0.8463971 0.5645792 0.5571975 0.4604546 0.5303618
#> loc10:lag1 0.8463971  0.9627632 0.4889026 0.5337057 0.4602372 0.5364123
#> loc1:lag2  0.5645792  0.4889026 0.9704988 0.8639790 0.7431693 0.8257624
#> loc2:lag2  0.5571975  0.5337057 0.8639790 0.9524050 0.7785441 0.9065053
#> loc3:lag2  0.4604546  0.4602372 0.7431693 0.7785441 0.9621333 0.8000485
#> loc4:lag2  0.5303618  0.5364123 0.8257624 0.9065053 0.8000485 0.9442536
#> loc5:lag2  0.4937699  0.5437398 0.7607768 0.8386384 0.7741986 0.8659361
#> loc6:lag2  0.4494614  0.4206389 0.7509679 0.7380747 0.8343490 0.7403896
#> loc7:lag2  0.5138613  0.4994845 0.8217702 0.8673102 0.8326318 0.8791794
#> loc8:lag2  0.4226284  0.4712480 0.6416376 0.6908474 0.6088082 0.6939989
#> loc9:lag2  0.6041345  0.5210036 0.9171879 0.9051960 0.7480322 0.8616000
#> loc10:lag2 0.5210036  0.5926333 0.7942474 0.8670323 0.7476790 0.8714294
#> loc1:lag3  0.3358038  0.2907924 0.5973950 0.5318262 0.4574612 0.5083018
#> loc2:lag3  0.3314132  0.3174407 0.5318262 0.5862572 0.4792364 0.5580035
#> loc3:lag3  0.2738719  0.2737426 0.4574612 0.4792364 0.5922455 0.4924735
#> loc4:lag3  0.3154518  0.3190505 0.5083018 0.5580035 0.4924735 0.5812396
#> loc5:lag3  0.2936874  0.3234088 0.4682996 0.5162277 0.4765615 0.5330309
#> loc6:lag3  0.2673333  0.2501901 0.4622617 0.4543253 0.5135873 0.4557502
#> loc7:lag3  0.3056375  0.2970864 0.5058444 0.5338767 0.5125303 0.5411829
#> loc8:lag3  0.2513734  0.2802917 0.3949629 0.4252542 0.3747547 0.4271942
#> loc9:lag3  0.3593307  0.3098856 0.5645792 0.5571975 0.4604546 0.5303618
#> loc10:lag3 0.3098856  0.3524899 0.4889026 0.5337057 0.4602372 0.5364123
#> loc1:lag4  0.2213593  0.1916881 0.3553221 0.3163228 0.2720915 0.3023308
#> loc2:lag4  0.2184651  0.2092545 0.3163228 0.3486976 0.2850430 0.3318926
#> loc3:lag4  0.1805343  0.1804491 0.2720915 0.2850430 0.3522593 0.2929163
#> loc4:lag4  0.2079434  0.2103157 0.3023308 0.3318926 0.2929163 0.3457132
#> loc5:lag4  0.1935965  0.2131886 0.2785380 0.3070450 0.2834521 0.3170393
#> loc6:lag4  0.1762241  0.1649234 0.2749468 0.2702263 0.3054745 0.2710738
#> loc7:lag4  0.2014739  0.1958371 0.3008691 0.3175424 0.3048458 0.3218880
#> loc8:lag4  0.1657035  0.1847662 0.2349184 0.2529352 0.2228988 0.2540891
#> loc9:lag4  0.2368680  0.2042742 0.3358038 0.3314132 0.2738719 0.3154518
#> loc10:lag4 0.2042742  0.2323587 0.2907924 0.3174407 0.2737426 0.3190505
#> loc1:lag5  0.1581522  0.1369533 0.2342256 0.2085176 0.1793607 0.1992941
#> loc2:lag5  0.1560844  0.1495038 0.2085176 0.2298588 0.1878982 0.2187811
#> loc3:lag5  0.1289844  0.1289235 0.1793607 0.1878982 0.2322067 0.1930882
#> loc4:lag5  0.1485671  0.1502620 0.1992941 0.2187811 0.1930882 0.2278915
#> loc5:lag5  0.1383168  0.1523146 0.1836101 0.2024017 0.1868494 0.2089899
#> loc6:lag5  0.1259049  0.1178310 0.1812428 0.1781311 0.2013665 0.1786898
#> loc7:lag5  0.1439449  0.1399176 0.1983306 0.2093215 0.2009520 0.2121861
#> loc8:lag5  0.1183884  0.1320079 0.1548564 0.1667330 0.1469332 0.1674936
#> loc9:lag5  0.1692325  0.1459456 0.2213593 0.2184651 0.1805343 0.2079434
#> loc10:lag5 0.1459456  0.1660108 0.1916881 0.2092545 0.1804491 0.2103157
#>            loc5:lag2 loc6:lag2 loc7:lag2 loc8:lag2 loc9:lag2 loc10:lag2
#> loc1:t+0   0.2785380 0.2749468 0.3008691 0.2349184 0.3358038  0.2907924
#> loc2:t+0   0.3070450 0.2702263 0.3175424 0.2529352 0.3314132  0.3174407
#> loc3:t+0   0.2834521 0.3054745 0.3048458 0.2228988 0.2738719  0.2737426
#> loc4:t+0   0.3170393 0.2710738 0.3218880 0.2540891 0.3154518  0.3190505
#> loc5:t+0   0.3393352 0.2540580 0.2991125 0.2662686 0.2936874  0.3234088
#> loc6:t+0   0.2540580 0.3656061 0.2865675 0.1998106 0.2673333  0.2501901
#> loc7:t+0   0.2991125 0.2865675 0.3402887 0.2368702 0.3056375  0.2970864
#> loc8:t+0   0.2662686 0.1998106 0.2368702 0.3303560 0.2513734  0.2802917
#> loc9:t+0   0.2936874 0.2673333 0.3056375 0.2513734 0.3593307  0.3098856
#> loc10:t+0  0.3234088 0.2501901 0.2970864 0.2802917 0.3098856  0.3524899
#> loc1:lag1  0.4682996 0.4622617 0.5058444 0.3949629 0.5645792  0.4889026
#> loc2:lag1  0.5162277 0.4543253 0.5338767 0.4252542 0.5571975  0.5337057
#> loc3:lag1  0.4765615 0.5135873 0.5125303 0.3747547 0.4604546  0.4602372
#> loc4:lag1  0.5330309 0.4557502 0.5411829 0.4271942 0.5303618  0.5364123
#> loc5:lag1  0.5705164 0.4271419 0.5028910 0.4476714 0.4937699  0.5437398
#> loc6:lag1  0.4271419 0.6146851 0.4817994 0.3359370 0.4494614  0.4206389
#> loc7:lag1  0.5028910 0.4817994 0.5721196 0.3982445 0.5138613  0.4994845
#> loc8:lag1  0.4476714 0.3359370 0.3982445 0.5554200 0.4226284  0.4712480
#> loc9:lag1  0.4937699 0.4494614 0.5138613 0.4226284 0.6041345  0.5210036
#> loc10:lag1 0.5437398 0.4206389 0.4994845 0.4712480 0.5210036  0.5926333
#> loc1:lag2  0.7607768 0.7509679 0.8217702 0.6416376 0.9171879  0.7942474
#> loc2:lag2  0.8386384 0.7380747 0.8673102 0.6908474 0.9051960  0.8670323
#> loc3:lag2  0.7741986 0.8343490 0.8326318 0.6088082 0.7480322  0.7476790
#> loc4:lag2  0.8659361 0.7403896 0.8791794 0.6939989 0.8616000  0.8714294
#> loc5:lag2  0.9268333 0.6939139 0.8169722 0.7272652 0.8021546  0.8833333
#> loc6:lag2  0.6939139 0.9985875 0.7827078 0.5457470 0.7301732  0.6833495
#> loc7:lag2  0.8169722 0.7827078 0.9294376 0.6469687 0.8347940  0.8114383
#> loc8:lag2  0.7272652 0.5457470 0.6469687 0.9023082 0.6865816  0.7655667
#> loc9:lag2  0.8021546 0.7301732 0.8347940 0.6865816 0.9814475  0.8463971
#> loc10:lag2 0.8833333 0.6833495 0.8114383 0.7655667 0.8463971  0.9627632
#> loc1:lag3  0.4682996 0.4622617 0.5058444 0.3949629 0.5645792  0.4889026
#> loc2:lag3  0.5162277 0.4543253 0.5338767 0.4252542 0.5571975  0.5337057
#> loc3:lag3  0.4765615 0.5135873 0.5125303 0.3747547 0.4604546  0.4602372
#> loc4:lag3  0.5330309 0.4557502 0.5411829 0.4271942 0.5303618  0.5364123
#> loc5:lag3  0.5705164 0.4271419 0.5028910 0.4476714 0.4937699  0.5437398
#> loc6:lag3  0.4271419 0.6146851 0.4817994 0.3359370 0.4494614  0.4206389
#> loc7:lag3  0.5028910 0.4817994 0.5721196 0.3982445 0.5138613  0.4994845
#> loc8:lag3  0.4476714 0.3359370 0.3982445 0.5554200 0.4226284  0.4712480
#> loc9:lag3  0.4937699 0.4494614 0.5138613 0.4226284 0.6041345  0.5210036
#> loc10:lag3 0.5437398 0.4206389 0.4994845 0.4712480 0.5210036  0.5926333
#> loc1:lag4  0.2785380 0.2749468 0.3008691 0.2349184 0.3358038  0.2907924
#> loc2:lag4  0.3070450 0.2702263 0.3175424 0.2529352 0.3314132  0.3174407
#> loc3:lag4  0.2834521 0.3054745 0.3048458 0.2228988 0.2738719  0.2737426
#> loc4:lag4  0.3170393 0.2710738 0.3218880 0.2540891 0.3154518  0.3190505
#> loc5:lag4  0.3393352 0.2540580 0.2991125 0.2662686 0.2936874  0.3234088
#> loc6:lag4  0.2540580 0.3656061 0.2865675 0.1998106 0.2673333  0.2501901
#> loc7:lag4  0.2991125 0.2865675 0.3402887 0.2368702 0.3056375  0.2970864
#> loc8:lag4  0.2662686 0.1998106 0.2368702 0.3303560 0.2513734  0.2802917
#> loc9:lag4  0.2936874 0.2673333 0.3056375 0.2513734 0.3593307  0.3098856
#> loc10:lag4 0.3234088 0.2501901 0.2970864 0.2802917 0.3098856  0.3524899
#> loc1:lag5  0.1836101 0.1812428 0.1983306 0.1548564 0.2213593  0.1916881
#> loc2:lag5  0.2024017 0.1781311 0.2093215 0.1667330 0.2184651  0.2092545
#> loc3:lag5  0.1868494 0.2013665 0.2009520 0.1469332 0.1805343  0.1804491
#> loc4:lag5  0.2089899 0.1786898 0.2121861 0.1674936 0.2079434  0.2103157
#> loc5:lag5  0.2236872 0.1674731 0.1971727 0.1755223 0.1935965  0.2131886
#> loc6:lag5  0.1674731 0.2410047 0.1889031 0.1317136 0.1762241  0.1649234
#> loc7:lag5  0.1971727 0.1889031 0.2243157 0.1561431 0.2014739  0.1958371
#> loc8:lag5  0.1755223 0.1317136 0.1561431 0.2177681 0.1657035  0.1847662
#> loc9:lag5  0.1935965 0.1762241 0.2014739 0.1657035 0.2368680  0.2042742
#> loc10:lag5 0.2131886 0.1649234 0.1958371 0.1847662 0.2042742  0.2323587
#>            loc1:lag3 loc2:lag3 loc3:lag3 loc4:lag3 loc5:lag3 loc6:lag3
#> loc1:t+0   0.2342256 0.2085176 0.1793607 0.1992941 0.1836101 0.1812428
#> loc2:t+0   0.2085176 0.2298588 0.1878982 0.2187811 0.2024017 0.1781311
#> loc3:t+0   0.1793607 0.1878982 0.2322067 0.1930882 0.1868494 0.2013665
#> loc4:t+0   0.1992941 0.2187811 0.1930882 0.2278915 0.2089899 0.1786898
#> loc5:t+0   0.1836101 0.2024017 0.1868494 0.2089899 0.2236872 0.1674731
#> loc6:t+0   0.1812428 0.1781311 0.2013665 0.1786898 0.1674731 0.2410047
#> loc7:t+0   0.1983306 0.2093215 0.2009520 0.2121861 0.1971727 0.1889031
#> loc8:t+0   0.1548564 0.1667330 0.1469332 0.1674936 0.1755223 0.1317136
#> loc9:t+0   0.2213593 0.2184651 0.1805343 0.2079434 0.1935965 0.1762241
#> loc10:t+0  0.1916881 0.2092545 0.1804491 0.2103157 0.2131886 0.1649234
#> loc1:lag1  0.3553221 0.3163228 0.2720915 0.3023308 0.2785380 0.2749468
#> loc2:lag1  0.3163228 0.3486976 0.2850430 0.3318926 0.3070450 0.2702263
#> loc3:lag1  0.2720915 0.2850430 0.3522593 0.2929163 0.2834521 0.3054745
#> loc4:lag1  0.3023308 0.3318926 0.2929163 0.3457132 0.3170393 0.2710738
#> loc5:lag1  0.2785380 0.3070450 0.2834521 0.3170393 0.3393352 0.2540580
#> loc6:lag1  0.2749468 0.2702263 0.3054745 0.2710738 0.2540580 0.3656061
#> loc7:lag1  0.3008691 0.3175424 0.3048458 0.3218880 0.2991125 0.2865675
#> loc8:lag1  0.2349184 0.2529352 0.2228988 0.2540891 0.2662686 0.1998106
#> loc9:lag1  0.3358038 0.3314132 0.2738719 0.3154518 0.2936874 0.2673333
#> loc10:lag1 0.2907924 0.3174407 0.2737426 0.3190505 0.3234088 0.2501901
#> loc1:lag2  0.5973950 0.5318262 0.4574612 0.5083018 0.4682996 0.4622617
#> loc2:lag2  0.5318262 0.5862572 0.4792364 0.5580035 0.5162277 0.4543253
#> loc3:lag2  0.4574612 0.4792364 0.5922455 0.4924735 0.4765615 0.5135873
#> loc4:lag2  0.5083018 0.5580035 0.4924735 0.5812396 0.5330309 0.4557502
#> loc5:lag2  0.4682996 0.5162277 0.4765615 0.5330309 0.5705164 0.4271419
#> loc6:lag2  0.4622617 0.4543253 0.5135873 0.4557502 0.4271419 0.6146851
#> loc7:lag2  0.5058444 0.5338767 0.5125303 0.5411829 0.5028910 0.4817994
#> loc8:lag2  0.3949629 0.4252542 0.3747547 0.4271942 0.4476714 0.3359370
#> loc9:lag2  0.5645792 0.5571975 0.4604546 0.5303618 0.4937699 0.4494614
#> loc10:lag2 0.4889026 0.5337057 0.4602372 0.5364123 0.5437398 0.4206389
#> loc1:lag3  0.9704988 0.8639790 0.7431693 0.8257624 0.7607768 0.7509679
#> loc2:lag3  0.8639790 0.9524050 0.7785441 0.9065053 0.8386384 0.7380747
#> loc3:lag3  0.7431693 0.7785441 0.9621333 0.8000485 0.7741986 0.8343490
#> loc4:lag3  0.8257624 0.9065053 0.8000485 0.9442536 0.8659361 0.7403896
#> loc5:lag3  0.7607768 0.8386384 0.7741986 0.8659361 0.9268333 0.6939139
#> loc6:lag3  0.7509679 0.7380747 0.8343490 0.7403896 0.6939139 0.9985875
#> loc7:lag3  0.8217702 0.8673102 0.8326318 0.8791794 0.8169722 0.7827078
#> loc8:lag3  0.6416376 0.6908474 0.6088082 0.6939989 0.7272652 0.5457470
#> loc9:lag3  0.9171879 0.9051960 0.7480322 0.8616000 0.8021546 0.7301732
#> loc10:lag3 0.7942474 0.8670323 0.7476790 0.8714294 0.8833333 0.6833495
#> loc1:lag4  0.5973950 0.5318262 0.4574612 0.5083018 0.4682996 0.4622617
#> loc2:lag4  0.5318262 0.5862572 0.4792364 0.5580035 0.5162277 0.4543253
#> loc3:lag4  0.4574612 0.4792364 0.5922455 0.4924735 0.4765615 0.5135873
#> loc4:lag4  0.5083018 0.5580035 0.4924735 0.5812396 0.5330309 0.4557502
#> loc5:lag4  0.4682996 0.5162277 0.4765615 0.5330309 0.5705164 0.4271419
#> loc6:lag4  0.4622617 0.4543253 0.5135873 0.4557502 0.4271419 0.6146851
#> loc7:lag4  0.5058444 0.5338767 0.5125303 0.5411829 0.5028910 0.4817994
#> loc8:lag4  0.3949629 0.4252542 0.3747547 0.4271942 0.4476714 0.3359370
#> loc9:lag4  0.5645792 0.5571975 0.4604546 0.5303618 0.4937699 0.4494614
#> loc10:lag4 0.4889026 0.5337057 0.4602372 0.5364123 0.5437398 0.4206389
#> loc1:lag5  0.3553221 0.3163228 0.2720915 0.3023308 0.2785380 0.2749468
#> loc2:lag5  0.3163228 0.3486976 0.2850430 0.3318926 0.3070450 0.2702263
#> loc3:lag5  0.2720915 0.2850430 0.3522593 0.2929163 0.2834521 0.3054745
#> loc4:lag5  0.3023308 0.3318926 0.2929163 0.3457132 0.3170393 0.2710738
#> loc5:lag5  0.2785380 0.3070450 0.2834521 0.3170393 0.3393352 0.2540580
#> loc6:lag5  0.2749468 0.2702263 0.3054745 0.2710738 0.2540580 0.3656061
#> loc7:lag5  0.3008691 0.3175424 0.3048458 0.3218880 0.2991125 0.2865675
#> loc8:lag5  0.2349184 0.2529352 0.2228988 0.2540891 0.2662686 0.1998106
#> loc9:lag5  0.3358038 0.3314132 0.2738719 0.3154518 0.2936874 0.2673333
#> loc10:lag5 0.2907924 0.3174407 0.2737426 0.3190505 0.3234088 0.2501901
#>            loc7:lag3 loc8:lag3 loc9:lag3 loc10:lag3 loc1:lag4 loc2:lag4
#> loc1:t+0   0.1983306 0.1548564 0.2213593  0.1916881 0.1673447 0.1489773
#> loc2:t+0   0.2093215 0.1667330 0.2184651  0.2092545 0.1489773 0.1642247
#> loc3:t+0   0.2009520 0.1469332 0.1805343  0.1804491 0.1281459 0.1342456
#> loc4:t+0   0.2121861 0.1674936 0.2079434  0.2103157 0.1423875 0.1563102
#> loc5:t+0   0.1971727 0.1755223 0.1935965  0.2131886 0.1311820 0.1446078
#> loc6:t+0   0.1889031 0.1317136 0.1762241  0.1649234 0.1294906 0.1272674
#> loc7:t+0   0.2243157 0.1561431 0.2014739  0.1958371 0.1416991 0.1495517
#> loc8:t+0   0.1561431 0.2177681 0.1657035  0.1847662 0.1106386 0.1191239
#> loc9:t+0   0.2014739 0.1657035 0.2368680  0.2042742 0.1581522 0.1560844
#> loc10:t+0  0.1958371 0.1847662 0.2042742  0.2323587 0.1369533 0.1495038
#> loc1:lag1  0.3008691 0.2349184 0.3358038  0.2907924 0.2342256 0.2085176
#> loc2:lag1  0.3175424 0.2529352 0.3314132  0.3174407 0.2085176 0.2298588
#> loc3:lag1  0.3048458 0.2228988 0.2738719  0.2737426 0.1793607 0.1878982
#> loc4:lag1  0.3218880 0.2540891 0.3154518  0.3190505 0.1992941 0.2187811
#> loc5:lag1  0.2991125 0.2662686 0.2936874  0.3234088 0.1836101 0.2024017
#> loc6:lag1  0.2865675 0.1998106 0.2673333  0.2501901 0.1812428 0.1781311
#> loc7:lag1  0.3402887 0.2368702 0.3056375  0.2970864 0.1983306 0.2093215
#> loc8:lag1  0.2368702 0.3303560 0.2513734  0.2802917 0.1548564 0.1667330
#> loc9:lag1  0.3056375 0.2513734 0.3593307  0.3098856 0.2213593 0.2184651
#> loc10:lag1 0.2970864 0.2802917 0.3098856  0.3524899 0.1916881 0.2092545
#> loc1:lag2  0.5058444 0.3949629 0.5645792  0.4889026 0.3553221 0.3163228
#> loc2:lag2  0.5338767 0.4252542 0.5571975  0.5337057 0.3163228 0.3486976
#> loc3:lag2  0.5125303 0.3747547 0.4604546  0.4602372 0.2720915 0.2850430
#> loc4:lag2  0.5411829 0.4271942 0.5303618  0.5364123 0.3023308 0.3318926
#> loc5:lag2  0.5028910 0.4476714 0.4937699  0.5437398 0.2785380 0.3070450
#> loc6:lag2  0.4817994 0.3359370 0.4494614  0.4206389 0.2749468 0.2702263
#> loc7:lag2  0.5721196 0.3982445 0.5138613  0.4994845 0.3008691 0.3175424
#> loc8:lag2  0.3982445 0.5554200 0.4226284  0.4712480 0.2349184 0.2529352
#> loc9:lag2  0.5138613 0.4226284 0.6041345  0.5210036 0.3358038 0.3314132
#> loc10:lag2 0.4994845 0.4712480 0.5210036  0.5926333 0.2907924 0.3174407
#> loc1:lag3  0.8217702 0.6416376 0.9171879  0.7942474 0.5973950 0.5318262
#> loc2:lag3  0.8673102 0.6908474 0.9051960  0.8670323 0.5318262 0.5862572
#> loc3:lag3  0.8326318 0.6088082 0.7480322  0.7476790 0.4574612 0.4792364
#> loc4:lag3  0.8791794 0.6939989 0.8616000  0.8714294 0.5083018 0.5580035
#> loc5:lag3  0.8169722 0.7272652 0.8021546  0.8833333 0.4682996 0.5162277
#> loc6:lag3  0.7827078 0.5457470 0.7301732  0.6833495 0.4622617 0.4543253
#> loc7:lag3  0.9294376 0.6469687 0.8347940  0.8114383 0.5058444 0.5338767
#> loc8:lag3  0.6469687 0.9023082 0.6865816  0.7655667 0.3949629 0.4252542
#> loc9:lag3  0.8347940 0.6865816 0.9814475  0.8463971 0.5645792 0.5571975
#> loc10:lag3 0.8114383 0.7655667 0.8463971  0.9627632 0.4889026 0.5337057
#> loc1:lag4  0.5058444 0.3949629 0.5645792  0.4889026 0.9704988 0.8639790
#> loc2:lag4  0.5338767 0.4252542 0.5571975  0.5337057 0.8639790 0.9524050
#> loc3:lag4  0.5125303 0.3747547 0.4604546  0.4602372 0.7431693 0.7785441
#> loc4:lag4  0.5411829 0.4271942 0.5303618  0.5364123 0.8257624 0.9065053
#> loc5:lag4  0.5028910 0.4476714 0.4937699  0.5437398 0.7607768 0.8386384
#> loc6:lag4  0.4817994 0.3359370 0.4494614  0.4206389 0.7509679 0.7380747
#> loc7:lag4  0.5721196 0.3982445 0.5138613  0.4994845 0.8217702 0.8673102
#> loc8:lag4  0.3982445 0.5554200 0.4226284  0.4712480 0.6416376 0.6908474
#> loc9:lag4  0.5138613 0.4226284 0.6041345  0.5210036 0.9171879 0.9051960
#> loc10:lag4 0.4994845 0.4712480 0.5210036  0.5926333 0.7942474 0.8670323
#> loc1:lag5  0.3008691 0.2349184 0.3358038  0.2907924 0.5973950 0.5318262
#> loc2:lag5  0.3175424 0.2529352 0.3314132  0.3174407 0.5318262 0.5862572
#> loc3:lag5  0.3048458 0.2228988 0.2738719  0.2737426 0.4574612 0.4792364
#> loc4:lag5  0.3218880 0.2540891 0.3154518  0.3190505 0.5083018 0.5580035
#> loc5:lag5  0.2991125 0.2662686 0.2936874  0.3234088 0.4682996 0.5162277
#> loc6:lag5  0.2865675 0.1998106 0.2673333  0.2501901 0.4622617 0.4543253
#> loc7:lag5  0.3402887 0.2368702 0.3056375  0.2970864 0.5058444 0.5338767
#> loc8:lag5  0.2368702 0.3303560 0.2513734  0.2802917 0.3949629 0.4252542
#> loc9:lag5  0.3056375 0.2513734 0.3593307  0.3098856 0.5645792 0.5571975
#> loc10:lag5 0.2970864 0.2802917 0.3098856  0.3524899 0.4889026 0.5337057
#>            loc3:lag4 loc4:lag4 loc5:lag4  loc6:lag4 loc7:lag4  loc8:lag4
#> loc1:t+0   0.1281459 0.1423875 0.1311820 0.12949059 0.1416991 0.11063859
#> loc2:t+0   0.1342456 0.1563102 0.1446078 0.12726740 0.1495517 0.11912391
#> loc3:t+0   0.1659022 0.1379536 0.1334963 0.14386812 0.1435720 0.10497777
#> loc4:t+0   0.1379536 0.1628192 0.1493147 0.12766656 0.1515983 0.11966734
#> loc5:t+0   0.1334963 0.1493147 0.1598153 0.11965268 0.1408718 0.12540349
#> loc6:t+0   0.1438681 0.1276666 0.1196527 0.17218804 0.1349636 0.09410402
#> loc7:t+0   0.1435720 0.1515983 0.1408718 0.13496356 0.1602644 0.11155784
#> loc8:t+0   0.1049778 0.1196673 0.1254035 0.09410402 0.1115578 0.15558645
#> loc9:t+0   0.1289844 0.1485671 0.1383168 0.12590493 0.1439449 0.11838835
#> loc10:t+0  0.1289235 0.1502620 0.1523146 0.11783104 0.1399176 0.13200788
#> loc1:lag1  0.1793607 0.1992941 0.1836101 0.18124282 0.1983306 0.15485642
#> loc2:lag1  0.1878982 0.2187811 0.2024017 0.17813110 0.2093215 0.16673299
#> loc3:lag1  0.2322067 0.1930882 0.1868494 0.20136647 0.2009520 0.14693320
#> loc4:lag1  0.1930882 0.2278915 0.2089899 0.17868979 0.2121861 0.16749361
#> loc5:lag1  0.1868494 0.2089899 0.2236872 0.16747309 0.1971727 0.17552227
#> loc6:lag1  0.2013665 0.1786898 0.1674731 0.24100473 0.1889031 0.13171364
#> loc7:lag1  0.2009520 0.2121861 0.1971727 0.18890311 0.2243157 0.15614306
#> loc8:lag1  0.1469332 0.1674936 0.1755223 0.13171364 0.1561431 0.21776815
#> loc9:lag1  0.1805343 0.2079434 0.1935965 0.17622410 0.2014739 0.16570345
#> loc10:lag1 0.1804491 0.2103157 0.2131886 0.16492341 0.1958371 0.18476616
#> loc1:lag2  0.2720915 0.3023308 0.2785380 0.27494678 0.3008691 0.23491840
#> loc2:lag2  0.2850430 0.3318926 0.3070450 0.27022628 0.3175424 0.25293525
#> loc3:lag2  0.3522593 0.2929163 0.2834521 0.30547451 0.3048458 0.22289881
#> loc4:lag2  0.2929163 0.3457132 0.3170393 0.27107381 0.3218880 0.25408911
#> loc5:lag2  0.2834521 0.3170393 0.3393352 0.25405798 0.2991125 0.26626865
#> loc6:lag2  0.3054745 0.2710738 0.2540580 0.36560607 0.2865675 0.19981063
#> loc7:lag2  0.3048458 0.3218880 0.2991125 0.28656750 0.3402887 0.23687025
#> loc8:lag2  0.2228988 0.2540891 0.2662686 0.19981063 0.2368702 0.33035598
#> loc9:lag2  0.2738719 0.3154518 0.2936874 0.26733334 0.3056375 0.25137343
#> loc10:lag2 0.2737426 0.3190505 0.3234088 0.25019011 0.2970864 0.28029171
#> loc1:lag3  0.4574612 0.5083018 0.4682996 0.46226173 0.5058444 0.39496294
#> loc2:lag3  0.4792364 0.5580035 0.5162277 0.45432526 0.5338767 0.42525425
#> loc3:lag3  0.5922455 0.4924735 0.4765615 0.51358730 0.5125303 0.37475468
#> loc4:lag3  0.4924735 0.5812396 0.5330309 0.45575020 0.5411829 0.42719422
#> loc5:lag3  0.4765615 0.5330309 0.5705164 0.42714188 0.5028910 0.44767139
#> loc6:lag3  0.5135873 0.4557502 0.4271419 0.61468512 0.4817994 0.33593704
#> loc7:lag3  0.5125303 0.5411829 0.5028910 0.48179939 0.5721196 0.39824453
#> loc8:lag3  0.3747547 0.4271942 0.4476714 0.33593704 0.3982445 0.55541996
#> loc9:lag3  0.4604546 0.5303618 0.4937699 0.44946144 0.5138613 0.42262841
#> loc10:lag3 0.4602372 0.5364123 0.5437398 0.42063891 0.4994845 0.47124805
#> loc1:lag4  0.7431693 0.8257624 0.7607768 0.75096792 0.8217702 0.64163758
#> loc2:lag4  0.7785441 0.9065053 0.8386384 0.73807472 0.8673102 0.69084736
#> loc3:lag4  0.9621333 0.8000485 0.7741986 0.83434895 0.8326318 0.60880822
#> loc4:lag4  0.8000485 0.9442536 0.8659361 0.74038961 0.8791794 0.69399895
#> loc5:lag4  0.7741986 0.8659361 0.9268333 0.69391392 0.8169722 0.72726517
#> loc6:lag4  0.8343490 0.7403896 0.6939139 0.99858754 0.7827078 0.54574698
#> loc7:lag4  0.8326318 0.8791794 0.8169722 0.78270785 0.9294376 0.64696869
#> loc8:lag4  0.6088082 0.6939989 0.7272652 0.54574698 0.6469687 0.90230824
#> loc9:lag4  0.7480322 0.8616000 0.8021546 0.73017319 0.8347940 0.68658155
#> loc10:lag4 0.7476790 0.8714294 0.8833333 0.68334951 0.8114383 0.76556666
#> loc1:lag5  0.4574612 0.5083018 0.4682996 0.46226173 0.5058444 0.39496294
#> loc2:lag5  0.4792364 0.5580035 0.5162277 0.45432526 0.5338767 0.42525425
#> loc3:lag5  0.5922455 0.4924735 0.4765615 0.51358730 0.5125303 0.37475468
#> loc4:lag5  0.4924735 0.5812396 0.5330309 0.45575020 0.5411829 0.42719422
#> loc5:lag5  0.4765615 0.5330309 0.5705164 0.42714188 0.5028910 0.44767139
#> loc6:lag5  0.5135873 0.4557502 0.4271419 0.61468512 0.4817994 0.33593704
#> loc7:lag5  0.5125303 0.5411829 0.5028910 0.48179939 0.5721196 0.39824453
#> loc8:lag5  0.3747547 0.4271942 0.4476714 0.33593704 0.3982445 0.55541996
#> loc9:lag5  0.4604546 0.5303618 0.4937699 0.44946144 0.5138613 0.42262841
#> loc10:lag5 0.4602372 0.5364123 0.5437398 0.42063891 0.4994845 0.47124805
#>            loc9:lag4 loc10:lag4  loc1:lag5  loc2:lag5  loc3:lag5  loc4:lag5
#> loc1:t+0   0.1581522  0.1369533 0.12662961 0.11273102 0.09696790 0.10774456
#> loc2:t+0   0.1560844  0.1495038 0.11273102 0.12426875 0.10158357 0.11827981
#> loc3:t+0   0.1289844  0.1289235 0.09696790 0.10158357 0.12553808 0.10438945
#> loc4:t+0   0.1485671  0.1502620 0.10774456 0.11827981 0.10438945 0.12320517
#> loc5:t+0   0.1383168  0.1523146 0.09926531 0.10942461 0.10101657 0.11298638
#> loc6:t+0   0.1259049  0.1178310 0.09798546 0.09630317 0.10886493 0.09660522
#> loc7:t+0   0.1439449  0.1399176 0.10722367 0.11316567 0.10864088 0.11471435
#> loc8:t+0   0.1183884  0.1320079 0.08372016 0.09014100 0.07943662 0.09055221
#> loc9:t+0   0.1692325  0.1459456 0.11967367 0.11810897 0.09760241 0.11242062
#> loc10:t+0  0.1459456  0.1660108 0.10363252 0.11312942 0.09755633 0.11370314
#> loc1:lag1  0.2213593  0.1916881 0.16734466 0.14897728 0.12814586 0.14238753
#> loc2:lag1  0.2184651  0.2092545 0.14897728 0.16422471 0.13424559 0.15631016
#> loc3:lag1  0.1805343  0.1804491 0.12814586 0.13424559 0.16590217 0.13795364
#> loc4:lag1  0.2079434  0.2103157 0.14238753 0.15631016 0.13795364 0.16281916
#> loc5:lag1  0.1935965  0.2131886 0.13118195 0.14460776 0.13349630 0.14931474
#> loc6:lag1  0.1762241  0.1649234 0.12949059 0.12726740 0.14386812 0.12766656
#> loc7:lag1  0.2014739  0.1958371 0.14169915 0.14955167 0.14357203 0.15159830
#> loc8:lag1  0.1657035  0.1847662 0.11063859 0.11912391 0.10497777 0.11966734
#> loc9:lag1  0.2368680  0.2042742 0.15815218 0.15608438 0.12898438 0.14856707
#> loc10:lag1 0.2042742  0.2323587 0.13695335 0.14950377 0.12892348 0.15026196
#> loc1:lag2  0.3358038  0.2907924 0.23422564 0.20851755 0.17936066 0.19929414
#> loc2:lag2  0.3314132  0.3174407 0.20851755 0.22985878 0.18789820 0.21878110
#> loc3:lag2  0.2738719  0.2737426 0.17936066 0.18789820 0.23220665 0.19308821
#> loc4:lag2  0.3154518  0.3190505 0.19929414 0.21878110 0.19308821 0.22789149
#> loc5:lag2  0.2936874  0.3234088 0.18361015 0.20240171 0.18684944 0.20898988
#> loc6:lag2  0.2673333  0.2501901 0.18124282 0.17813110 0.20136647 0.17868979
#> loc7:lag2  0.3056375  0.2970864 0.19833065 0.20932151 0.20095205 0.21218610
#> loc8:lag2  0.2513734  0.2802917 0.15485642 0.16673299 0.14693320 0.16749361
#> loc9:lag2  0.3593307  0.3098856 0.22135930 0.21846509 0.18053430 0.20794340
#> loc10:lag2 0.3098856  0.3524899 0.19168814 0.20925446 0.18044906 0.21031567
#> loc1:lag3  0.5645792  0.4889026 0.35532213 0.31632276 0.27209152 0.30233077
#> loc2:lag3  0.5571975  0.5337057 0.31632276 0.34869756 0.28504304 0.33189264
#> loc3:lag3  0.4604546  0.4602372 0.27209152 0.28504304 0.35225930 0.29291633
#> loc4:lag3  0.5303618  0.5364123 0.30233077 0.33189264 0.29291633 0.34571316
#> loc5:lag3  0.4937699  0.5437398 0.27853803 0.30704498 0.28345206 0.31703929
#> loc6:lag3  0.4494614  0.4206389 0.27494678 0.27022628 0.30547451 0.27107381
#> loc7:lag3  0.5138613  0.4994845 0.30086915 0.31754237 0.30484583 0.32188797
#> loc8:lag3  0.4226284  0.4712480 0.23491840 0.25293525 0.22289881 0.25408911
#> loc9:lag3  0.6041345  0.5210036 0.33580378 0.33141324 0.27387194 0.31545176
#> loc10:lag3 0.5210036  0.5926333 0.29079241 0.31744065 0.27374263 0.31905051
#> loc1:lag4  0.9171879  0.7942474 0.59739497 0.53182622 0.45746124 0.50830182
#> loc2:lag4  0.9051960  0.8670323 0.53182622 0.58625724 0.47923635 0.55800350
#> loc3:lag4  0.7480322  0.7476790 0.45746124 0.47923635 0.59224551 0.49247352
#> loc4:lag4  0.8616000  0.8714294 0.50830182 0.55800350 0.49247352 0.58123964
#> loc5:lag4  0.8021546  0.8833333 0.46829962 0.51622770 0.47656147 0.53303090
#> loc6:lag4  0.7301732  0.6833495 0.46226173 0.45432526 0.51358730 0.45575020
#> loc7:lag4  0.8347940  0.8114383 0.50584442 0.53387673 0.51253032 0.54118288
#> loc8:lag4  0.6865816  0.7655667 0.39496294 0.42525425 0.37475468 0.42719422
#> loc9:lag4  0.9814475  0.8463971 0.56457922 0.55719751 0.46045462 0.53036183
#> loc10:lag4 0.8463971  0.9627632 0.48890263 0.53370571 0.46023723 0.53641233
#> loc1:lag5  0.5645792  0.4889026 0.97049881 0.86397901 0.74316928 0.82576241
#> loc2:lag5  0.5571975  0.5337057 0.86397901 0.95240499 0.77854406 0.90650535
#> loc3:lag5  0.4604546  0.4602372 0.74316928 0.77854406 0.96213325 0.80004853
#> loc4:lag5  0.5303618  0.5364123 0.82576241 0.90650535 0.80004853 0.94425364
#> loc5:lag5  0.4937699  0.5437398 0.76077679 0.83863841 0.77419859 0.86593608
#> loc6:lag5  0.4494614  0.4206389 0.75096792 0.73807472 0.83434895 0.74038961
#> loc7:lag5  0.5138613  0.4994845 0.82177023 0.86731017 0.83263182 0.87917939
#> loc8:lag5  0.4226284  0.4712480 0.64163758 0.69084736 0.60880822 0.69399895
#> loc9:lag5  0.6041345  0.5210036 0.91718794 0.90519597 0.74803219 0.86160003
#> loc10:lag5 0.5210036  0.5926333 0.79424743 0.86703233 0.74767901 0.87142938
#>             loc5:lag5  loc6:lag5  loc7:lag5  loc8:lag5  loc9:lag5 loc10:lag5
#> loc1:t+0   0.09926531 0.09798546 0.10722367 0.08372016 0.11967367 0.10363252
#> loc2:t+0   0.10942461 0.09630317 0.11316567 0.09014100 0.11810897 0.11312942
#> loc3:t+0   0.10101657 0.10886493 0.10864088 0.07943662 0.09760241 0.09755633
#> loc4:t+0   0.11298638 0.09660522 0.11471435 0.09055221 0.11242062 0.11370314
#> loc5:t+0   0.12093218 0.09054112 0.10659763 0.09489275 0.10466424 0.11525635
#> loc6:t+0   0.09054112 0.13029460 0.10212685 0.07120846 0.09527219 0.08916269
#> loc7:t+0   0.10659763 0.10212685 0.12127199 0.08441576 0.10892300 0.10587557
#> loc8:t+0   0.09489275 0.07120846 0.08441576 0.11773218 0.08958440 0.09989029
#> loc9:t+0   0.10466424 0.09527219 0.10892300 0.08958440 0.12805818 0.11043696
#> loc10:t+0  0.11525635 0.08916269 0.10587557 0.09989029 0.11043696 0.12562027
#> loc1:lag1  0.13118195 0.12949059 0.14169915 0.11063859 0.15815218 0.13695335
#> loc2:lag1  0.14460776 0.12726740 0.14955167 0.11912391 0.15608438 0.14950377
#> loc3:lag1  0.13349630 0.14386812 0.14357203 0.10497777 0.12898438 0.12892348
#> loc4:lag1  0.14931474 0.12766656 0.15159830 0.11966734 0.14856707 0.15026196
#> loc5:lag1  0.15981534 0.11965268 0.14087182 0.12540349 0.13831679 0.15231457
#> loc6:lag1  0.11965268 0.17218804 0.13496356 0.09410402 0.12590493 0.11783104
#> loc7:lag1  0.14087182 0.13496356 0.16026441 0.11155784 0.14394486 0.13991759
#> loc8:lag1  0.12540349 0.09410402 0.11155784 0.15558645 0.11838835 0.13200788
#> loc9:lag1  0.13831679 0.12590493 0.14394486 0.11838835 0.16923255 0.14594560
#> loc10:lag1 0.15231457 0.11783104 0.13991759 0.13200788 0.14594560 0.16601079
#> loc1:lag2  0.18361015 0.18124282 0.19833065 0.15485642 0.22135930 0.19168814
#> loc2:lag2  0.20240171 0.17813110 0.20932151 0.16673299 0.21846509 0.20925446
#> loc3:lag2  0.18684944 0.20136647 0.20095205 0.14693320 0.18053430 0.18044906
#> loc4:lag2  0.20898988 0.17868979 0.21218610 0.16749361 0.20794340 0.21031567
#> loc5:lag2  0.22368715 0.16747309 0.19717267 0.17552227 0.19359649 0.21318863
#> loc6:lag2  0.16747309 0.24100473 0.18890311 0.13171364 0.17622410 0.16492341
#> loc7:lag2  0.19717267 0.18890311 0.22431570 0.15614306 0.20147388 0.19583708
#> loc8:lag2  0.17552227 0.13171364 0.15614306 0.21776815 0.16570345 0.18476616
#> loc9:lag2  0.19359649 0.17622410 0.20147388 0.16570345 0.23686805 0.20427424
#> loc10:lag2 0.21318863 0.16492341 0.19583708 0.18476616 0.20427424 0.23235868
#> loc1:lag3  0.27853803 0.27494678 0.30086915 0.23491840 0.33580378 0.29079241
#> loc2:lag3  0.30704498 0.27022628 0.31754237 0.25293525 0.33141324 0.31744065
#> loc3:lag3  0.28345206 0.30547451 0.30484583 0.22289881 0.27387194 0.27374263
#> loc4:lag3  0.31703929 0.27107381 0.32188797 0.25408911 0.31545176 0.31905051
#> loc5:lag3  0.33933516 0.25405798 0.29911248 0.26626865 0.29368739 0.32340881
#> loc6:lag3  0.25405798 0.36560607 0.28656750 0.19981063 0.26733334 0.25019011
#> loc7:lag3  0.29911248 0.28656750 0.34028868 0.23687025 0.30563745 0.29708637
#> loc8:lag3  0.26626865 0.19981063 0.23687025 0.33035598 0.25137343 0.28029171
#> loc9:lag3  0.29368739 0.26733334 0.30563745 0.25137343 0.35933068 0.30988562
#> loc10:lag3 0.32340881 0.25019011 0.29708637 0.28029171 0.30988562 0.35248994
#> loc1:lag4  0.46829962 0.46226173 0.50584442 0.39496294 0.56457922 0.48890263
#> loc2:lag4  0.51622770 0.45432526 0.53387673 0.42525425 0.55719751 0.53370571
#> loc3:lag4  0.47656147 0.51358730 0.51253032 0.37475468 0.46045462 0.46023723
#> loc4:lag4  0.53303090 0.45575020 0.54118288 0.42719422 0.53036183 0.53641233
#> loc5:lag4  0.57051645 0.42714188 0.50289098 0.44767139 0.49376990 0.54373985
#> loc6:lag4  0.42714188 0.61468512 0.48179939 0.33593704 0.44946144 0.42063891
#> loc7:lag4  0.50289098 0.48179939 0.57211957 0.39824453 0.51386126 0.49948453
#> loc8:lag4  0.44767139 0.33593704 0.39824453 0.55541996 0.42262841 0.47124805
#> loc9:lag4  0.49376990 0.44946144 0.51386126 0.42262841 0.60413446 0.52100360
#> loc10:lag4 0.54373985 0.42063891 0.49948453 0.47124805 0.52100360 0.59263327
#> loc1:lag5  0.76077679 0.75096792 0.82177023 0.64163758 0.91718794 0.79424743
#> loc2:lag5  0.83863841 0.73807472 0.86731017 0.69084736 0.90519597 0.86703233
#> loc3:lag5  0.77419859 0.83434895 0.83263182 0.60880822 0.74803219 0.74767901
#> loc4:lag5  0.86593608 0.74038961 0.87917939 0.69399895 0.86160003 0.87142938
#> loc5:lag5  0.92683326 0.69391392 0.81697223 0.72726517 0.80215456 0.88333330
#> loc6:lag5  0.69391392 0.99858754 0.78270785 0.54574698 0.73017319 0.68334951
#> loc7:lag5  0.81697223 0.78270785 0.92943762 0.64696869 0.83479401 0.81143827
#> loc8:lag5  0.72726517 0.54574698 0.64696869 0.90230824 0.68658155 0.76556666
#> loc9:lag5  0.80215456 0.73017319 0.83479401 0.68658155 0.98144745 0.84639711
#> loc10:lag5 0.88333330 0.68334951 0.81143827 0.76556666 0.84639711 0.96276319