Extract, calculate, or assign standard deviations for an mcgf or mcgf_rs object.

Extract, calculate, or assign standard deviations for an mcgf or mcgf_rs object.

Usage

# S3 method for mcgf
sds(x, ...)

# S3 method for mcgf_rs
sds(x, replace = FALSE, ...)

sds(x) <- value

Arguments

x

An mcgf or mcgf_rs object.

...

Additional parameters or attributes. Not in use.

replace

Logical; if TRUE, sds are recalculated.

value

A vector (or list of vectors) of standard deviations for all stations (under each regime and combined).

Value

sds() returns empirical (regime-switching) standard deviations.

Details

For mcgf objects, sds() extracts or computes the empirical standard deviations. The output is a vector of sds.

For mcgf_rs objects, sds() extracts or computes the regime-switching empirical standard deviations. The output is a list of vectors of sds. Each element in the list corresponds to the sds for a regime.

sds<- assigns sds to x. Use add_ccfs() to add both ccfs and sds to x.

Examples

data(sim1)
sim1_mcgf <- mcgf(sim1$data, dists = sim1$dists)
#> `time` is not provided, assuming rows are equally spaced temporally.
sds(sim1_mcgf)
#>      loc1      loc2      loc3      loc4      loc5      loc6      loc7      loc8 
#> 0.9851390 0.9759124 0.9808839 0.9717271 0.9627218 0.9992935 0.9640735 0.9498991 
#>      loc9     loc10 
#> 0.9906803 0.9812050 

data(sim2)
sim2_mcgf <- mcgf_rs(sim2$data, dists = sim2$dists, label = sim2$label)
#> `time` is not provided, assuming rows are equally spaced temporally.
sds(sim2_mcgf)
#> $sds
#>      loc1      loc2      loc3      loc4      loc5      loc6      loc7      loc8 
#> 0.9199169 1.0698542 0.9206324 1.0232810 0.9811647 0.9087964 0.9931767 1.0542091 
#>      loc9     loc10 
#> 0.9726674 0.9144110 
#> 
#> $sds_rs
#> $sds_rs$`Regime 1`
#>      loc1      loc2      loc3      loc4      loc5      loc6      loc7      loc8 
#> 0.9372674 1.0410025 0.8916340 1.0393791 0.9463280 0.8165009 1.0285441 1.0408619 
#>      loc9     loc10 
#> 0.9618800 0.9063334 
#> 
#> $sds_rs$`Regime 2`
#>      loc1      loc2      loc3      loc4      loc5      loc6      loc7      loc8 
#> 0.8616473 1.1068999 0.9547834 0.9847940 0.9794145 1.0053082 0.9338493 1.0725529 
#>      loc9     loc10 
#> 0.9690595 0.9249724 
#> 
#> 
data(sim1)
sim1_mcgf <- mcgf(sim1$data, dists = sim1$dists)
#> `time` is not provided, assuming rows are equally spaced temporally.
sim1_sds <- sds(sim1_mcgf)
sds(sim1_mcgf) <- sim1_sds

data(sim2)
sim2_mcgf <- mcgf_rs(sim2$data, dists = sim2$dists, label = sim2$label)
#> `time` is not provided, assuming rows are equally spaced temporally.
sim2_sds <- sds(sim2_mcgf)
sds(sim2_mcgf) <- sim2_sds